### 2022-06-25 #546 @s3alfisc
fwildclusterboot submission
Submitting Author Name: Alexander Fischer
Submitting Author Github Handle: @s3alfisc
Repository:  https://github.com/s3alfisc/fwildclusterboot/tree/master
Version submitted: 0.9.1 (in ropensci branch)
Submission type: Stats
Badge grade: silver
Version accepted: TBD
---

```
Package: fwildclusterboot
Title: Fast Wild Cluster Bootstrap Inference for Linear Models
Version: 0.13.2
    person("Alexander", "Fischer", , "alexander-fischer1801@t-online.de", role = c("aut", "cre")),
    person("David", "Roodman", role = "aut"),
    person("Achim", "Zeileis", role = "ctb",
           comment = "Author of included sandwich fragments"),
    person("Nathaniel", "Graham", role = "ctb",
           comment = "Contributor to included sandwich fragments"),
    person("Susanne", "Koell", role = "ctb",
           comment = "Contributor to included sandwich fragments"),
    person("Laurent", "Berge", role = "ctb",
           comment = "Author of included fixest fragments"),
    person("Sebastian", "Krantz", role = "ctb")
  )
Description: Implementation of fast algorithms for wild cluster bootstrap
    inference developed in 'Roodman et al' (2019, 'STATA' Journal,
    <doi:10.1177/1536867X19830877>) and 'MacKinnon et al' (2022), which
    makes it feasible to quickly calculate bootstrap test statistics based
    on a large number of bootstrap draws even for large samples.  Multiple
    bootstrap types as described in 'MacKinnon, Nielsen & Webb' (2022) are
    supported.  Further, 'multiway' clustering, regression weights,
    bootstrap weights, fixed effects and 'subcluster' bootstrapping are
    supported. Further, both restricted ('WCR') and unrestricted ('WCU')
    bootstrap are supported. Methods are provided for a variety of fitted
    models, including 'lm()', 'feols()' (from package 'fixest') and
    'felm()' (from package 'lfe').  Additionally implements a
    'heteroskedasticity-robust' ('HC1') wild bootstrap.  Last, the package
    provides an R binding to 'WildBootTests.jl', which provides additional
    speed gains and functionality, including the 'WRE' bootstrap for
    instrumental variable models (based on models of type 'ivreg()' from
    package 'ivreg') and hypotheses with q > 1.
    collapse,
    dqrng,
    dreamerr,
    Formula,
    generics,
    gtools,
    JuliaConnectoR,
    MASS,
    Matrix,
    Rcpp,
    summclust
    bench,
    broom,
    clubSandwich,
    covr,
    fabricatr,
    fixest,
    gt,
    ivreg,
    lfe,
    lmtest,
    modelsummary,
    sandwich,
    tibble
    Rcpp,
    RcppArmadillo,
    RcppEigen
    knitr
Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd",
    "srr::srr_stats_roclet"))
SystemRequirements: Version Requirements to run the wild bootstrap through
    Julia - Julia (>= 1.8), WildBootTests.jl (>=0.9).  Julia is
    downloadable via the official Julia website
    (https://julialang.org/downloads/), WildBootTests.jl via Julia's
    package manager (https://docs.julialang.org/en/v1/stdlib/Pkg/) or its
    github repository (https://github.com/droodman/WildBootTests.jl)
```

- [x] A pre-submission inquiry has been approved in #542

The target audience is academic social scientists (economics, political science, sociology). fwildclusterboot should be used whenever regression errors are "clustered" into few groups, in which case inference based on asymptotic approximations might fail.

    - *The first implementation of a novel algorithm*; or
    - *The first implementation within* **R** *of an algorithm which has previously been implemented in other languages or contexts*; or
    - *An improvement on other implementations of similar algorithms in* **R**.

`fwildclusterboot` is the first R-implementation of the 'fast' wild cluster bootstrap algorithm outlined in [Roodman et al (2019, Stata Journal)](https://journals.sagepub.com/doi/pdf/10.1177/1536867X19830877?casa_token=5zZJJ-enmT4AAAAA:4-QPcny5T05lBEaYd9mIP82Y4Ol03zd3kwDB08GnNq7lR1p4mYCW-52lbYNLPD6zbk0lZRc4ZjQ2SA) and implemented in Stata via the [boottest package](https://github.com/droodman/boottest). Recently, @droodman has also [ported the algorithm to Julia](https://github.com/droodman/WildBootTests.jl). Besides implementing the "fast & wild" algorithm in native R, `fwildclusterboot` also allows to call `WildBootTests.jl` via the `JuliaConnectoR` package. The "fast" algorithm is significantly faster than any other R-implementations of the wild cluster bootstrap to date (e.g. see [here](https://s3alfisc.github.io/blog/post/1000x-faster-wild-cluster-bootstrap-inference-in-r-with-fwildclusterboot/)).

Yes.

At a minimal level, I strive for a silver badge.

1) Compliance with a good number of standards; 2) I strive for compliance with all standards regarding a) documentation and b) testing.

`pkgcheck::pkgcheck()` currently fails on my machine with the following error message: "Error in normalizePath(path.expand(path), winslash, mustWork).: path[1]="C:/.../fwildclusterboot/NA" The system cannot find the specified file. It also fails on github-actions on the [`ropensci` branch](https://github.com/s3alfisc/fwildclusterboot/runs/7054793724?check_suite_focus=true).

Also, note that the CMD checks currently fail on github actions as the `gtools` package has been orphaned - I am [monitoring the situation](https://github.com/r-gregmisc/gtools/issues/15), but it looks like the package will not be removed from CRAN.

- [x] Do you intend for this package to go on CRAN? The package is already on CRAN. In case major changes are suggested during the review, I would consider integrating them into a successor package (e.g. 'fwildclusterboot2').
