+++ R/srr-stats-standards.R
@@ -15,7 +15,7 @@
 #' @srrstats {G1.3} *All statistical terminology should be clarified and unambiguously defined.* This is done in an extra vignette. 
 #' @srrstats {G1.4} *`roxygen2` is used throughout the package to document all functions.*
 #' @srrstats {G1.4a} *All internal (non-exported) functions should also be documented in standard [`roxygen2`](https://roxygen2.r-lib.org/) format, along with a final `@noRd` tag to suppress automatic generation of `.Rd` files.* Done.
-#' @srrstatsTODO {G1.6} *Software should include code necessary to compare performance claims with alternative implementations in other R packages.* 
+#' @srrstats {G1.6} *Software should include code necessary to compare performance claims with alternative implementations in other R packages.* Is a link to this blog post sufficient? https://s3alfisc.github.io/blog/post/1000x-faster-wild-cluster-bootstrap-inference-in-r-with-fwildclusterboot/
 #' @srrstats {G2.0} *Function argument checks are implemented throughout with via the `dreamerr` package.* Function values checked via dreamerr.
 #' @srrstats {G2.0a} Provide explicit secondary documentation of any expectations on lengths of inputs
 #' @srrstats {G2.1} *Implement assertions on types of inputs (see the initial point on nomenclature above).*Function values checked via dreamerr.
@@ -39,6 +39,7 @@
 #' @srrstats {G2.14a} *see G2.13*
 #' @srrstats {G2.14b} *see G2.13*
 #' @srrstats {G2.14c} *see G2.13*
+#' @srrstats {G2.15} *Functions should never assume non-missingness, and should never pass data with potential missing values to any base routines with default `na.rm = FALSE`-type parameters* Done
 #' @srrstats {G3.0} *No floating point numbers (rational numbers) are compared with equality.* Yes, checked.
 #' @srrstats {G3.1} *Statistical software which relies on covariance calculations should enable users to choose between different algorithms for calculating covariances, and should not rely solely on covariances from the `stats::cov` function.* The package deals with "clustered" covariances. Non-clustered tests based on HC1 covariances are also supported.
 #' @srrstats {G3.1a} *The ability to use arbitrarily specified covariance methods should be documented (typically in examples or vignettes).* See above.
@@ -48,7 +49,6 @@
 #' @srrstats {G5.2} *Appropriate error and warning behaviour of all functions should be explicitly demonstrated through tests. In particular,* Tested in test-warning-error test file. 
 #' @srrstats {G5.2a} *Every message produced within R code by `stop()`, `warning()`, `message()`, or equivalent should be unique* Done.
 #' @srrstats {G5.2b} *Explicit tests should demonstrate conditions which trigger every one of those messages, and should compare the result with expected values.* Done. See test-warning-error test file.
-#' @srrstatsTODO {G5.3} *For functions which are expected to return objects containing no missing (`NA`) or undefined (`NaN`, `Inf`) values, the absence of any such values in return objects should be explicitly tested.* 
 #' @srrstats {G5.4} **Correctness tests** *to test that statistical algorithms produce expected results to some fixed test data sets (potentially through comparisons using binding frameworks such as [RStata](https://github.com/lbraglia/RStata)).* Several correctness tests are implemented. First, it is tested if the non-bootstrapped t-statistics 
 #' produced via boottest() *exactly* match those computed by the fixest package (see test_tstat_equivalence). Second, `fwildclusterboot` is heavily tested against `WildBootTests.jl` - see "test-r-vs-julia".
 #' @srrstats {G5.4b} *For new implementations of existing methods, correctness tests should include tests against previous implementations. Such testing may explicitly call those implementations in testing, preferably from fixed-versions of other software, or use stored outputs from those where that is not possible.* Extensive tests against WildBootTests.jl
@@ -56,16 +56,10 @@
[…]
