### 2022-06-25 92cea90
fix minor docs errors
