### 2022-06-25 #546 @s3alfisc
Hi @emilyriederer, is it possible that the editor check bot fails because I have submitted the package via a side-branch? @mpadge mentioned in the pre-submission that some of the automated tools might fail in this case, but suggested to nevertheless submit via a non default branch for testing purposes.
