+++ tests/testthat/test-seed.R
@@ -1,7 +1,6 @@
 test_that("seed works for OLS", {
-  
   skip_on_cran()
-  
+
   data1 <<- fwildclusterboot:::create_data(N = 5000, N_G1 = 40, icc1 = 0.01, N_G2 = 10, icc2 = 0.01, numb_fe1 = 10, numb_fe2 = 10, seed = 1)
   lm_fit <- lm(proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
     data = data1
@@ -79,7 +78,7 @@ test_that("seed works for OLS", {
 
     set.seed(9)
     dqrng::dqset.seed(9)
-    
+
     boot_lm_s1 <- suppressMessages(
       boottest(
         object = lm_fit,
@@ -205,13 +204,7 @@ test_that("seed works for OLS", {
         seed = 2,
         boot_algo = boot_algo
       )
-      
     )
     expect_true(boot_lm_s1$p_val != boot_lm_s2$p_val)
   }
-  
-  
 })
-
-
-
