+++ R/set_seed.R
@@ -2,15 +2,14 @@ get_seed <- function() {
   #' creates an integer based on the global random seed set via set.seed()
   #' for using set.seed() for controlling rcpp's seed, see this
   #' blog post http://rorynolan.rbind.io/2018/09/30/rcsetseed/
-  
-  #max_int <- .Machine$integer.max
+
+  # max_int <- .Machine$integer.max
   max_int <- 2147483647L
   x <- sample.int(max_int, 1)
   x
 }
 
 set_seed <- function(seed, boot_algo, type) {
-  
   if (!is.null(seed)) {
     if (boot_algo == "R") {
       if (type %in% c("rademacher", "webb", "norm")) {
@@ -24,7 +23,7 @@ set_seed <- function(seed, boot_algo, type) {
       set.seed(seed)
       internal_seed <- NULL
     } else if (boot_algo == "WildBootTests.jl") {
-      JuliaConnectoR::juliaEval('using StableRNGs')
+      JuliaConnectoR::juliaEval("using StableRNGs")
       set.seed(seed)
       seed <- get_seed()
       internal_seed <-
@@ -33,14 +32,13 @@ set_seed <- function(seed, boot_algo, type) {
   } else if (is.null(seed)) {
     if (boot_algo == "WildBootTests.jl") {
       seed <- get_seed()
-      JuliaConnectoR::juliaEval('using StableRNGs')
+      JuliaConnectoR::juliaEval("using StableRNGs")
       internal_seed <-
         JuliaConnectoR::juliaEval(paste0("rng = StableRNG(", seed, ")"))
     } else {
       internal_seed <- NULL
     }
   }
-  
+
   internal_seed
-  
 }
