+++ R/boottest_lm.R
@@ -6,8 +6,8 @@
 #' the fast wild bootstrap algorithm developed in Roodman et al., 2019.
 #'
 #' @param object An object of class lm
-#' @param clustid A character vector or rhs formula containing the names of the cluster variables. If NULL, 
-#'        a heteroskedasticity-robust (HC1) wild bootstrap is run. 
+#' @param clustid A character vector or rhs formula containing the names of the cluster variables. If NULL,
+#'        a heteroskedasticity-robust (HC1) wild bootstrap is run.
 #' @param param A character vector or rhs formula. The name of the regression
 #'        coefficient(s) for which the hypothesis is to be tested
 #' @param B Integer. The number of bootstrap iterations. When the number of clusters is low,
@@ -26,11 +26,11 @@
 #'        intervals by test inversion. If FALSE, only the p-value is returned.
 #' @param boot_algo Character scalar. Either "R" or "WildBootTests.jl". Controls the algorithm employed by boottest().
 #'                  "R" is the default and implements the cluster bootstrap as in Roodman (2019). "WildBootTests.jl" executes the wild cluster bootstrap via the WildBootTests.jl
-#'                  package. For it to run, Julia and WildBootTests.jl need to be installed. 
+#'                  package. For it to run, Julia and WildBootTests.jl need to be installed.
 #                   The "R-lean" algorithm is a memory friendly, but less performant rcpp-armadillo based implementation of the wild cluster bootstrap.
 #'                  Note that if no cluster is provided, boottest() always defaults to the "lean" algorithm. You can set the employed algorithm globally by using the
 #'                  `setBoottest_boot_algo()` function.
-#' @param seed An integer. Allows to set a random seed. For details, see below.  
+#' @param seed An integer. Allows to set a random seed. For details, see below.
 #' @param R Hypothesis Vector giving linear combinations of coefficients. Must be either NULL or a vector of the same length as `param`. If NULL, a vector of ones of length param.
 #' @param r A numeric. Shifts the null hypothesis
 #'        H0: param = r vs H1: param != r
@@ -101,15 +101,15 @@
 #' \item{internal_seed}{The integer value -inherited from set.seed() - used within boottest() to set the random seed in either R or Julia. If NULL, no internal seed was created.}
 #' @export
 #'
-#' @section Setting Seeds: 
-#' To guarantee reproducibility, you can either use `boottest()'s` `seed` function argument, or 
-#' set a global random seed via 
+#' @section Setting Seeds:
+#' To guarantee reproducibility, you can either use `boottest()'s` `seed` function argument, or
+#' set a global random seed via
 #' + `set.seed()` when using
-#'    1) the lean algorithm (via `boot_algo = "R-lean"`) including the heteroskedastic wild bootstrap 
-#'    2) the wild cluster bootstrap via `boot_algo = "R"` with Mammen weights or 
+#'    1) the lean algorithm (via `boot_algo = "R-lean"`) including the heteroskedastic wild bootstrap
+#'    2) the wild cluster bootstrap via `boot_algo = "R"` with Mammen weights or
 #'    3) `boot_algo = "WildBootTests.jl"`
 #' + `dqrng::dqset.seed()` when using `boot_algo = "R"` for Rademacher, Webb or Normal weights
-#' 
+#'
 #' @section Confidence Intervals:
 #' \code{boottest} computes confidence intervals by inverting p-values.
 #'       In practice, the following procedure is used:
[…]

[…]
