### 2022-06-25 36ce08c
change sapply to vapply
