+++ man/boottest.fixest.Rd
@@ -238,13 +238,13 @@ if (requireNamespace("fixest")) {
     R = c(1, 1),
     r = 2
   )
-summary(boot1)
-print(boot1)
-plot(boot1)
-nobs(boot1)
-pval(boot1)
-confint(boot1)
-generics::tidy(boot1)
+  summary(boot1)
+  print(boot1)
+  plot(boot1)
+  nobs(boot1)
+  pval(boot1)
+  confint(boot1)
+  generics::tidy(boot1)
 }
 }
 
