### 2022-06-25 e5ecf4f
document missing params in invert_p_val
