### 2022-06-26 e44e522
switch back to sapply
