### 2022-06-26 be9603b
temporarily drop dependence on gtools, as orphaned #48
