### 2022-06-28 #546 @mpadge
Thanks @emilyriederer, because of this submission we now have an ability to process packages with non-default review branches. You (or indeed @s3alfisc) may call `check package` at any time. (But I've noted to @s3alfisc elsewhere that R CMD check currently fails on his package because our check systems don't have Julia installed.)
