+++ tests/testthat/test-global_vars.R
@@ -1,6 +1,7 @@
 test_that("global boot_algo", {
   skip_on_cran()
-
+  skip_on_ci()
+  
   library(fwildclusterboot)
 
   data(voters)
