+++ tests/testthat/test-error_warning.R
@@ -1,6 +1,7 @@
 test_that("errors and warnings q = 1", {
   skip_on_cran()
-
+  skip_on_ci()
+  
   # ------------------------------------------------------------------ #
   # test for warnings and errors
   # ------------------------------------------------------------------ #
