+++ tests/testthat/test-seed.R
@@ -1,6 +1,7 @@
 test_that("seed works for OLS", {
   skip_on_cran()
-
+  skip_on_ci()
+  
   data1 <<- fwildclusterboot:::create_data(N = 5000, N_G1 = 40, icc1 = 0.01, N_G2 = 10, icc2 = 0.01, numb_fe1 = 10, numb_fe2 = 10, seed = 1)
   lm_fit <- lm(proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
     data = data1
