### 2022-07-02 83348db
add contributing md, add examples / noRd tags
