+++ R/set_seed.R
@@ -2,6 +2,7 @@ get_seed <- function() {
   #' creates an integer based on the global random seed set via set.seed()
   #' for using set.seed() for controlling rcpp's seed, see this
   #' blog post http://rorynolan.rbind.io/2018/09/30/rcsetseed/
+  #' @noRd
 
   # max_int <- .Machine$integer.max
   max_int <- 2147483647L
