### 2022-07-03 96e822c
#18 some code reorg, all linelengths <= 80
