+++ tests/testthat/test-global_vars.R
@@ -3,26 +3,58 @@ test_that("global boot_algo", {
   skip_on_ci()
   
   library(fwildclusterboot)
-
+  
   data(voters)
-  lm_fit <- lm(proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
-    data = voters
-  )
-
-  boot1 <- boottest(lm_fit, param = "treatment", B = 999, clustid = "group_id1")
-  boot2 <- suppressWarnings(boottest(lm_fit, param = "treatment", B = 999, clustid = "group_id1", boot_algo = "WildBootTests.jl"))
-
+  lm_fit <-
+    lm(proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
+       data = voters)
+  
+  boot1 <-
+    boottest(lm_fit,
+             param = "treatment",
+             B = 999,
+             clustid = "group_id1")
+  boot2 <-
+    suppressWarnings(
+      boottest(
+        lm_fit,
+        param = "treatment",
+        B = 999,
+        clustid = "group_id1",
+        boot_algo = "WildBootTests.jl"
+      )
+    )
+  
   expect_equal(boot1$boot_algo, "R")
   expect_equal(boot2$boot_algo, "WildBootTests.jl")
-
+  
   # leave state of global variables clean
   setBoottest_boot_algo(boot_algo = "WildBootTests.jl")
   # switch back to "R" as global variable at end of function
   on.exit(setBoottest_boot_algo(boot_algo = "R"), add = TRUE)
-  boot1 <- suppressWarnings(boottest(lm_fit, param = "treatment", B = 999, clustid = "group_id1"))
-  boot2 <- suppressWarnings(boottest(lm_fit, param = "treatment", B = 999, clustid = "group_id1"))
-  boot3 <- boottest(lm_fit, param = "treatment", B = 999, clustid = "group_id1", boot_algo = "R")
-
+  boot1 <-
+    suppressWarnings(boottest(
+      lm_fit,
+      param = "treatment",
+      B = 999,
+      clustid = "group_id1"
+    ))
+  boot2 <-
+    suppressWarnings(boottest(
+      lm_fit,
+      param = "treatment",
+      B = 999,
+      clustid = "group_id1"
+    ))
+  boot3 <-
+    boottest(
+      lm_fit,
+      param = "treatment",
+      B = 999,
+      clustid = "group_id1",
+      boot_algo = "R"
+    )
+  
   expect_equal(boot1$boot_algo, "WildBootTests.jl")
   expect_equal(boot2$boot_algo, "WildBootTests.jl")
   expect_equal(boot3$boot_algo, "R")
@@ -32,33 +64,61 @@ test_that("global boot_algo", {
 
 
 # test_that("global nthreads", {
-#
 #   skip_on_ci()
 #   skip_on_cran()
-#
+#   
 #   library(fwildclusterboot)
 #   data(voters)
-#   lm_fit <- lm(proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
-#     data = voters
-#   )
-#
-#   boot1 <- boottest(lm_fit, param = "treatment", B = 999, clustid = "group_id1")
-#   boot2 <- boottest(lm_fit, param = "treatment", B = 999, clustid = "group_id1", nthreads = 4)
-#   boot3 <- boottest(lm_fit, param = "treatment", B = 999, clustid = "group_id1", nthreads = 0.5)
-#
+#   lm_fit <-
+#     lm(proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
+#        data = voters)
+#   
+#   boot1 <-
+#     boottest(lm_fit,
+#              param = "treatment",
[…]
