### 2022-07-03 419112e
#18 replace sapply with vapply
