+++ tests/testthat/test-method_equivalence.R
@@ -1,1194 +1,14 @@
 test_that("Do different, but equivalent ways to specify
-          linear models lead to equivalent results?",
-          {
-            skip_on_cran()
-            skip_on_ci()
-            
-            library(fixest)
-            library(lfe)
-            # library(fwildclusterboot)
-            
-            print_results <- FALSE
-            
-            data1 <<-
-              fwildclusterboot:::create_data(
-                N = 10000,
-                N_G1 = 20,
-                icc1 = 0.01,
-                N_G2 = 10,
-                icc2 = 0.01,
-                numb_fe1 = 10,
-                numb_fe2 = 10,
-                seed = 71986045
-              )
-            sapply(data1, class)
-            
-            lm_fit <-
-              lm(
-                proposition_vote ~ treatment + ideology1 + log_income + 
-                  Q1_immigration + Q2_defense,
-                data = data1
-              )
-            feols_fit1 <-
-              feols(
-                proposition_vote ~ treatment + ideology1 + log_income + 
-                  Q1_immigration + Q2_defense,
-                data = data1
-              )
-            feols_fit2 <-
-              feols(
-                proposition_vote ~ treatment + ideology1 + log_income + 
-                  Q1_immigration + Q2_defense,
-                data = data1,
-                cluster = "group_id1"
-              )
-            feols_fit3 <-
-              feols(
-                proposition_vote ~ treatment + ideology1 + log_income + 
-                  Q1_immigration + Q2_defense,
-                data = data1,
-                cluster = ~ group_id1
-              )
-            feols_fit4 <-
-              feols(
-                proposition_vote ~ treatment + ideology1 + log_income +
-                  Q1_immigration + Q2_defense,
-                data = data1,
-                cluster = c("group_id1", "group_id2")
-              )
-            feols_fit5 <-
-              feols(
-                proposition_vote ~ treatment + ideology1 + log_income +
-                  Q1_immigration + Q2_defense,
-                data = data1,
-                cluster = ~ group_id1 + group_id2
-              )
-            feols_fit6 <-
-              feols(
-                proposition_vote ~ treatment + ideology1 + log_income |
-                  Q1_immigration + Q2_defense,
-                data = data1
-              )
-            feols_fit7 <-
-              feols(
-                proposition_vote ~ treatment + ideology1 + log_income |
-                  Q1_immigration + Q2_defense,
-                data = data1,
-                cluster = "group_id1"
-              )
-            feols_fit8 <-
-              feols(
-                proposition_vote ~ treatment + ideology1 + log_income |
-                  Q1_immigration + Q2_defense,
-                data = data1,
-                cluster = ~ group_id1
-              )
-            feols_fit9 <-
-              feols(
-                proposition_vote ~ treatment + ideology1 + log_income |
-                  Q1_immigration + Q2_defense,
-                data = data1,
-                cluster = c("group_id1", "group_id2")
-              )
-            feols_fit10 <-
-              feols(
-                proposition_vote ~ treatment + ideology1 + log_income |
-                  Q1_immigration + Q2_defense,
-                data = data1,
-                cluster = ~ group_id1 + group_id2
-              )
-            feols_fit11 <-
-              feols(
-                proposition_vote ~ treatment + ideology1 + log_income +
-                  Q1_immigration |
-                  Q2_defense,
-                data = data1
-              )
-            feols_fit12 <-
-              feols(
-                proposition_vote ~ treatment + ideology1 + log_income +
-                  Q1_immigration,
-                fixef = "Q2_defense",
-                data = data1
-              )
-            feols_fit13 <-
-              feols(
-                proposition_vote ~ treatment + ideology1 + log_income,
[…]
