+++ R/boottest_felm.R
@@ -16,20 +16,20 @@
 #'        increasing B adds little additional runtime.
 #' @param bootcluster A character vector or rhs formula of length 1.
 #'  Specifies the bootstrap clustering variable or variables. If more
-#'        than one variable is specified, then bootstrapping is clustered 
+#'        than one variable is specified, then bootstrapping is clustered
 #'        by the intersections of
-#'        clustering implied by the listed variables. To mimic the behavior 
+#'        clustering implied by the listed variables. To mimic the behavior
 #'        of stata's boottest command,
-#'        the default is to cluster by the intersection of all the variables 
+#'        the default is to cluster by the intersection of all the variables
 #'        specified via the `clustid` argument,
 #'        even though that is not necessarily recommended (see the paper by
 #'         Roodman et al cited below, section 4.2).
-#'        Other options include "min", where bootstrapping is clustered by 
+#'        Other options include "min", where bootstrapping is clustered by
 #'        the cluster variable with the fewest clusters.
 #'        Further, the subcluster bootstrap (MacKinnon & Webb, 2018) is
-#'         supported - see the \code{vignette("fwildclusterboot", 
+#'         supported - see the \code{vignette("fwildclusterboot",
 #'         package = "fwildclusterboot")} for details.
-#' @param fe A character vector or rhs formula of length one which contains 
+#' @param fe A character vector or rhs formula of length one which contains
 #' the name of the fixed effect to be projected
 #'        out in the bootstrap. Note: if regression weights are used, fe
 #'        needs to be NULL.
@@ -38,17 +38,17 @@
 #'        returns 0.95% confidence intervals. By default, sign_level = 0.05.
 #' @param conf_int A logical vector. If TRUE, boottest computes confidence
 #'        intervals by test inversion. If FALSE, only the p-value is returned.
-#' @param boot_algo Character scalar. Either "R" or "WildBootTests.jl". 
+#' @param boot_algo Character scalar. Either "R" or "WildBootTests.jl".
 #' Controls the algorithm employed by boottest().
-#'                  "R" is the default and implements the cluster bootstrap 
-#'                  as in Roodman (2019). "WildBootTests.jl" executes the 
+#'                  "R" is the default and implements the cluster bootstrap
+#'                  as in Roodman (2019). "WildBootTests.jl" executes the
 #'                  wild cluster bootstrap via the WildBootTests.jl
 #'                  package. For it to run, Julia and WildBootTests.jl
 #'                  need to be installed.
 #'                  The "R-lean" algorithm is a memory friendly, but less
-#'                  performant rcpp-armadillo based implementation of the 
+#'                  performant rcpp-armadillo based implementation of the
 #'                  wild cluster bootstrap.
-#'                  Note that if no cluster is provided, boottest() always 
+#'                  Note that if no cluster is provided, boottest() always
 #'                  defaults to the "lean" algorithm. You can set the employed
 #'                   algorithm globally by using the
 #'                  `setBoottest_boot_algo()` function.
@@ -63,7 +63,7 @@
 #'        of boostrap to use: One of "rademacher", "mammen", "norm"
 #'        and "webb". Alternatively, type can be a function(n) for drawing
[…]
