+++ R/boottest_felm.R
@@ -124,7 +124,6 @@
 #' "WildBootTests.jl"'
 #' @param t_boot Logical. Should bootstrapped t-statistics be returned?
 #' @param ... Further arguments passed to or from other methods.
-#' @import JuliaConnectoR
 #'
 #' @importFrom dreamerr check_arg validate_dots
 #' @return An object of class \code{boottest}
