+++ R/boottest_fixest.R
@@ -105,7 +105,7 @@
 #' 'boot_algo = "WildBootTests.jl"'
 #' @param t_boot Logical. Should bootstrapped t-statistics be returned?
 #' @param ... Further arguments passed to or from other methods.
-#' @import JuliaConnectoR
+
 #' @importFrom dreamerr check_arg validate_dots
 
 #' @return An object of class \code{boottest}
