+++ R/model_matrix.R
@@ -1,9 +1,9 @@
 model_matrix <- function(object, ...) {
+  
   #' enhanced model.matrix functionalities
   #' @param object An object of class `lm` or `felm``
   #' @param ... Other arguments
   #' @export
-  #' @noRd
 
 
   UseMethod("model_matrix")
