+++ NAMESPACE
@@ -32,7 +32,8 @@ export(model_matrix)
 export(pval)
 export(setBoottest_boot_algo)
 export(teststat)
-import(JuliaConnectoR)
+importFrom(JuliaConnectoR,juliaEval)
+importFrom(JuliaConnectoR,juliaImport)
 importFrom(Matrix,Diagonal)
 importFrom(Matrix,t)
 importFrom(Matrix.utils,aggregate.Matrix)
