+++ tests/testthat/test-error_warning.R
@@ -1,1226 +1,68 @@
 test_that("errors and warnings q = 1", {
+  
   skip_on_cran()
   skip_on_ci()
 
-  # ------------------------------------------------------------------ #
-  # test for warnings and errors
-  # ------------------------------------------------------------------ #
-  library(fwildclusterboot)
-  library(fixest)
-  library(lfe)
+  if(juliaconnector_prepared){
 
-
-  for (boot_algo in c("R", "WildBootTests.jl", "R-lean")) {
-    cat(boot_algo, "\n")
-
-    # for(boot_algo in c("R-lean")){
-    # test boottest function arguments for errors
-    lm_fit <-
-      lm(
-        proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
-        data = fwildclusterboot:::create_data(
-          N = 1000,
-          N_G1 = 10,
-          icc1 = 0.01,
-          N_G2 = 10,
-          icc2 = 0.01,
-          numb_fe1 = 10,
-          numb_fe2 = 10,
-          seed = 1234
+    # ------------------------------------------------------------------ #
+    # test for warnings and errors
+    # ------------------------------------------------------------------ #
+    # library(fwildclusterboot)
+    # library(fixest)
+    # library(lfe)
+    
+    
+    for (boot_algo in c("R", "WildBootTests.jl", "R-lean")) {
+      cat(boot_algo, "\n")
+      
+      # for(boot_algo in c("R-lean")){
+      # test boottest function arguments for errors
+      lm_fit <-
+        lm(
+          proposition_vote ~ treatment + ideology1 + log_income +
+            Q1_immigration,
+          data = fwildclusterboot:::create_data(
+            N = 1000,
+            N_G1 = 10,
+            icc1 = 0.01,
+            N_G2 = 10,
+            icc2 = 0.01,
+            numb_fe1 = 10,
+            numb_fe2 = 10,
+            seed = 1234
+          )
         )
-      )
-    feols_fit <-
-      feols(
-        proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
-        data = fwildclusterboot:::create_data(
-          N = 1000,
-          N_G1 = 10,
-          icc1 = 0.01,
-          N_G2 = 10,
-          icc2 = 0.01,
-          numb_fe1 = 10,
-          numb_fe2 = 10,
-          seed = 1234
+      feols_fit <-
+        feols(
+          proposition_vote ~ treatment + ideology1 + log_income 
+          + Q1_immigration,
+          data = fwildclusterboot:::create_data(
+            N = 1000,
+            N_G1 = 10,
+            icc1 = 0.01,
+            N_G2 = 10,
+            icc2 = 0.01,
+            numb_fe1 = 10,
+            numb_fe2 = 10,
+            seed = 1234
+          )
         )
-      )
-    felm_fit <-
-      felm(
-        proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
-        data = fwildclusterboot:::create_data(
-          N = 1000,
-          N_G1 = 10,
-          icc1 = 0.01,
-          N_G2 = 10,
-          icc2 = 0.01,
-          numb_fe1 = 10,
-          numb_fe2 = 10,
-          seed = 1234
+      felm_fit <-
+        felm(
+          proposition_vote ~ treatment + ideology1 + log_income 
+          + Q1_immigration,
+          data = fwildclusterboot:::create_data(
+            N = 1000,
+            N_G1 = 10,
+            icc1 = 0.01,
+            N_G2 = 10,
+            icc2 = 0.01,
+            numb_fe1 = 10,
+            numb_fe2 = 10,
+            seed = 1234
+          )
         )
-      )
-    feols_fit_c <-
-      feols(
[…]
