+++ tests/testthat/test-convergence.R
@@ -1,4 +1,5 @@
 test_that("algorithm performance test", {
+  
   skip_on_cran()
   skip_on_ci()
 

[…]
