### 2022-07-08 ead6385
#50 add warning message when fixed effects or cluster vars in felm or feols specified as characters
