+++ DESCRIPTION
@@ -1,7 +1,6 @@
 Package: fwildclusterboot
 Title: Fast Wild Cluster Bootstrap Inference for Linear Models
-Version: 0.9
-Date: 2022-06-06
+Version: 0.9.1
 Authors@R: 
     c(
     person(given = "Alexander",
