### 2022-07-09 db5e96b
#50 clarifying error message when fixed effects or clusters are of class character
