### 2022-08-08 4a07694
fix bug in r-lean, add enumeration tests
