+++ R/mboottest_fixest.R
@@ -63,8 +63,6 @@
 #'  point numbers in Julia be represented as 32 or 64 bit?
 #' @param getauxweights Logical. FALSE by default. Whether to save auxilliary
 #'  weight matrix (v)
-#' @param teststat_boot Logical. Should bootstrapped test statistics be
-#'  returned?
 #' @param maxmatsize NULL by default = no limit. Else numeric scalar to set
 #' the maximum size of auxilliary weight matrix (v), in gigabytes
 #' @param bootstrapc Logical scalar, FALSE by default. TRUE  to request
@@ -180,7 +178,6 @@ mboottest.fixest <- function(object,
                              tol = 1e-6,
                              floattype = "Float64",
                              getauxweights = FALSE,
-                             teststat_boot = FALSE,
                              maxmatsize = NULL,
                              bootstrapc = FALSE,
                              ssc = boot_ssc(
@@ -214,7 +211,6 @@ mboottest.fixest <- function(object,
   check_arg(floattype, "charin(Float32, Float64)")
   check_arg(maxmatsize, "scalar integer | NULL")
   check_arg(bootstrapc, "scalar logical")
-  check_arg(teststat_boot, "logical scalar")
 
   if (inherits(clustid, "formula")) {
     clustid <- attr(terms(clustid), "term.labels")
