### 2022-08-09 815e405
add test for multi-param HC1 bootstrap
