+++ src/wildboottest.cpp
@@ -78,73 +78,6 @@ arma::mat sample_weights(int G,
   
 }
 
-// // [[Rcpp::export]]
-// SEXP test(){
-//   Environment pkg = Environment::namespace_env("stats");
-//   Function f = pkg["runif"];
-//   return(f(Named("n") = 1));
-// }
-// 
-// // [[Rcpp::export]]
-// NumericMatrix v_enumerated(int N_G_bootcluster){
-// 
-//   Environment pkg = Environment::namespace_env("gtools");
-// 
-//   Function f = pkg["permutations"];
-// 
-//   NumericVector s(2);
-//   NumericVector prob = NumericVector::create();
-//   s[0] = 1;
-//   s[1] = -1;
-// 
-//   return( f( Named("n") =  2, Named("r")=N_G_bootcluster, Named("v") =s,
-// Named("repeats.allowed") = TRUE) );
-// 
-// }
-
-
-// // [[Rcpp::export]]
-// arma::mat sample_weights2(int G,
-//                          int type){
-//   
-//   // R random function based random numbers
-//   // type == 0 for rademacher
-//   // type == 1 for mammen
-//   arma::vec weights(G);
-//   
-//   if(type == 0){
-//     for(int g = 0; g < G; g++){
-//       double v = R::runif(0, 1);
-//       if(v < 0.5){
-//         weights(g) = -1;
-//       } else {
-//         weights(g) = 1;
-//       }
-//     }
-//   } else if(type == 1){
-//     for(int g = 0; g < G; g++){
-//       double v = R::runif(0, 1);
-//       if(v < 1/6){
-//         weights(g) = -sqrt(1.5);
-//       } else if (v < 1/3){
-//         weights(g) = -1;
-//       } else if (v < 1/2){
-//         weights(g) = -sqrt(0.5);
-//       }  else if (v < 2/3){
-//         weights(g) = sqrt(0.5);
-//       } else if (v < 5/6){
-//         weights(g) = 1;
-//       } else if (v >= 5/6){
-//         weights(g) = sqrt(1.5);
-//       }
-//     }
-//   }
-//   
-//   
-//   return weights;
-//   
-// }
-
 
 //' Implementation of the heteroskedastic wild bootstrap. Computes
 //' HC robust variance estimators. For use in fwildclusterboot when no
