### 2022-08-09 ebb5d02
Ãdelete unfinished documentation drafts (for now)
