+++ vignettes/articles/technical_vignette.Rmd
@@ -1,465 +1,0 @@
----
-title: "The (Wild) Cluster Bootstrap and Implementation Details - Preliminary Draft"
----
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-```
-
-```{r setup}
-library(fwildclusterboot)
-```
-
-This vignette presents several implementation details of the algorithms underlying `fwildclusterboot`. It introduces the wild (cluster) bootstrap, illustrates how `fwildclusterboot::boottest()` inverts p-values to obtain confidence intervals, and explains how the "fast wild cluster bootstrap algorithm" is implemented in `fwildclusterboot` for the "R" algorithm. 
-
-## The Wild (Cluster) Bootstrap 
-
-We want to test a hypothesis $R\beta = r$ for a linear regression model
-
-$$
-  Y = X\beta + u.
-$$
-
-with $E(u|X) = 0$ via a wild bootstrap. Following Roodman et al (2019), the wild bootstrap proceeds in the following steps: 
-
-1) - for the WCU (uncrestricted bootstrap): regress Y on X; obtain $\hat{\beta}$, $\hat{u}$ and $\hat{V}$. 
-   - for the WCR (restricted bootstrap): regress Y on X imposing the null (via restricted least squares), obtain $\tilde{\beta}$, $\tilde{u}$ and $\tilde{V}$
-2) calculate the (cluster) robust t-statistic for the null hypothesis of interest
-  $$
-    t = \frac{R\hat{\beta} - r}{\sqrt(R'\hat{V} R)}
-  $$
-  where $\beta$ is an estimate of the parameter of interest and $V$ the associated variance-covariance matrix.
-3) for all $b = 1, ..., B$ bootstrap iterations, create
-   - a set of bootstrap error terms $v_{b}^{*}$
-   - a new bootstrap sample $y_{b}^{*} = X \beta + f(\hat{u}) \times v_{b}^{*}$
-   - a bootstrapped t-statistic
-      $$
-        t_{b}^{*} = \frac{R\hat{\beta_{b}^{*}} - r}{\sqrt(R'\hat{V}_{b}^{*} R)}
-      $$
-      where $\beta_{b}^{*}$ is the OLS estimate of a regression of X on $y_{b}^{*}$ and 
-      $\hat{V}_{b}^{*}$ the associated (cluster) robust variance-covariance estimate
-4) based on all bootsrapped t-statistics, calculate a bootstrap p-value as 
-  
-  $$
-  \begin{align}
-    \text{left-tailed: } \hat{P}_{L}^{*} = \frac{1}{B} \sum_{b = 1}^{B} 1(t^{*^{b}}_{j} < t_j) && \text{right tailed: } \hat{P}_{R}^{*} = \frac{1}{B} \sum_{b = 1}^{B} 1(t^{*^{b}}_{j} > t_j)
-  \end{align}
-  $$
-
-
-  $$
-  \begin{align}
-   \text{two-tailed: }   \hat{P}_{S}^{*} = \frac{1}{B} \sum_{b = 1}^{B} 1(|t^{*^{b}}_{j}| > |t_j|) && \text{equal-tailed: } \hat{P}_{ET}^{*} = 2 min(\hat{P}_{L}^{*}, \hat{P}_{R}^{*})
-  \end{align}
-  $$
-For the **heteroskedastic wild bootstrap**, the variance-covariance matrices are calculated as 
[…]
