+++ vignettes/articles/literature.Rmd
@@ -1,90 +1,0 @@
----
-title: "Literature on the Wild Cluster Bootstrap"
----
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-```
-
-The following section compiles a list of papers, software implementations and blog posts on the Wild Cluster Bootstrap. Let me know if you have any suggestions for other papers / blog posts to include! 
-
-
-## Software
-
-+ Stata: [boottest]()
-
-+ Julia: [WildBootTests.jl]()
-
-+ R: [fwildclusterboot]()
-
-  Other R implementations are included in the [sandwich]() and [clusterSEs]() packages. 
-  
-## Academic Papers
-
-+ [Roodman, David, et al. "Fast and wild: Bootstrap inference in Stata using boottest." The Stata Journal 19.1 (2019): 4-60.](https://journals.sagepub.com/doi/pdf/10.1177/1536867X19830877?casa_token=vapijGHUIIoAAAAA:S9AOnL7mdY3yeteZf-gXg1miBS96qrT7X53H5FWDlqokdSiqc8oPa8scKwJeF2_WeGFMxpGJzAwTwA)
-
-The paper introduces the `boottest` Stata module, discusses its implementation, and suggests when and how to use the wild cluster bootstrap. 
-
-+ [Cameron, A. Colin, Jonah B. Gelbach, and Douglas L. Miller. "Bootstrap-based improvements for inference with clustered errors." The review of economics and statistics 90.3 (2008): 414-427.](https://www.nber.org/system/files/working_papers/t0344/t0344.pdf) 
-
-The original simulation study that argued that the wild cluster bootstrap might be preferred to "standard" cluster robust sandwich estimators when the number of clusters is small.
-
-+ [MacKinnon, James G., Morten Ørregaard Nielsen, and Matthew D. Webb. Fast and reliable jackknife and bootstrap methods for cluster-robust inference. No. 1485. 2022.](https://www.econ.queensu.ca/sites/econ.queensu.ca/files/wpaper/qed_wp_1485.pdf) 
-
-Suggests new fast jackknife methods to compute CRV2 and CRV3 estimators and variants of the wild cluster bootstrap based on these new estimators.
-
-+ [Niccodemi, Gianmaria, and Tom Wansbeek. "A New Estimator for Standard Errors with Few Unbalanced Clusters." Econometrics 10.1 (2022): 6.](https://www.mdpi.com/2225-1146/10/1/6/htm) 
-
-Introduces new CRV2 and CRV3 estimators, which the paper by MNW above build upon. 
-
-+ [MacKinnon, James G., Morten Ørregaard Nielsen, and Matthew D. Webb. "Cluster-robust inference: A guide to empirical practice." Journal of Econometrics (2022).](https://www.sciencedirect.com/science/article/pii/S0304407622000781)
-
-An up-to date survey paper on recent advances how to deal with clustered errors. 
-
-+ [Cameron, A. Colin, and Douglas L. Miller. "A practitioner’s guide to cluster-robust inference." Journal of human resources 50.2 (2015): 317-372.](http://jhr.uwpress.org/content/50/2/317.full.pdf?casa_token=m_Akn2a2l1sAAAAA:96lg39sv-UbrK-p4TH2UEOpF2QGdxwNLQ32vbU8ig7LhDCKes_QeVZCsXBQvYfEhQmM1DATXqBw)
-
-The survey paper on inference with clustered errors I read during my masters program. 
-
[…]
