### 2022-08-09 4455912
remove several skip_on_ci flags to increase test coverage
