+++ tests/testthat/test-r-lean.R
@@ -1,4 +1,6 @@
 test_that("test lean cpp boottest", {
+  
+  
   set.seed(9809873)
   # library(fixest)
   # library(lfe)
