+++ tests/testthat/test-uncategorized.R
@@ -1,4 +1,5 @@
 test_that("uncategorized tests", {
+  
   skip_on_cran()
 
   # see this issue: https://github.com/s3alfisc/fwildclusterboot/issues/26
