+++ tests/testthat/test-seed.R
@@ -1,7 +1,6 @@
 test_that("seed works for OLS", {
   
   skip_on_cran()
-  skip_on_ci()
 
   if(juliaconnector_prepared){
     data1 <<-
