+++ tests/testthat/test-error_warning.R
@@ -1,7 +1,6 @@
 test_that("errors and warnings q = 1", {
   
   skip_on_cran()
-  skip_on_ci()
 
   if(juliaconnector_prepared){
 
