### 2022-08-09 bd632f3
run r-vs-julia tests on ci
