+++ tests/testthat/test-r-vs-julia.R
@@ -1,6 +1,8 @@
 test_that("test r against Julia I: stochastic tests", {
+  
+  
   skip_on_cran()
-  skip_on_ci()
+  #skip_on_ci()
 
 
   if(juliaconnector_prepared){
