### 2022-08-09 a03599b
move around julia_bindir check for tests
