+++ tests/testthat/test-r-vs-julia.R
@@ -2,10 +2,10 @@ test_that("test r against Julia I: stochastic tests", {
   
   
   skip_on_cran()
-  #skip_on_ci()
+  skip_on_ci()
 
 
-  if(juliaconnector_prepared){
+  if(is_juliaconnector_prepared()){
     reltol <- 0.05
     
     N <- 10000
