+++ tests/testthat/test_tidy.R
@@ -2,7 +2,7 @@ test_that("test tidiers with q = 1", {
 
   skip_on_cran()
   
-  if(juliaconnector_prepared){
+  if(is_juliaconnector_prepared()){
 
     lm_fit <<-
       lm(
@@ -59,7 +59,7 @@ test_that("test tidiers with q > 1", {
   
   skip_on_cran()
 
-  if(juliaconnector_prepared){
+  if(is_juliaconnector_prepared()){
     
   lm_fit <<-
     lm(
