+++ tests/testthat/test-full-enumeration.R
@@ -5,7 +5,7 @@ test_that("test full enumeration cases", {
   skip_on_cran()
   # if not skipped, codecov fails after ~6h
 
-  if(juliaconnector_prepared){
+  if(is_juliaconnector_prepared()){
     reltol <- 0.05
     
     N <- 2000
