+++ R/p_val_null.R
@@ -44,6 +44,9 @@ p_val_null2 <-
     numer <- A + B * r
     names_clustid <- names(clustid)
 
+    cat("numer: ", numer[3], "\n")
+    cat("ssc:", small_sample_correction, "\n")
+    
     JJ <- list()
     for (x in seq_along(names_clustid)) {
       JJ[[x]] <-
@@ -59,6 +62,7 @@ p_val_null2 <-
     }
 
     denom <- suppressWarnings(sqrt(JJ_sum))
+    cat("denom:", denom[3], "\n")
 
     t <- numer / denom
     t_boot <- t[2:(boot_iter + 1)]
