+++ R/get_scores.R
@@ -8,53 +8,29 @@ get_scores <- function(bootstrap_type,
                        beta_g_hat, 
                        beta_1g_tilde){
   
-  # scores_template <- function(g, a, b, beta){
-  #   a[[g]] - b[[g]] %*% beta
-  # }
+
+  bootstrap_type <- substr(bootstrap_type, 1, 4)
   
-  
-  score_num <- 
+  scores <- 
     switch(
       bootstrap_type, 
-      WCR11 = function(g) tXgyg[[g]] - tXgXg[[g]] %*% beta_tilde, 
-      WCR13 = function(g) tXgyg[[g]] - tXgXg[[g]] %*% beta_tilde, 
-      WCR31 = function(g) tXgyg[[g]] - tXgX1g[[g]] %*% beta_1g_tilde[[g]], 
-      WCR33 = function(g) tXgyg[[g]] - tXgX1g[[g]] %*% beta_1g_tilde[[g]], 
-      WCU11 = function(g) tXgyg[[g]] - tXgXg[[g]] %*% beta_hat, 
-      WCU13 = function(g) tXgyg[[g]] - tXgXg[[g]] %*% beta_hat, 
-      WCU31 = function(g) tXgyg[[g]] - tXgXg[[g]] %*% beta_g_hat[[g]],
-      WCU33 = function(g) tXgyg[[g]] - tXgXg[[g]] %*% beta_g_hat[[g]]
+      WCR1 = function(g) tXgyg[[g]] - tXgXg[[g]] %*% beta_tilde, 
+      # WCR13 = function(g) tXgyg[[g]] - tXgXg[[g]] %*% beta_tilde, 
+      WCU1 = function(g) tXgyg[[g]] - tXgXg[[g]] %*% beta_hat, 
+      # WCU13 = function(g) tXgyg[[g]] - tXgXg[[g]] %*% beta_hat, 
+      WCR3 = function(g) tXgyg[[g]] - tXgX1g[[g]] %*% beta_1g_tilde[[g]], 
+      # WCR33 = function(g) tXgyg[[g]] - tXgX1g[[g]] %*% beta_1g_tilde[[g]], 
+      WCU3 = function(g) tXgyg[[g]] - tXgXg[[g]] %*% beta_g_hat[[g]],
+      # WCU33 = function(g) tXgyg[[g]] - tXgXg[[g]] %*% beta_g_hat[[g]]
     )
   
-  score_denom <- 
-    switch(
-      bootstrap_type, 
-      WCR11 = function(g) tXgyg[[g]] - tXgXg[[g]] %*% beta_tilde, 
-      WCR13 = function(g) tXgyg[[g]] - tXgX1g[[g]] %*% beta_1g_tilde[[g]], 
-      WCR31 = function(g) tXgyg[[g]] - tXgXg[[g]] %*% beta_tilde,
-      WCR33 = function(g) tXgyg[[g]] - tXgX1g[[g]] %*% beta_1g_tilde[[g]], 
-      WCU11 = function(g) tXgyg[[g]] - tXgXg[[g]] %*% beta_hat, 
-      WCU13 = function(g) tXgyg[[g]] - tXgXg[[g]] %*% beta_g_hat[[g]],
-      WCU31 = function(g) tXgyg[[g]] - tXgXg[[g]] %*% beta_hat, 
-      WCU33 = function(g) tXgyg[[g]] - tXgXg[[g]] %*% beta_g_hat[[g]]
-    )
-  
-  score_num <- lapply(
+
+  scores_list <- lapply(
     1:G, 
-    function(g) eval(score_num(g))
+    function(g) eval(scores(g))
   )
+
   
-  score_denom <- lapply(
-    1:G, 
[…]
