+++ NAMESPACE
@@ -71,4 +71,6 @@ importFrom(stats,terms)
 importFrom(stats,uniroot)
 importFrom(stats,update)
 importFrom(stats,weights)
+importFrom(summclust,coeftable)
+importFrom(summclust,summclust)
 useDynLib(fwildclusterboot)

[…]
