### 2022-08-16 3c8f5a5
fix bugs in boot algo 3
