+++ R/boot_algo3.R
@@ -51,14 +51,17 @@ boot_algo3 <- function(preprocessed_object,
   # other checks: only test of one param, no weights, no fixed effects, 
   # ...
   
+  #here for debugging
+  #preprocessed_object <- preprocess
+  
+  
   if(substr(bootstrap_type, 5, 5) == 1){
     crv_type <- "crv1"  
   } else {
     crv_type <- "crv3"
   }
   
-  #preprocessed_object <- preprocess
-  
+
   X <- preprocessed_object$X
   y <- preprocessed_object$Y
   R <- preprocessed_object$R0
@@ -68,6 +71,8 @@ boot_algo3 <- function(preprocessed_object,
   bootcluster <- preprocessed_object$bootcluster
   G <- N_G_bootcluster <- length(unique(bootcluster[[1]]))
   k <- length(R)
+
+  bootstrap_type <- paste0(substr(bootstrap_type, 1, 4), "x")
   
   v <- get_weights(
     type = type, 
@@ -76,30 +81,14 @@ boot_algo3 <- function(preprocessed_object,
     boot_iter = B
   )
   
-  # X1: X without parameter beta for which hypothesis beta = 0 is tested
-  X1 <- X[, which(R == 0)]
-  
   # create X_g's, X1_g's, y_g's etc 
   X_list <- matrix_split(X, cluster, "row")
-  X1_list <- matrix_split(X1, cluster, "row")
   y_list <- split(y, cluster, drop = FALSE)
-  
-  colnames(X_list)
   
   # precompute a range of other objects
   tXgXg <- lapply(
-    seq_along(1:G),
+    seq_along(1:G), 
     function(g) crossprod(X_list[[g]])
-  )
-  
-  tX1gX1g <- lapply(
-    seq_along(1:G),
-    function(g) crossprod(X1_list[[g]])
-  )
-  
-  tX1gyg <- lapply(
-    seq_along(1:G),
-    function(g) t(X1_list[[g]]) %*% y_list[[g]]
   )
   
   tXgyg <- lapply(
@@ -108,54 +97,94 @@ boot_algo3 <- function(preprocessed_object,
   )
   
   
-  tXgX1g <- lapply(
-    seq_along(1:G),
-    function(g) t(X_list[[g]]) %*% X1_list[[g]]
-  )
-  
   tXX <- Reduce("+", tXgXg) # crossprod(X)  
-  tX1X1 <- Reduce("+", tX1gX1g) # crossprod(X1)
-  tX1y <- Reduce("+", tX1gyg) #t(X1) %*% y
   tXy <- Reduce("+", tXgyg) # t(X) %*% y
-  
   tXXinv <- solve(tXX)
-  tX1X1inv <- solve(tX1X1)
   
-  # OLS estimate, ROLS estimate
-  beta_hat <- tXXinv %*% tXy # mean(c(beta_hat) - coef(object)) #essentially zero
-  beta_tilde <- beta_hat - tXXinv %*% R %*% solve(t(R) %*% tXXinv %*% R) %*% (R %*% beta_hat - 0)
   
-  # precompute more objects 
+  tXgX1g <- NULL 
+  beta_hat <- NULL
+  beta_tilde <- NULL
+  beta_g_hat <- NULL
+  beta_1g_tilde <- NULL
+  inv_tXX_tXgXg <- NULL
   
-  inv_tXX_tXgXg <- lapply(
-    1:G,
-    function(x) MASS::ginv(tXX - tXgXg[[x]])
-  )
   
-  beta_1g_tilde <- lapply(
-    1:G, 
-    function(g) MASS::ginv(tX1X1 - tX1gX1g[[g]]) %*% (tX1y - tX1gyg[[g]])
-  )
-  
[…]
