+++ DESCRIPTION
@@ -61,7 +61,8 @@ Imports:
     generics,
     dqrng, 
     JuliaConnectoR, 
-    summclust
+    summclust, 
+    sandwich
 Suggests: 
     fixest,
     lfe,
@@ -79,6 +80,7 @@ Suggests:
     bench,
     testthat (>= 3.0.0), 
     tibble
+Remotes: s3alfisc/summclust
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
