+++ tests/testthat/test-new-bootstrap-variants.R
@@ -2,7 +2,7 @@ test_that("new bootstrap variants", {
   
   # run Library(testthat) and library(fwildclusterboot) prior
   
-  B <- 9999
+  B <- 999
   type <- "mammen"
   N_G1 <- 21
   
