+++ DESCRIPTION
@@ -61,14 +61,11 @@ Imports:
     generics,
     dqrng, 
     JuliaConnectoR, 
-    summclust, 
-    sandwich
 Suggests: 
     fixest,
     lfe,
     ivreg,
     clubSandwich,
-    sandwich,
     lmtest,
     data.table,
     fabricatr,
@@ -80,7 +77,6 @@ Suggests:
     bench,
     testthat (>= 3.0.0), 
     tibble
-Remotes: s3alfisc/summclust
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
