+++ R/model_matrix.R
@@ -3,7 +3,8 @@ model_matrix <- function(object, ...) {
   #' enhanced model.matrix functionalities
   #' @param object An object of class `lm` or `felm``
   #' @param ... Other arguments
-  #' @export
+  #' @noRd
+
 
 
   UseMethod("model_matrix")
@@ -13,7 +14,6 @@ model_matrix <- function(object, ...) {
 model_matrix.lm <- function(object, collin.rm = TRUE, ...) {
   #' Enhanced model.matrix for objects of type lm
   #' @method model_matrix lm
-  #' @export
   #' @param object An object of class lm
   #' @param collin.rm Should collinear variables be dropped?
   #' @param ... Other arguments
@@ -31,7 +31,6 @@ model_matrix.lm <- function(object, collin.rm = TRUE, ...) {
 model_matrix.felm <- function(object, type, collin.rm = TRUE, ...) {
   #' Enhanced model.matrix for objects of type felm
   #' @method model_matrix felm
-  #' @export
   #' @param object An object of class felm
   #' @param collin.rm Should collinear variables be dropped?
   #' @param type 'rhs' for right-hand side variables, 'fixef' for fixed effects
@@ -64,7 +63,6 @@ model_matrix.fixest <-
   function(object, type, collin.rm = TRUE, ...) {
     #' Enhanced model.matrix for objects of type fixest
     #' @method model_matrix fixest
-    #' @export
     #' @param object An object of class fixest
     #' @param type rhs lhs or fixef
     #' @param collin.rm Should collinear variables be dropped?
