+++ NAMESPACE
@@ -30,6 +30,7 @@ export(setBoottest_boot_algo)
 export(teststat)
 importFrom(JuliaConnectoR,juliaEval)
 importFrom(JuliaConnectoR,juliaImport)
+importFrom(MASS,ginv)
 importFrom(Matrix,Diagonal)
 importFrom(Matrix,t)
 importFrom(Matrix.utils,aggregate.Matrix)
