+++ R/p_val_null.R
@@ -60,12 +60,10 @@ p_val_null2 <-
 
     denom <- suppressWarnings(sqrt(JJ_sum))
 
-    cat("ssc:", small_sample_correction, "\n")
-    cat("numer: ", numer[3], "\n")
-    cat("denom:", denom[3], "\n")
+    # cat("ssc:", small_sample_correction, "\n")
+    # cat("numer: ", numer[3], "\n")
+    # cat("denom:", denom[3], "\n")
 
-    numer_old <<- numer
-    
     t <- numer / denom
     t_boot <- t[2:(boot_iter + 1)]
 
