+++ tests/testthat/test-method_equivalence.R
@@ -26,89 +26,89 @@ test_that("Do different, but equivalent ways to specify
         data = data1
       )
     feols_fit1 <-
-      feols(
+      fixest::feols(
         proposition_vote ~ treatment + ideology1 + log_income +
           Q1_immigration + Q2_defense,
         data = data1
       )
     feols_fit2 <-
-      feols(
+     fixest::feols(
         proposition_vote ~ treatment + ideology1 + log_income +
           Q1_immigration + Q2_defense,
         data = data1,
         cluster = "group_id1"
       )
     feols_fit3 <-
-      feols(
+     fixest::feols(
         proposition_vote ~ treatment + ideology1 + log_income +
           Q1_immigration + Q2_defense,
         data = data1,
         cluster = ~group_id1
       )
     feols_fit4 <-
-      feols(
+     fixest::feols(
         proposition_vote ~ treatment + ideology1 + log_income +
           Q1_immigration + Q2_defense,
         data = data1,
         cluster = c("group_id1", "group_id2")
       )
     feols_fit5 <-
-      feols(
+     fixest::feols(
         proposition_vote ~ treatment + ideology1 + log_income +
           Q1_immigration + Q2_defense,
         data = data1,
         cluster = ~ group_id1 + group_id2
       )
     feols_fit6 <-
-      feols(
+     fixest::feols(
         proposition_vote ~ treatment + ideology1 + log_income |
           Q1_immigration + Q2_defense,
         data = data1
       )
     feols_fit7 <-
-      feols(
+     fixest::feols(
         proposition_vote ~ treatment + ideology1 + log_income |
           Q1_immigration + Q2_defense,
         data = data1,
         cluster = "group_id1"
       )
     feols_fit8 <-
-      feols(
+     fixest::feols(
         proposition_vote ~ treatment + ideology1 + log_income |
           Q1_immigration + Q2_defense,
         data = data1,
         cluster = ~group_id1
       )
     feols_fit9 <-
-      feols(
+     fixest::feols(
         proposition_vote ~ treatment + ideology1 + log_income |
           Q1_immigration + Q2_defense,
         data = data1,
         cluster = c("group_id1", "group_id2")
       )
     feols_fit10 <-
-      feols(
+     fixest::feols(
         proposition_vote ~ treatment + ideology1 + log_income |
           Q1_immigration + Q2_defense,
         data = data1,
         cluster = ~ group_id1 + group_id2
       )
     feols_fit11 <-
-      feols(
+     fixest::feols(
         proposition_vote ~ treatment + ideology1 + log_income +
           Q1_immigration |
           Q2_defense,
         data = data1
       )
     feols_fit12 <-
-      feols(
+     fixest::feols(
         proposition_vote ~ treatment + ideology1 + log_income +
           Q1_immigration,
         fixef = "Q2_defense",
         data = data1
       )
     feols_fit13 <-
-      feols(
+     fixest::feols(
         proposition_vote ~ treatment + ideology1 + log_income,
         fixef = c("Q1_immigration", "Q2_defense"),
         data = data1
@@ -117,52 +117,52 @@ test_that("Do different, but equivalent ways to specify
     # can also assign fixed effects via fixef = c("") arguments ...
     
     felm_fit1 <-
-      felm(
+     lfe::felm(
         proposition_vote ~ treatment + ideology1 + log_income +
           Q1_immigration + Q2_defense,
         data = data1
       )
     felm_fit2 <-
-      felm(
+     lfe::felm(
         proposition_vote ~ treatment + ideology1 + log_income +
           Q1_immigration + Q2_defense |
[…]
