+++ tests/testthat/test-r-lean.R
@@ -20,10 +20,10 @@ test_that("test lean cpp boottest", {
   lm_fit <- lm(proposition_vote ~ treatment + ideology1 + log_income,
     data = data1
   )
-  feols_fit <- feols(proposition_vote ~ treatment + ideology1 + log_income,
+  feols_fit <-fixest::feols(proposition_vote ~ treatment + ideology1 + log_income,
     data = data1
   )
-  felm_fit <- felm(proposition_vote ~ treatment + ideology1 + log_income,
+  felm_fit <-lfe::felm(proposition_vote ~ treatment + ideology1 + log_income,
     data = data1
   )
 
