### 2022-08-17 #546 @s3alfisc
Hi, I simply wanted to give some feedback on the state of the package:
- Over the last weeks, I have integrated new bootstrap variants to the package [as suggested in MacKinnon, Nielsen & Webb)[https://www.econ.queensu.ca/sites/econ.queensu.ca/files/wpaper/qed_wp_1485.pdf]
- I have also spent quite some time on trying to make the package compliant with all `pkgchecks` - unfortunately, I still fail to run the entire process locally, so please apologize if it will take me one or two more attempts to fully align the package.
- As suggested by @mpadge , I have added a clause that will prohibit that unit tests will run if Julia is not installed. Unfortunately, this has affected the CI statistics of the package by disabling *any* Julia based testing on CI. I am currently investigating this - the computed coverage of 56% does not look very good & is quite a bit higher when adding all tests against Julia
