### 2022-08-17 49d09ff
add pkgcheck action for commits to master
