### 2022-08-17 f69ea7b
temporarily do not check for julia inst before running test
