### 2022-08-17 9046d78
add fixest::ssc in tests
