### 2022-08-18 5ad8396
fix bug in random number generation for Rlean algo
