+++ R/methods.R
@@ -14,7 +14,7 @@
 #'  \link[fwildclusterboot]{boottest.ivreg}
 #'
 #' @examples
-#' library(fwildclusterboot)
+#' requireNamespace("fwildclusterboot")
 #' data(voters)
 #' lm_fit <- lm(
 #'   proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
@@ -134,7 +134,7 @@ boottest <- function(object,
 #' . No. 1315. Queen's Economics Department Working Paper, 2013.
 #' @examples
 #' \dontrun{
-#' library(clubSandwich)
+#' requireNamespace("clubSandwich")
 #' R <- clubSandwich::constrain_zero(2:3, coef(lm_fit))
 #' wboottest <-
 #'   mboottest(
@@ -164,7 +164,7 @@ mboottest <- function(object,
 #' @export
 #' 
 #' @examples
-#' library(fwildclusterboot)
+#' requireNamespace("fwildclusterboot")
 #' data(voters)
 #' lm_fit <- lm(
 #' proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
@@ -191,7 +191,7 @@ pval <- function(object,
 #'
 #' @export
 #' @examples
-#' library(fwildclusterboot)
+#' requireNamespace("fwildclusterboot")
 #' data(voters)
 #' lm_fit <- lm(
 #'   proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
@@ -217,7 +217,7 @@ teststat <- function(object,
 #' @return A vector containing the boundaries of the wild cluster
 #'  bootstrapped confidence interval
 #' @examples
-#' library(fwildclusterboot)
+#' requireNamespace("fwildclusterboot")
 #' data(voters)
 #' lm_fit <- lm(
 #'   proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
@@ -245,7 +245,7 @@ confint.boottest <- function(object, ...) {
 #' @return A vector containing the boundaries of the wild cluster
 #'  bootstrapped p-value
 #' @examples
-#' #' library(fwildclusterboot)
+#' #' requireNamespace("fwildclusterboot")
 #' data(voters)
 #' lm_fit <- lm(
 #'   proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
@@ -273,7 +273,7 @@ pval.boottest <- function(object, ...) {
 #' @return A vector containing the non-bootstrapped t-statistic
 #'  calculated in `boottest()`
 #' @examples
-#' library(fwildclusterboot)
+#' requireNamespace("fwildclusterboot")
 #' data(voters)
 #' lm_fit <- lm(
 #'   proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
@@ -301,7 +301,7 @@ teststat.boottest <- function(object, ...) {
 #' @return A scalar containing the effective number of observations
 #'  used in `boottest()`
 #' @examples
-#' library(fwildclusterboot)
+#' requireNamespace("fwildclusterboot")
 #' data(voters)
 #' lm_fit <- lm(
 #' proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
@@ -330,7 +330,7 @@ tidy.boottest <- function(object, ...) {
   #' @return A tidy data.frame with estimation results for objects of type
   #'         boottest
   #' @examples
-  #' library(fwildclusterboot)
+  #' requireNamespace("fwildclusterboot")
   #' data(voters)
   #' lm_fit <- lm(
   #' proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
[…]
