### 2022-08-18 8df41f0
Ãfix ssc() namespace error in test
