+++ tests/testthat/test_fixest_sugar.R
@@ -13,7 +13,7 @@ test_that("test fixest formula sugar", {
     seed = 908369,
     weights = 1:N / N
   )
-
+  
   fit <-
     fixest::feols(c(income, proposition_vote) ~ treatment, data = voters)
   res <-
@@ -26,7 +26,7 @@ test_that("test fixest formula sugar", {
         clustid = "group_id1"
       )
     )
-
+  
   fit <-
     fixest::feols(proposition_vote ~ csw(treatment, ideology1), data = voters)
   res <-
@@ -39,24 +39,28 @@ test_that("test fixest formula sugar", {
         clustid = "group_id1"
       )
     )
-
+  
   fit <-
-    fixest::feols(proposition_vote ~ fixest::i(treatment, ideology1), data = voters)
+    fixest::feols(
+      proposition_vote ~ fixest::i(treatment, ideology1), 
+      data = voters
+    )
+  
   res <-
     boottest(fit,
-      B = 999,
-      param = "fixest::treatment::0:ideology1",
-      clustid = "group_id1"
+             B = 999,
+             param = "fixest::treatment::0:ideology1",
+             clustid = "group_id1"
     )
-
-
+  
+  
   fit <- fixest::feols(
     proposition_vote ~ treatment,
     split = ~Q2_defense,
     cluster = ~group_id1,
     data = voters
   )
-
+  
   res <- lapply(
     fit,
     \(x) boottest(
@@ -66,7 +70,7 @@ test_that("test fixest formula sugar", {
       clustid = "group_id1"
     )
   )
-
+  
   expect_equal(unlist(lapply(res, function(x) {
     x$t_stat
   })),
