+++ R/boottest_lm.R
@@ -279,7 +279,10 @@ boottest.lm <- function(object,
   check_arg(tol, "numeric scalar GT{0}")
   check_arg(maxiter, "scalar integer GT{5}")
   check_arg(boot_ssc, "class(ssc) | class(boot_ssc)")
-  check_arg(boot_algo, "charin(R, R-lean, WildBootTests.jl, WCR13, WCR33, WCU13, WCU33, WCR11, WCR31, WCU11, WCU31)")
+  check_arg(boot_algo, "charin(R, R-lean, WildBootTests.jl)")
+  check_arg(
+    boot_algo, 
+    "charin(WCR13, WCR33, WCU13, WCU33, WCR11, WCR31, WCU11, WCU31)")
   check_arg(floattype, "charin(Float32, Float64)")
   check_arg(maxmatsize, "scalar integer | NULL")
   check_arg(bootstrapc, "scalar logical")
