+++ R/boottest_fixest.R
@@ -177,7 +177,7 @@
 #' The Review of Economics and Statistics 90.3 (2008): 414-427.
 #' @references Cameron, A.Colin & Douglas L. Miller.
 #' "A practitioner's guide to cluster-robust inference"
-#' Journal of Human Resources (2015) \doi{doi: 10.3368/jhr.50.2.317}
+#' Journal of Human Resources (2015) \doi{doi:10.3368/jhr.50.2.317}
 #' @references Davidson & MacKinnon. "Wild Bootstrap Tests for IV regression"
 #' Journal of Economics and Business Statistics (2010)
 #'  \doi{https://doi.org/10.1198/jbes.2009.07221}
