+++ tests/testthat/test-r-vs-julia.R
@@ -5,7 +5,8 @@ test_that("test r against Julia I: stochastic tests", {
   skip_on_ci()
 
 
-  if(is_juliaconnector_prepared()){
+  #if(is_juliaconnector_prepared()){
+  if(FALSE){
     reltol <- 0.05
     
     N <- 10000
