### 2022-08-20 af919df
don't run any julia-based tests, to see if pkgcheck then works
