+++ R/boottest_lm.R
@@ -338,7 +338,6 @@ boottest.lm <- function(object,
     )
   }
   
-  
   check_params_in_model(object = object, param = param)
   
   check_boottest_args_plus(
