### 2022-08-21 8e92d30
fix small bug for WCRx3 and WCUx3
