### 2022-08-21 442060b
fix two bugs in boot_algo3
