### 2022-08-21 9a1f1e0
fix another minor bug, add tests for t-stat equivalence
