+++ NAMESPACE
@@ -48,6 +48,7 @@ importFrom(dreamerr,warn_up)
 importFrom(generics,glance)
 importFrom(generics,tidy)
 importFrom(parallel,detectCores)
+importFrom(sandwich,vcovCL)
 importFrom(stats,aggregate)
 importFrom(stats,as.formula)
 importFrom(stats,coef)
@@ -68,4 +69,5 @@ importFrom(stats,terms)
 importFrom(stats,uniroot)
 importFrom(stats,update)
 importFrom(stats,weights)
+importFrom(summclust,vcov_CR3J)
 useDynLib(fwildclusterboot)
