### 2022-08-21 d09e3dc
minor cleanups
