### 2022-08-21 2101f67
enable julia tests
