+++ pkgdown.yml
@@ -2,9 +2,10 @@ pandoc: 2.7.3
 pkgdown: 2.0.6
 pkgdown_sha: ~
 articles:
+  Different-Variants-of-the-Wild-Cluster-Bootstrap: Different-Variants-of-the-Wild-Cluster-Bootstrap.html
   WildBootTests: WildBootTests.html
   fwildclusterboot: fwildclusterboot.html
-last_built: 2022-08-21T13:31Z
+last_built: 2022-08-21T15:29Z
 urls:
   reference: https://s3alfisc.github.io/fwildclusterboot/reference
   article: https://s3alfisc.github.io/fwildclusterboot/articles
