+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   Different-Variants-of-the-Wild-Cluster-Bootstrap: Different-Variants-of-the-Wild-Cluster-Bootstrap.html
   WildBootTests: WildBootTests.html
   fwildclusterboot: fwildclusterboot.html
-last_built: 2022-08-21T15:29Z
+last_built: 2022-08-21T15:33Z
 urls:
   reference: https://s3alfisc.github.io/fwildclusterboot/reference
   article: https://s3alfisc.github.io/fwildclusterboot/articles
