+++ README.Rmd
@@ -34,6 +34,8 @@ knitr::opts_chunk$set(
 
 The `{fwildclusterboot}` package provides a native R implementation of the fast wild cluster bootstrap algorithm developed in [Roodman et al (2019)](https://econpapers.repec.org/paper/qedwpaper/1406.htm) for regression objects in R.
 
+Additionally, it provides support for multiple wild cluster bootstrap variants as described in [MacKinnon, Nielsen & Webb (2022)](https://www.econ.queensu.ca/sites/econ.queensu.ca/files/wpaper/qed_wp_1485.pdf).
+
 It also ports functionality of [WildBootTests.jl](https://github.com/droodman/WildBootTests.jl) to R via the [JuliaConnectoR](https://github.com/stefan-m-lenz/JuliaConnectoR).
 
 The package's central function is `boottest()`. It allows to test univariate hypotheses using a wild cluster bootstrap at extreme speed: via the 'fast' algorithm, it is possible to run a wild cluster bootstrap with $B = 100.000$ iterations in less than a second!
@@ -139,4 +141,4 @@ If you are in `R`, you can simply run the following command to get the BibTeX ci
 citation("fwildclusterboot")
 ```
 
-Alternatively, if you prefer to cite the "Fast & Wild" paper by Roodman et al, I would be super happy if you mentioned `{fwildclusterboot}` in a footnote! 
+Alternatively, if you prefer to cite the "Fast & Wild" paper by Roodman et al, it would be great if you mentioned `{fwildclusterboot}` in a footnote =) ! 
