### 2022-08-23 47f2443
enable julia based tests
