+++ R/boot_algo3.R
@@ -197,13 +197,15 @@ boot_algo3 <- function(preprocessed_object,
     beta_1g_tilde = beta_1g_tilde
   )
   
-  
+  beta_2g_tilde <<- beta_1g_tilde
   # pre-allocate space for bootstrap 
   # start the bootstrap loop 
   t_boot <- vector(mode = "numeric", B + 1)
   se <- vector(mode = "numeric", B + 1)
   
   # numer <- (( R %*% tXXinv) %*%  (Reduce("cbind", scores_list) %*% v))
+  
+
   
   dim(R) <- c(1, k) # turn R into matrix
   
