### 2022-08-24 c99c914
review: add return values, increase coverage, skip failing test for new features
