+++ R/boot_ssc.R
@@ -22,6 +22,8 @@ boot_ssc <-
     #' @examples
     #' boot_ssc(adj = TRUE, cluster.adj = TRUE)
     #' boot_ssc(adj = TRUE, cluster.adj = TRUE, cluster.df = "min")
+    #' @return 
+    #' A list with encoded info on how to form small sample corrections
 
     dreamerr::check_arg_plus(adj, "loose logical scalar conv")
     dreamerr::check_arg_plus(fixef.K, "match(none, full)")
