### 2022-08-24 #546 @s3alfisc
I just ran the `pkgcheck` action in my own repo, and - even though the `pkgcheck` still complained about my compliance with standards - [it finished](https://github.com/s3alfisc/fwildclusterboot/issues/57)! This of course means that from now on, I will be able to test in my own repo and can stop polluting this space 😄
