### 2022-08-24 s3alfisc/fwildclusterboot#57 @s3alfisc
- `cmd check` fails because `pkgcheck` does not find the `JULIA_BINDIR` environment flag - the workflow of course does not install `Julia`, `WildBootTests.jl` etc, and it fails because I failed to add the clause that exludes the test if Julia is not found. I have changed this in a recent commit
- coverage is still low because I am excluding all Julia tests (except the one that fails)
- the concerned duplicated function / method names are `pval` and `teststat` - I like these function names and think it is in fact a benefit that they are used in other packages
- [the only remaining `library()` calls are in vignettes, readme and news files](https://github.com/s3alfisc/fwildclusterboot/search?q=library%28). Should I get rid of all of these and replace with `requireNamespace()`?
- when running `pkgcheck` locally, I only find [one line with more than 80 characters ](https://github.com/s3alfisc/fwildclusterboot/blob/aba1d7687835c2111c156668e3bbb61bfa373be3/R/boottest_lm.R#L282) - I have kept this line because otherwise, the argument checks via the `dreamerr` package would fail (dreamerr does not work nicely with linebreaks)
