### 2022-08-24 6e9afb9
add return values to two funs
