### 2022-08-24 25d6a9c
separate some r and julia tests so that coverage increases for ropensci submission
