+++ tests/testthat/test-tstat_equivalence.R
@@ -1,16 +1,4 @@
-test_that("t-stat equivalence OLS", {
-  
-  skip_on_cran()
-  #skip_on_ci()
-
-
-  if(is_juliaconnector_prepared()){
-  
-    ols_test <- function(run_this_test) {
-      if (run_this_test) {
-        # data(voters)
-        # adj <- cluster.adj <- TRUE; cluster.df <- "conventional";
-        # impose_null = TRUE
+test_that("t-stat equivalence OLS - R and R-lean", {
   
   
         lm_fit <-
@@ -36,7 +24,7 @@ test_that("t-stat equivalence OLS", {
         # impose_null= TRUE
         # boot_algo = "R"
   
-        for (boot_algo in c("R", "WildBootTests.jl")) {
+        for (boot_algo in c("R")) {
           for (adj in c(TRUE, FALSE)) {
             for (cluster.adj in c(TRUE, FALSE)) {
               for (cluster.df in c("conventional", "min")) {
@@ -353,17 +341,269 @@ test_that("t-stat equivalence OLS", {
             }
           }
         }
-      }
-    }
+        
   
-    ols_test(run_this_test = TRUE)
-  } else {
-    message("test-tstat_equivalence.R I skipped as JULIA_BINDR not found.")
-  }
+
 })
 
+
+test_that("t-stat equivalence OLS - WildBootTests", {
+  
+  skip_on_cran()
+  #skip_on_ci()
+  
+  
+  if(is_juliaconnector_prepared()){
+    
+        # data(voters)
+        # adj <- cluster.adj <- TRUE; cluster.df <- "conventional";
+        # impose_null = TRUE
+        
+        
+        lm_fit <-
+          lm(
+            proposition_vote ~ treatment + ideology1 + log_income,
+            data = fwildclusterboot:::create_data(
+              N = 1000,
+              N_G1 = 20,
+              icc1 = 0.81,
+              N_G2 = 10,
+              icc2 = 0.01,
+              numb_fe1 = 10,
+              numb_fe2 = 10,
+              seed = 970691
+            )
+          )
+        
+        B <- 999
+        
+        # adj= FALSE
+        # cluster.adj= FALSE
+        # cluster.df= "conventional"
+        # impose_null= TRUE
+        # boot_algo = "R"
+        
+        for (boot_algo in c("WildBootTests.jl")) {
+          for (adj in c(TRUE, FALSE)) {
+            for (cluster.adj in c(TRUE, FALSE)) {
+              for (cluster.df in c("conventional", "min")) {
+                for (impose_null in c(TRUE, FALSE)) {
+                  # cat("--------------------------------", "\n")
+                  #
+                  # cat("adj:", adj, "\n")
+                  # cat("cluster.adj:", cluster.adj, "\n")
+                  # cat("cluster.df:", cluster.df, "\n")
+                  # #cat("impose_null:", impose_null, "\n")
+                  #
+                  # cat("--------------------------------", "\n")
+                  # cat("oneway:", "\n")
+                  
+                  # oneway clustering
+                  feols_fit <-
+                    fixest::feols(
+                      proposition_vote ~ treatment + ideology1 + log_income,
+                      data = fwildclusterboot:::create_data(
+                        N = 1000,
+                        N_G1 = 20,
+                        icc1 = 0.81,
+                        N_G2 = 10,
+                        icc2 = 0.01,
+                        numb_fe1 = 10,
+                        numb_fe2 = 10,
+                        seed = 970691
+                      ),
+                      cluster = ~group_id1,
+                      ssc = fixest::ssc(
+                        adj = adj,
+                        cluster.adj = cluster.adj,
+                        cluster.df = cluster.df
+                      )
+                    )
+                  
+                  dof_tstat <-
+                    fixest::tstat(feols_fit)[c(
[…]
