+++ tests/testthat/test-full-enumeration.R
@@ -1,11 +1,9 @@
-test_that("test full enumeration cases", {
+test_that("test full enumeration cases: r and r-lean", {
   
   # note: these tests are deterministic! therefore exact.
   
-  skip_on_cran()
   # if not skipped, codecov fails after ~6h
 
-  if(is_juliaconnector_prepared()){
     reltol <- 0.05
     
     N <- 2000
@@ -95,19 +93,7 @@ test_that("test full enumeration cases", {
                 conf_int = FALSE
               )
             )
-            boot_jl <- suppressWarnings(
-              boottest(object, 
-                       clustid = "group_id1",
-                       B = 999, 
-                       param = "treatment", 
-                       type = type, 
-                       p_val_type = p_val_type, 
-                       impose_null = impose_null,
-                       boot_algo = "WildBootTests.jl", 
-                       conf_int = FALSE
-              )
-            )
-            
+
     # note: difference in p-values due to discrete jumps: 
     # t-statistics are not directly identical (discrepancies at order e-14)
     # therefore, the actual t-statistic t can be smaller, larger, or lie within
@@ -123,32 +109,18 @@ test_that("test full enumeration cases", {
     # I therefore set the tolerance to 1 / (2^N_G_bootcluster) for all p-values
     # see sum(sort(boot_r$t_boot) - sort(boot_jl$t_boot)) != 0L
             
+
             expect_true(
-              boot_r$p_val %in% (boot_jl$p_val +  -2:2 * 1 / 2^boot_r$N_G)
-            )
-            
-            expect_equal(
-              boot_r$t_stat, 
-              boot_jl$t_stat, 
-              ignore_attr = TRUE
-            )
-            expect_equal(
-              sort(boot_r$t_boot), 
-              sort(boot_jl$t_boot), 
-              ignore_attr = TRUE
-            )
-            
-            expect_true(
-              boot_r_lean$p_val %in% (boot_jl$p_val +  -2:2 * 1 / 2^boot_r$N_G)
+              boot_r_lean$p_val %in% (boot_r$p_val +  -2:2 * 1 / 2^boot_r$N_G)
             )
             expect_equal(
               boot_r_lean$t_stat,
-              boot_jl$t_stat, 
+              boot_r$t_stat, 
               ignore_attr = TRUE
             )
             expect_equal(
               sort(boot_r_lean$t_boot), 
-              sort(boot_jl$t_boot),
+              sort(boot_r$t_boot),
               ignore_attr = TRUE
             )
             
@@ -162,9 +134,7 @@ test_that("test full enumeration cases", {
             expect_true(
               (boot_r_lean$p_val * 2^boot_r_lean$N_G) %% 1 == 0
             )
-            expect_true(
-              (boot_jl$p_val * 2^boot_jl$N_G) %% 1 == 0
-            )
+        
             
             # multi-param hypotheses
             # cat("Check 2:", "\n")
@@ -198,6 +168,202 @@ test_that("test full enumeration cases", {
                 conf_int = FALSE
               )
             )
+            
+    
+            expect_true(
+              boot_r_lean$p_val %in% (boot_r$p_val +  -2:2 * 1 / 2^boot_r$N_G)
+            )
+            expect_equal(
+              boot_r_lean$t_stat,
+              boot_r$t_stat, 
+              ignore_attr = TRUE
+            )
+            expect_equal(
+              sort(boot_r_lean$t_boot),
+              sort(boot_r$t_boot), 
+              ignore_attr = TRUE
+            )
+            
+            # check: with 9 clusters, 2^9 unique draws - 
+            # hence the p-values calculated
+            # must be an integer if multiplied by 512
+            
+            expect_true(
[…]
