### 2022-08-24 96de679
install Julia etc in pkgcheck workflow and run one julia based test
