+++ tests/testthat/test-full-enumeration.R
@@ -211,7 +211,7 @@ test_that("test full enumeration cases: r and julia", {
   skip_on_cran()
   # if not skipped, codecov fails after ~6h
   
-  if(is_juliaconnector_prepared()){
+  # if(is_juliaconnector_prepared()){
     reltol <- 0.05
     
     N <- 2000
@@ -412,6 +412,6 @@ test_that("test full enumeration cases: r and julia", {
         }
       }
     }
-  }
+  #}
 })
 
