+++ .github/workflows/pkgcheck.yaml
@@ -18,4 +18,16 @@ jobs:
   pkgcheck:
     runs-on: ubuntu-latest
     steps:
+      # install julia
+      - uses: julia-actions/setup-julia@v1
+      # add julia to renviron
+      - name: Create and populate .Renviron file
+        run: echo JULIA_BINDIR= "${{ env.juliaLocation }}" >> ~/.Renviron
+        shell: bash
+      # install WildBootTests.jl
+      - name: install WildBootTests.jl 
+        run: julia -e 'using Pkg; Pkg.add("WildBootTests")'
+        # use shell bash to ensure consistent behavior across OS
+        shell: bash
+      # run pkgcheck action
       - uses: ropensci-review-tools/pkgcheck-action@main
