### 2022-08-27 1738403
rename boot_algo arg, introduce bootstrap_type arg
