### 2022-08-27 s3alfisc/fwildclusterboot#57 @assignUser
@s3alfisc You can change your [pkgcheck action file](https://github.com/s3alfisc/fwildclusterboot/blob/master/.github/workflows/pkgcheck.yaml) too install additional system dependencies like Julia (e.g. via an additional step using https://github.com/julia-actions/setup-julia)
We have an open issue to add documentation about that (with a small example on how to do it: https://github.com/ropensci-review-tools/pkgcheck-action/issues/23#issuecomment-1027035722
