### 2022-08-27 s3alfisc/fwildclusterboot#57 @s3alfisc
Hi @assignUser, thanks for your feedback! I had in fact [done this the other day ](https://github.com/s3alfisc/fwildclusterboot/commit/96de679b48dd5071be8bdf9587fe897cd6f15cf5) but then reverted the commit as the unit tests run for a while, and I first wanted to make sure that `pkgcheck()` runs without me fiddling with the yaml file. I'll try to keep an eye on the issue and would be happy to eventually add a Julia example to the documentation.
