### 2022-08-27 154bff8
add tests for new bootstrap variants
