+++ R/boottest_felm.R
@@ -374,14 +374,14 @@ boottest.felm <- function(object,
     heteroskedastic <- FALSE
   }
   
-  R <- process_R(
+  R_long <- process_R(
     R = R,
     param = param
   )
   
   if (engine != "WildBootTests.jl") {
     r_algo_checks(
-      R = R,
+      R = R_long,
       p_val_type = p_val_type,
       conf_int = conf_int,
       B = B
@@ -392,7 +392,7 @@ boottest.felm <- function(object,
   
   check_boottest_args_plus(
     object = object,
-    R = R,
+    R = R_long,
     param = param,
     sign_level = sign_level,
     B = B,
@@ -402,7 +402,7 @@ boottest.felm <- function(object,
   preprocess <- preprocess2.felm(
     object = object,
     clustid = clustid,
-    R = R,
+    R = R_long,
     param = param,
     bootcluster = bootcluster,
     fe = fe,
@@ -472,7 +472,9 @@ boottest.felm <- function(object,
         weights = stats::weights(object), 
         clustid = clustid,
         fe = fe,
-        bootstrap_type = bootstrap_type
+        bootstrap_type = bootstrap_type, 
+        R = R_long, 
+        r = r
       )
       
       res <- boot_algo3(
@@ -488,7 +490,8 @@ boottest.felm <- function(object,
         full_enumeration = full_enumeration,
         small_sample_correction = small_sample_correction,
         seed = internal_seed, 
-        object = object
+        object = object, 
+        impose_null = impose_null
       )
       conf_int <- p_grid_vals <- grid_vals <- FALSE
       
@@ -499,7 +502,7 @@ boottest.felm <- function(object,
     check_r_lean(
       weights = stats::weights(object),
       clustid = clustid,
-      fe = NULL, 
+      fe = fe, 
       impose_null = impose_null
     )
     
@@ -525,6 +528,8 @@ boottest.felm <- function(object,
     small <- julia_ssc$small
     clusteradj <- julia_ssc$clusteradj
     clustermin <- julia_ssc$clustermin
+    
+    fedfadj <- 0L
     
     if (ssc[["fixef.K"]] != "none") {
       message(paste("Currently, boottest() only supports fixef.K = 'none'."))
@@ -579,7 +584,7 @@ boottest.felm <- function(object,
     call = call,
     type = type,
     impose_null = impose_null,
-    R = R,
+    R = R_long,
     r = r,
     engine = engine,
     nthreads = nthreads,
