### 2022-08-27 93e54d1
add installation instructions for julia, wildboottests.jl in pkgcheck yaml
