### 2022-08-27 fe3d8c7
fix error in pkgcheck yaml
