+++ R/boottest_felm.R
@@ -57,7 +57,8 @@
 #' run. Options are "fnw11","11", "13", "31" and "33". For more information, see the
 #' details section. "fnw11" is the default, which runs a "11" type wild cluster 
 #' bootstrap via the algorithm outlined in "fast and wild" 
-#' (Roodman et al (2019)). #' @param seed An integer. Allows to set a random seed. For details, see below.
+#' (Roodman et al (2019)). 
+#' @param seed An integer. Allows to set a random seed. For details, see below.
 #' @param R Hypothesis Vector giving linear combinations of coefficients.
 #'  Must be either NULL or a vector of the same length as `param`. If NULL,
 #'   a vector of ones of length param.
@@ -194,9 +195,7 @@
 #'      (\url{https://journals.sagepub.com/doi/full/10.1177/1536867X19830877})
 #' @references MacKinnon, James G., Morten Ørregaard Nielsen, and 
 #' Matthew D. Webb. Fast and reliable jackknife and bootstrap
-#'  methods for cluster-robust inference. No. 1485. 2022. 
-#'  (\url{https://www.econ.queensu.ca
-#'  /sites/econ.queensu.ca/files/wpaper/qed_wp_1485.pdf})
+#' methods for cluster-robust inference. No. 1485. 2022. 
 #' @references Cameron, A. Colin, Jonah B. Gelbach, and Douglas L. Miller.
 #' "Bootstrap-based improvements for inference with clustered errors."
 #' The Review of Economics and Statistics 90.3 (2008): 414-427.
