+++ tests/testthat/test-new-bootstrap-variants.R
@@ -1,110 +1,129 @@
-test_that("new bootstrap variants I", {
-
-
-  B <- 999
-  type <- "mammen"
-  N_G1 <- 21
-
-  data2 <- fwildclusterboot:::create_data(N = 1000,
-                                          N_G1 = N_G1,
-                                          icc1 = 0.8,
-                                          N_G2 = N_G1,
-                                          icc2 = 0.8,
-                                          numb_fe1 = 10,
-                                          numb_fe2 = 5,
-                                          seed = 41224,
-                                          #seed = 123,
-                                          weights = 1:N / N)
-
-  object <- lm(
-    proposition_vote ~ treatment + ideology1 + log_income   ,
-    data = data2
-  )
-
-
-  # # new WCR11 ("fast and reliable") vs old WCR11 ("fast and wild")
-  #  <- "R"
-  # bootstrap_type <-  "11"
-  # impose_null <- TRUE
-
-  boot1 <- boottest(object,
-                   param = "log_income",
-                   clustid = c("group_id2"),
-                   B = B,
-                   impose_null = TRUE,
-                   engine = "R",
-                   bootstrap_type = "fnw11",
-                   seed = 432,
-                   type = type,
-                   conf_int = FALSE,
-                   ssc = boot_ssc(adj = FALSE, cluster.adj = FALSE)
-  )
-
-  boot2 <- boottest(object,
-                   param = "log_income",
-                   clustid = c("group_id2"),
-                   B = B,
-                   impose_null = TRUE,
-                   engine = "R",
-                   bootstrap_type = "11",
-                   seed = 432,
-                   type = type,
-                   conf_int = FALSE,
-                   ssc = boot_ssc(adj = FALSE, cluster.adj = FALSE)
-  )
-
-  expect_equal(
-    teststat(boot1), teststat(boot2), ignore_attr = TRUE
-  )
-  expect_equal(
-    pval(boot1), pval(boot2), ignore_attr = TRUE
-  )
-  expect_equal(
-    boot1$t_boot, boot2$t_boot, ignore_attr = TRUE
-  )
-
-
-  # new WCU11 ("fast and reliable") vs old WCR11 ("fast and wild")
-
-
-  boot1 <- boottest(object,
-                    param = "log_income",
-                    clustid = c("group_id2"),
-                    B = B,
-                    impose_null = FALSE,
-                    bootstrap_type = "11",
-                    engine = "R",
-                    seed = 432,
-                    type = type,
-                    conf_int = FALSE,
-                    ssc = boot_ssc(adj = FALSE, cluster.adj = FALSE)
-  )
-
-  boot2 <- boottest(object,
-                    param = "log_income",
-                    clustid = c("group_id2"),
-                    B = B,
-                    impose_null = FALSE,
-                    engine = "R",
-                    bootstrap_type = "fnw11",
-                    seed = 432,
-                    type = type,
-                    conf_int = FALSE,
-                    ssc = boot_ssc(adj = FALSE, cluster.adj = FALSE)
-  )
-
-  expect_equal(
-    teststat(boot1), teststat(boot2), ignore_attr = TRUE
-  )
-  expect_equal(
-    pval(boot1), pval(boot2), ignore_attr = TRUE
-  )
-  expect_equal(
-    boot1$t_boot, boot2$t_boot, ignore_attr = TRUE
-  )
-
-
+test_that("test r-fnw vs r-, stochastic", {
+  
+      reltol <- 0.05
+      B <- 999
+      
+      data1 <<- fwildclusterboot:::create_data(
+        N = 10000,
+        N_G1 = 20,
+        icc1 = 0.5,
+        N_G2 = 20,
[…]
