### 2022-08-27 89513ed
run some julia based tests via pkgcheck & covr
