+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   Different-Variants-of-the-Wild-Cluster-Bootstrap: Different-Variants-of-the-Wild-Cluster-Bootstrap.html
   WildBootTests: WildBootTests.html
   fwildclusterboot: fwildclusterboot.html
-last_built: 2022-08-21T16:08Z
+last_built: 2022-08-28T10:08Z
 urls:
   reference: https://s3alfisc.github.io/fwildclusterboot/reference
   article: https://s3alfisc.github.io/fwildclusterboot/articles
