### 2022-08-28 15766fd
vignette cleanup
