+++ man/reexports.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/methods.R
+\docType{import}
+\name{reexports}
+\alias{reexports}
+\alias{tidy}
+\title{Objects exported from other packages}
+\keyword{internal}
+\description{
+These objects are imported from other packages. Follow the links
+below to see their documentation.
+
+\describe{
+  \item{generics}{\code{\link[generics]{tidy}}, \code{\link[generics]{tidy}}}
+}}
+
