+++ NAMESPACE
@@ -28,6 +28,7 @@ export(mboottest)
 export(pval)
 export(setBoottest_engine)
 export(teststat)
+export(tidy)
 importFrom(JuliaConnectoR,juliaEval)
 importFrom(JuliaConnectoR,juliaImport)
 importFrom(MASS,ginv)
