### 2022-09-07 64e9fff
add explicit error when feols run with sunab
