+++ man/reexports.Rd
@@ -4,6 +4,7 @@
 \name{reexports}
 \alias{reexports}
 \alias{tidy}
+\alias{glance}
 \title{Objects exported from other packages}
 \keyword{internal}
 \description{
@@ -11,6 +12,6 @@ These objects are imported from other packages. Follow the links
 below to see their documentation.
 
 \describe{
-  \item{generics}{\code{\link[generics]{tidy}}, \code{\link[generics]{tidy}}}
+  \item{generics}{\code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}
 }}
 
