+++ NAMESPACE
@@ -24,6 +24,7 @@ S3method(tidy,boottest)
 S3method(tidy,mboottest)
 export(boot_ssc)
 export(boottest)
+export(glance)
 export(mboottest)
 export(pval)
 export(setBoottest_engine)
