+++ R/RcppExports.R
@@ -1,6 +1,18 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
+boot_algo3_crv1 <- function(B, G, k, v, scores_mat, scores_boot, tXXinv, Ag, ssc, cores, R) {
+    .Call('_fwildclusterboot_boot_algo3_crv1', PACKAGE = 'fwildclusterboot', B, G, k, v, scores_mat, scores_boot, tXXinv, Ag, ssc, cores, R)
+}
+
+get_c_all_boot_cpp <- function(coef_boot, shares, B, cores) {
+    .Call('_fwildclusterboot_get_c_all_boot_cpp', PACKAGE = 'fwildclusterboot', coef_boot, shares, B, cores)
+}
+
+get_se_all_boot_cpp <- function(V_boot, s, B, cores) {
+    .Call('_fwildclusterboot_get_se_all_boot_cpp', PACKAGE = 'fwildclusterboot', V_boot, s, B, cores)
+}
+
 #' Implementation of the heteroskedastic wild bootstrap. Computes
 #' HC robust variance estimators. For use in fwildclusterboot when no
 #' cluster variable is provided
