+++ R/bootagg.R
@@ -1,0 +1,314 @@
+bootagg = function(
+    x,
+    agg,
+    full = FALSE,
+    use_weights = TRUE,
+    B = 999, 
+    bootstrap_type = "11", 
+    clustid,
+    nthreads = 1,
+    ...){
+  
+  #' Aggregates the value of coefficients. Inference via a wild cluster
+  #' bootstrap 
+  #' Most of this function is written by Laurent Bergé and used 
+  #' in the fixest package published under GPL-3, 
+  #' https://cran.r-project.org/web/packages/fixest/index.html
+  #' minor changes by Alexander Fischer
+  #' 
+  #' @importFrom dreamerr check_value_plus
+  #' @examples 
+  #' 
+  #' #' library(fixest)
+  #' #' library(fwildclusterboot)
+  #' 
+  #' set.seed(123456L)
+  #' 
+  #' # code by Grant McDermott
+  #' # 60 time periods, 30 individuals, and 5 waves of treatment
+  #' tmax = 22; imax = 500; nlvls = 5
+  #' 
+  #' dat =
+  #'   expand.grid(time = 1:tmax, id = 1:imax) |>
+  #'   within({
+  #' 
+  #'     cohort      = NA
+  #'     effect      = NA
+  #'     first_treat = NA
+  #' 
+  #'     for (chrt in 1:imax) {
+  #'       cohort = ifelse(id==chrt, sample.int(nlvls, 1), cohort)
+  #'     }
+  #' 
+  #'     for (lvls in 1:nlvls) {
+  #'       effect      = ifelse(cohort==lvls, sample(2:10 , 1), effect)
+  #'       first_treat = ifelse(cohort==lvls, sample(1:(tmax+20), 1), first_treat)
+  #'     }
+  #' 
+  #'     first_treat = ifelse(first_treat>tmax, Inf, first_treat)
+  #'     treat       = time>=first_treat
+  #'     rel_time    = time - first_treat
+  #'     y           = id + time + ifelse(treat, effect*rel_time, 0) + rnorm(imax*tmax)
+  #' 
+  #'     rm(chrt, lvls, cohort, effect)
+  #'   })
+  #' 
+  #' head(dat)
+  #' dim(dat)
+  #' length(unique(dat$time))
+  #' dat <- na.omit(dat)
+  #' 
+  #' object = feols(
+  #'   y ~ sunab(first_treat, rel_time) + id + time,
+  #'   data = dat, vcov = ~id
+  #' )
+  #' 
+  #' # run - else error in expand.data.frame
+  #' sunab <- fixest:::sunab
+  #' res <-
+  #'   fwildclusterboot:::bootagg(
+  #'     object,
+  #'     agg = TRUE,
+  #'     full = FALSE,
+  #'     use_weights = TRUE,
+  #'     B = 9999,
+  #'     bootstrap_type = "11",
+  #'     clustid = ~ time,
+  #'     nthreads = 8
+  #'   )
+  #' 
+  #' 
+  #' res
+  #' sunab_p <- pvalue(summary(object, agg = TRUE))
+  #' nparams <- length(sunab_p)
+  #' cbind(
+  #'   res,
+  #'   round(sunab_p[-nparams][-(nparams-1)][-1], 3)
+  #' )
+
+
+  check_arg(x, "class(fixest) mbt")
+  if(isTRUE(x$is_sunab)){
+    check_arg(agg, "scalar(character, logical)")
+  } else {
+    check_arg(agg, "character scalar")
+  }
+
+  check_arg(full, "logical scalar")
+  # => later => extend it to more than one set of vars to agg
+
+  dots = list(...)
+  from_summary = isTRUE(dots$from_summary)
+
+  no_agg = FALSE
+  agg_rm = NULL
[…]
