### 2022-09-10 a684d0c
support for confidence intervals for bootagg
