+++ R/bootagg.R
@@ -1,94 +1,134 @@
 bootagg = function(
     x,
     agg,
+    clustid = NULL, 
+    bootcluster = "max",
     full = FALSE,
     use_weights = TRUE,
     B = 999, 
-    bootstrap_type = "11", 
-    clustid,
+    impose_null = TRUE,
+    fe = NULL,
     nthreads = 1,
     alpha = 0.05,
     ...){
   
-  #' Aggregates the value of coefficients. Inference via a wild cluster
-  #' bootstrap 
+  #' Simple tool that aggregates the value of CATT coefficients in
+  #' staggered difference-in-difference setups with inference based on 
+  #' a wild cluster bootstrap (see details) - similar to `fixest::aggregate()`
+  #' 
+  #' This is a function helping to replicate the estimator from Sun and
+  #' Abraham (2021, Journal of Econometrics). You first need to perform 
+  #' an estimation with cohort and relative periods dummies 
+  #' (typically using the function i), this leads to estimators of the 
+  #' cohort average treatment effect on the treated (CATT). Then you can
+  #' use this function to retrieve the average treatment effect on each 
+  #' relative period,or for any other way you wish to aggregate the CATT.
+  #' 
+  #' Note that contrary to the SA article, here the cohort share 
+  #' in the sample is considered to be a perfect measure for the 
+  #' cohort share in the population.
+  
+  
   #' Most of this function is written by Laurent Bergé and used 
   #' in the fixest package published under GPL-3, 
   #' https://cran.r-project.org/web/packages/fixest/index.html
   #' minor changes by Alexander Fischer
   #' 
+  #' @param x An object of type fixest estimated using `sunab()`
+  #' @param agg A character scalar describing the variable names to be aggregated, it is pattern-based. All variables that match the pattern will be aggregated. It must be of the form `"(root)"`, the parentheses must be there and the resulting variable name will be `"root"`. You can add another root with parentheses: `"(root1)regex(root2)"`, in which case the resulting name is `"root1::root2"`. To name the resulting variable differently you can pass a named vector: `c("name" = "pattern")` or `c("name" = "pattern(root2)")`. It's a bit intricate sorry, please see the examples.
+  #' @param full Logical scalar, defaults to `FALSE`. If `TRUE`, then all coefficients are returned, not only the aggregated coefficients.
+  #' @param use_weights Logical, default is `TRUE`. If the estimation was weighted, whether the aggregation should take into account the weights. Basically if the weights reflected frequency it should be `TRUE`.
+  #' @param clustid A character vector or rhs formula containing the names of the
+  #' cluster variables. If NULL,
+  #'        a heteroskedasticity-robust (HC1) wild bootstrap is run.
+  #' @param B Integer. The number of bootstrap iterations. When the number of
+  #'  clusters is low,
+  #'        increasing B adds little additional runtime.  
+  #' @param fe A character vector or rhs formula of length one which contains
[…]
