+++ src/boot_algo3_cpp.cpp
@@ -56,197 +56,100 @@ List boot_algo3_crv1(
 }
 
 
-
-// 
-// // This is a simple example of exporting a C++ function to R. You can
-// // source this function into an R session using the Rcpp::sourceCpp
-// // function (or via the Source button on the editor toolbar). Learn
-// // more about Rcpp at:
-// //
-// //   http://www.rcpp.org/
-// //   http://adv-r.had.co.nz/Rcpp.html
-// //   http://gallery.rcpp.org/
-// //
-// 
 // // [[Rcpp::export]]
-// List boot_algo3cpp(
-//     const int G,
-//     const int B,
-//     arma::mat X,
-//     arma::mat scores_list2,
-//     arma::mat RtXXinv,
-//     Rcpp::List Ag, 
-//     arma::mat v, 
-//     const int ssc, 
-//     const int crv_type, 
-//     const List inv_tXX_tXgXg, 
-//     arma::vec R) {
+// List boot_algo3_crv3( const int B,
+//                       const int G,
+//                       const int k,
+//                       arma::mat v,
+//                       arma::mat scores_mat,
+//                       arma::mat scores_boot,
+//                       arma::cube inv_tXX_tXgXg,
+//                       int cores,
+//                       arma::mat R, 
+//                       arma::mat delta_b_star){
 // 
-//   int k = X.n_cols;
+//   arma::vec se(B+1);
+//   arma::cube boot_vcov(k, k, B+1);
 //   
-//   arma::vec delta_b_star(B);
-//   arma::vec se(B);
+//   double ssc = (G-1) / G;
 // 
-//   if(crv_type == 0){
-//     
-//     for(int b = 0; b < B; b++){
+//   for(int b = 0; b < B +1; b++){
+// 
+//     arma::mat delta_diff(G,k);
+// 
+//     for(int g = 0; g < G; g++){
 //       
-//       // Step 1: calculate numerator
+//       arma::vec scores_g_boot = scores_mat.col(g) * v(g, b);
 //       
-//       arma::mat scores_g_boot(G, k);
-//       arma::mat scores_boot(1,k);
-//       
-//       for(int g = 0; g < G; g++){
-//         for (int j = 0; j < k; j++){
-//           scores_g_boot(g, j) = scores_list2(j,g) * v(g,b);
-//         }
-//         scores_boot += scores_g_boot.row(g);
-//       }
-//       
-//       delta_b_star(b) = arma::as_scalar(RtXXinv * scores_boot.t());
-//       
-//       // 2 denominator (variance) calculation
-//       
-//       arma::mat Sigma(k,k);
-//       arma::mat score_hat_g_boot(1,k);
-//       // arma::mat Sigma;
-//       
-//       for(int g = 0; g < G; g++){
-//         score_hat_g_boot = scores_g_boot.row(g).t() - as<arma::mat>(Ag[g]) * scores_boot.t();
-//         Sigma += score_hat_g_boot * score_hat_g_boot.t();
-//       }
-//       
-//       se(b) = sqrt(arma::as_scalar(ssc * RtXXinv * Sigma * RtXXinv.t()));
-//       
+//       arma::mat score_diff = scores_boot.col(b) - scores_g_boot;
+//       delta_diff.row(g) = (inv_tXX_tXgXg.slice(g) * score_diff
+//          - delta_b_star.col(g)).t();
 //     }
 //     
-//   } else if(crv_type == 1){
-//     
-//     // Step 1: calculate numerator
-//   
-//     
-//     
[…]
