+++ R/preprocess2.R
@@ -676,308 +676,6 @@ preprocess2.ivreg <-
   }
 
 
-get_cluster <-
-  function(object,
-           clustid_char,
-           bootcluster,
-           N,
-           call_env) {
-    #' function creates a data.frame with cluster variables
-    #'
-    #' @param object An object of type lm, fixest, felm or ivreg
-    #' @param clustid_char the name of the cluster variable(s) as
-    #' a character vector
-    #' @param bootcluster the name of the bootcluster variable(s)
-    #'  as a character vector, or "min" or "max" for multiway clustering
-    #' @param N the number of observations used in the bootstrap
-    #' @param call_env the environment in which the 'object' was evaluated
-    #'
-    #' @noRd
-    #'
-    #' @return a list, containing, among other things, a data.frame of the
-    #'  cluster variables,
-    #'         a data.frame of the bootcluster variable(s), and a helper
-    #'         matrix, all_c, used in `engine_julia()`
-
-    # ----------------------------------------------------------------------- #
-    # Note: a large part of the following code was taken and adapted from the
-    # sandwich R package, which is distributed under GPL-2 | GPL-3
-    # Zeileis A, Köll S, Graham N (2020). "Various Versatile Variances:
-    # An object-Oriented Implementation of Clustered Covariances in R."
-    # _Journal of Statistical Software_,  *95*(1), 1-36.
-    # doi: 10.18637/jss.v095.i01 (URL: https://doi.org/10.18637/jss.v095.i01).
-
-    # changes by Alexander Fischer:
-    # no essential changes, but slight reorganization of pieces of code
-
-    dreamerr::check_arg(clustid_char, "character scalar|charakter vector")
-    dreamerr::check_arg(bootcluster, "character scalar | character vector")
-
-    clustid_fml <- reformulate(clustid_char)
-    # Step 1: create cluster df
-
-    cluster_tmp <-
-      try(if ("Formula" %in% loadedNamespaces()) {
-        ## FIXME to suppress potential warnings due to | in Formula
-        suppressWarnings(
-          expand.model.frame(
-            model = object,
-            extras = clustid_fml,
-            na.expand = FALSE,
-            envir = call_env
-          )
-        )
-      } else {
-        expand.model.frame(object,
-          clustid_fml,
-          na.expand = FALSE,
-          envir = call_env
-        )
-      })
-
-    if(inherits(cluster_tmp, "try-error")){
-      if(inherits(object, "fixest") || inherits(object, "felm")){
-        if(grepl("non-numeric argument to binary operator$", 
-                 attr(cluster_tmp, "condition")$message)){
-          stop("In your model, you have specified multiple fixed effects,
-               none of which are of type factor. While `fixest::feols()` and
-               `lfe::felm()` handle this case without any troubles,  
-               `boottest()` currently cannot handle this case - please 
-               change the type of (at least one) fixed effect(s) to factor.
-               If this does not solve the error, please report the issue
-               at https://github.com/s3alfisc/fwildclusterboot.")
-        }
-        if(grepl("operations are possible only for numeric, logical
-                 or complex types$", 
-                 attr(cluster_tmp, "condition")$message)){
-          stop("Either a fixed effect or a cluster variable in your fixest()
-               or felm() model is currently specified as a character. 
-               'boottest()' relies on 'expand.model.frame',
-               which can not handle these variable types in models.
-               Please change these character variables to factors. ")
-        }
-      }
-    }
-    
-    cluster_df <-
[…]

[…]
