### 2022-09-11 f34e0e2
bug fixes in cluster preprocessing
