+++ R/boot_algo3.R
@@ -1,3 +1,368 @@
+#' boot_algo3 <- function(preprocessed_object, 
+#'                        B, 
+#'                        bootstrap_type, 
+#'                        r = 0, 
+#'                        sign_level, 
+#'                        param,
+#'                        p_val_type,
+#'                        nthreads,
+#'                        type,
+#'                        full_enumeration,
+#'                        small_sample_correction,
+#'                        seed, 
+#'                        object, 
+#'                        impose_null){
+#'   
+#'   #' WRE13, WRE33, WRU13 and WRU33 bootstraps as in MNW (2022) "Fast and 
+#'   #' reliable" 
+#'   #'
+#'   #' @param preprocessed_object A list: output of the preprocess2 function.
+#'   #' @param B number of bootstrap iterations
+#'   #' @param r Shifts the null hypothesis.
+#'   #' @param sign_level The significance level.
+#'   #' @param param name of the test parameter.
+#'   #' @param p_val_type type Type of p-value. By default "two-tailed".
+#'   #' Other options: "equal-tailed", ">", "<"
+#'   #' @param nthreads The number of threads. Can be: a) an integer lower than,
+#'   #'                 or equal to, the maximum number of threads; b) 0: meaning
+#'   #'                 all available threads will be used; c) a number strictly
+#'   #'                 between 0 and 1 which represents the fraction of all
+#'   #'                 threads to use. The default is to use 50\% of all
+#'   #'                 threads. You can set permanently the number of threads
+#'   #'                 used within this package using the function ...
+#'   #' @param type character or function. The character string specifies the
+#'   #'        type of boostrap to use: One of "rademacher", "mammen", "norm"
+#'   #'        and "webb". Alternatively, type can be a function(n) for drawing
+#'   #'        wild bootstrap factors. "rademacher" by default.
+#'   #' @param full_enumeration Is full enumeration employed? Full enum.
+#'   #' is used if N_G^2 < B for Mammen and Rademacher weights
+#'   #' @param small_sample_correction The small sample correction to be applied.
+#'   #' See ssc().
+#'   #' If FALSE, run wild cluster bootstrap.
+#'   #' @param seed Integer scalar. Either set via boottest()'s seed argument
+#'   #' or inherited from R's global seed (set via set.seed)
+#'   #' @param object the regression object
+#'   #' @param impose_null logical scalar. Should the null be imposed on the
+#'   #' bootstrap dgp or not?
+#'   #' @importFrom MASS ginv
+#'   #' @importFrom sandwich vcovCL
+#'   #' @importFrom summclust vcov_CR3J
+#'   #' @return A list of ...
+#'   #' @noRd
+#'   
+#'   #check_arg(bootstrap_type, "charin(WCR13, WCR33, WCU13, WCU33)")
[…]
