+++ NAMESPACE
@@ -74,4 +74,6 @@ importFrom(stats,uniroot)
 importFrom(stats,update)
 importFrom(stats,weights)
 importFrom(summclust,vcov_CR3J)
+importFrom(utils,setTxtProgressBar)
+importFrom(utils,txtProgressBar)
 useDynLib(fwildclusterboot)
