+++ tests/testthat/test_multicollinearity.R
@@ -5,16 +5,16 @@ test_that("test multicollinearity", {
   X1 <- rnorm(N)
   X2 <- X1
   Y <- rnorm(N)
-  data <- data.frame(
+  data2 <- data.frame(
     Y = Y,
     X1 = X1,
     X2 = X2,
     clustid = sample(1:10, N, TRUE)
   )
 
-  lm_fit <- lm(Y ~ X1 + X2, data = data)
-  feols_fit <- fixest::feols(Y ~ X1 + X2, data = data)
-  felm_fit <- lfe::felm(Y ~ X1 + X2, data = data)
+  lm_fit <- lm(Y ~ X1 + X2, data = data2)
+  suppressWarnings(feols_fit <- fixest::feols(Y ~ X1 + X2, data = data2))
+  suppressWarnings(felm_fit <- lfe::felm(Y ~ X1 + X2, data = data2))
 
   boot_lm <-
     boottest(
