+++ tests/testthat/test_multicollinearity.R
@@ -4,17 +4,17 @@ test_that("test multicollinearity", {
   N <- 1000
   X1 <- rnorm(N)
   X2 <- X1
-  Y <- rnorm(N)
-  data2 <- data.frame(
-    Y = Y,
+  depvar <- rnorm(N)
+  data2 <<- data.frame(
+    depvar = depvar,
     X1 = X1,
     X2 = X2,
     clustid = sample(1:10, N, TRUE)
   )
 
-  lm_fit <- lm(Y ~ X1 + X2, data = data2)
-  suppressWarnings(feols_fit <- fixest::feols(Y ~ X1 + X2, data = data2))
-  suppressWarnings(felm_fit <- lfe::felm(Y ~ X1 + X2, data = data2))
+  lm_fit <- lm(depvar ~ X1 + X2, data = data2)
+  suppressWarnings(feols_fit <- fixest::feols(depvar ~ X1 + X2, data = data2))
+  suppressWarnings(felm_fit <- lfe::felm(depvar ~ X1 + X2, data = data2))
 
   boot_lm <-
     boottest(
