+++ R/bootagg.R
@@ -1,16 +1,33 @@
-bootagg = function(
+boot_aggregate = function(
     x,
     agg,
-    clustid = NULL, 
-    bootcluster = "max",
     full = FALSE,
     use_weights = TRUE,
-    B = 999, 
-    impose_null = TRUE,
+    clustid = NULL, 
+    B,
+    bootcluster = "max",
     fe = NULL,
+    sign_level = 0.05,
+    seed = NULL,
+    beta0 = NULL,
+    type = "rademacher",
+    impose_null = TRUE,
+    bootstrap_type = "fnw11",
+    p_val_type = "two-tailed",
     nthreads = getBoottest_nthreads(),
-    alpha = 0.05,
-    engine = "R",
+    tol = 1e-06, 
+    maxiter = 10,
+    ssc = boot_ssc(
+      adj = TRUE,
+      fixef.K = "none",
+      cluster.adj = TRUE,
+      cluster.df = "conventional"
+    ),
+    engine = getBoottest_engine(),
+    floattype = "Float64",
+    maxmatsize = FALSE,
+    bootstrapc = FALSE,
+    getauxweights = FALSE,
     ...){
   
   #' Simple tool that aggregates the value of CATT coefficients in
@@ -28,33 +45,37 @@ bootagg = function(
   #' Note that contrary to the SA article, here the cohort share 
   #' in the sample is considered to be a perfect measure for the 
   #' cohort share in the population.
-  
-  
+  #' 
   #' Most of this function is written by Laurent Bergé and used 
   #' in the fixest package published under GPL-3, 
   #' https://cran.r-project.org/web/packages/fixest/index.html
   #' minor changes by Alexander Fischer
   #' 
   #' @param x An object of type fixest estimated using `sunab()`
-  #' @param agg A character scalar describing the variable names to be aggregated, it is pattern-based. All variables that match the pattern will be aggregated. It must be of the form `"(root)"`, the parentheses must be there and the resulting variable name will be `"root"`. You can add another root with parentheses: `"(root1)regex(root2)"`, in which case the resulting name is `"root1::root2"`. To name the resulting variable differently you can pass a named vector: `c("name" = "pattern")` or `c("name" = "pattern(root2)")`. It's a bit intricate sorry, please see the examples.
-  #' @param full Logical scalar, defaults to `FALSE`. If `TRUE`, then all coefficients are returned, not only the aggregated coefficients.
-  #' @param use_weights Logical, default is `TRUE`. If the estimation was weighted, whether the aggregation should take into account the weights. Basically if the weights reflected frequency it should be `TRUE`.
+  #' @param agg A character scalar describing the variable names to be
+  #'  aggregated, it is pattern-based. All variables that match the pattern 
+  #'  will be aggregated. It must be of the form `"(root)"`, the parentheses 
+  #'  must be there and the resulting variable name will be `"root"`. You 
+  #'  can add another root with parentheses: `"(root1)regex(root2)"`, in
+  #'  which case the resulting name is `"root1::root2"`. To name the resulting 
+  #'  variable differently you can pass a named vector: `c("name" = "pattern")`
+  #'  or `c("name" = "pattern(root2)")`. It's a bit intricate sorry, please
+  #'  see the examples.
+  #' 
+  #' @param full Logical scalar, defaults to `FALSE`. If `TRUE`, then all 
+  #' coefficients are returned, not only the aggregated coefficients.
+  #' @param use_weights Logical, default is `TRUE`. If the estimation was 
+  #' weighted, whether the aggregation should take into account the weights.
[…]
