+++ src/boot_algo3_cpp.cpp
@@ -10,6 +10,8 @@
 
 using namespace Rcpp;
 
+// function returns CRV1 bootstrapped vcov matrix and standard errors
+
 // [[Rcpp::export]]
 List boot_algo3_crv1(
     const int B,
@@ -43,7 +45,7 @@ List boot_algo3_crv1(
     }
     
     boot_vcov.slice(b) = ssc * tXXinv * score_hat_boot * tXXinv;
-    se(b) = arma::as_scalar(R * boot_vcov.slice(b) * R.t());
+    se(b) = sqrt(arma::as_scalar(R * boot_vcov.slice(b) * R.t()));
   }
   
   List res_list; 
@@ -56,52 +58,54 @@ List boot_algo3_crv1(
 }
 
 
-// // [[Rcpp::export]]
-// List boot_algo3_crv3( const int B,
-//                       const int G,
-//                       const int k,
-//                       arma::mat v,
-//                       arma::mat scores_mat,
-//                       arma::mat scores_boot,
-//                       arma::cube inv_tXX_tXgXg,
-//                       int cores,
-//                       arma::mat R, 
-//                       arma::mat delta_b_star){
-// 
-//   arma::vec se(B+1);
-//   arma::cube boot_vcov(k, k, B+1);
-//   
-//   double ssc = (G-1) / G;
-// 
-//   for(int b = 0; b < B +1; b++){
-// 
-//     arma::mat delta_diff(G,k);
-// 
-//     for(int g = 0; g < G; g++){
-//       
-//       arma::vec scores_g_boot = scores_mat.col(g) * v(g, b);
-//       
-//       arma::mat score_diff = scores_boot.col(b) - scores_g_boot;
-//       delta_diff.row(g) = (inv_tXX_tXgXg.slice(g) * score_diff
-//          - delta_b_star.col(g)).t();
-//     }
-//     
-//     boot_vcov.slice(b) = ssc * delta_diff.t() * delta_diff;
-//     se(b) = arma::as_scalar(sqrt(ssc * R * boot_vcov.slice(b) * R.t()));
-// 
-// 
-//   }
-//   
-//   List res_list;
-//   res_list["boot_vcov"] = boot_vcov;
-//   res_list["se"] = se;
-//   
-//   return res_list; 
-//   
-// 
-// 
-// 
-// }
+// [[Rcpp::export]]
+
+List boot_algo3_crv3( const int B,
+                      const int G,
+                      const int k,
+                      arma::mat v,
+                      arma::mat scores_mat,
+                      arma::mat scores_boot,
+                      arma::cube inv_tXX_tXgXg,
+                      int cores,
+                      arma::mat R,
+                      arma::mat delta_b_star){
+
+  arma::vec se(B+1);
+  arma::cube boot_vcov(k, k, B+1);
+
+  double ssc = (G-1) / G;
+
+#pragma omp parallel for num_threads(cores)
+  for(int b = 0; b < B +1; b++){
+
+    arma::mat delta_diff(G,k);
+
+    for(int g = 0; g < G; g++){
+
+      arma::vec scores_g_boot = scores_mat.col(g) * v(g, b);
+
+      arma::mat score_diff = scores_boot.col(b) - scores_g_boot;
+      delta_diff.row(g) = (inv_tXX_tXgXg.slice(g) * score_diff
+         - delta_b_star.col(g)).t();
+    }
+
+    boot_vcov.slice(b) = ssc * delta_diff.t() * delta_diff;
+    se(b) = arma::as_scalar(sqrt(ssc * R * boot_vcov.slice(b) * R.t()));
+
+
+  }
+
+  List res_list;
+  res_list["boot_vcov"] = boot_vcov;
+  res_list["se"] = se;
+
+  return res_list;
+
+
+
+
+}
 
 
 /*** R
