+++ NAMESPACE
@@ -22,8 +22,8 @@ S3method(teststat,boottest)
 S3method(teststat,mboottest)
 S3method(tidy,boottest)
 S3method(tidy,mboottest)
+export(boot_aggregate)
 export(boot_ssc)
-export(bootagg)
 export(boottest)
 export(glance)
 export(mboottest)

[…]
