### 2022-09-18 977d245
rename boot_algo3 funs
