+++ man/boot_aggregate.Rd
@@ -1,11 +1,40 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/bootagg.R
+% Please edit documentation in R/boot_aggregate.R, R/bootagg.R
 \name{boot_aggregate}
 \alias{boot_aggregate}
 \title{Simple tool that aggregates the value of CATT coefficients in
 staggered difference-in-difference setups with inference based on
 a wild cluster bootstrap (see details) - similar to \code{fixest::aggregate()}}
 \usage{
+boot_aggregate(
+  x,
+  agg,
+  full = FALSE,
+  use_weights = TRUE,
+  clustid = NULL,
+  B,
+  bootcluster = "max",
+  fe = NULL,
+  sign_level = 0.05,
+  seed = NULL,
+  beta0 = NULL,
+  type = "rademacher",
+  impose_null = TRUE,
+  bootstrap_type = "fnw11",
+  p_val_type = "two-tailed",
+  nthreads = getBoottest_nthreads(),
+  tol = 1e-06,
+  maxiter = 10,
+  ssc = boot_ssc(adj = TRUE, fixef.K = "none", cluster.adj = TRUE, cluster.df =
+    "conventional"),
+  engine = getBoottest_engine(),
+  floattype = "Float64",
+  maxmatsize = FALSE,
+  bootstrapc = FALSE,
+  getauxweights = FALSE,
+  ...
+)
+
 boot_aggregate(
   x,
   agg,
@@ -176,8 +205,19 @@ coefficient(s) for which the hypothesis is to be tested}
 \value{
 A data frame with aggregated coefficients, p-values and
 confidence intervals.
+
+A data frame with aggregated coefficients, p-values and
+confidence intervals.
 }
 \description{
+This is a function helping to replicate the estimator from Sun and
+Abraham (2021, Journal of Econometrics). You first need to perform
+an estimation with cohort and relative periods dummies
+(typically using the function i), this leads to estimators of the
+cohort average treatment effect on the treated (CATT). Then you can
+use this function to retrieve the average treatment effect on each
+relative period,or for any other way you wish to aggregate the CATT.
+
 This is a function helping to replicate the estimator from Sun and
 Abraham (2021, Journal of Econometrics). You first need to perform
 an estimation with cohort and relative periods dummies
@@ -195,8 +235,63 @@ Most of this function is written by Laurent Bergé and used
 in the fixest package published under GPL-3,
 https://cran.r-project.org/web/packages/fixest/index.html
 minor changes by Alexander Fischer
+
+Note that contrary to the SA article, here the cohort share
+in the sample is considered to be a perfect measure for the
+cohort share in the population.
+
+Most of this function is written by Laurent Bergé and used
+in the fixest package published under GPL-3,
+https://cran.r-project.org/web/packages/fixest/index.html
+minor changes by Alexander Fischer
 }
 \examples{
+
+library(fwildclusterboot)
+library(fixest)
+library(did2s)
+data("df_het", package = "did2s")
+object = feols(
+  dep_var ~ sunab(year - rel_year, year) | state + year, 
+  df_het, 
+  cluster = ~state
+)
+
+#sunab <- fixest:::sunab
+
+agg1 = fwildclusterboot:::boot_aggregate(
+  x = object,
+  agg = TRUE,
+  full = FALSE,
+  use_weights = TRUE,
+  B = 9999, 
+  clustid = ~state,
+  nthreads = 1,
+  sign_level = 0.05,
+  impose_null = TRUE,
+  fe = ~ state
+)
+
+agg1 |> head()
[…]
