### 2022-09-18 11f2d4b
enable all julia based tests
