+++ tests/testthat/test-r-vs-julia.R
@@ -1,730 +1,730 @@
-# test_that("test r against Julia I: stochastic tests", {
-#   
-#   
-#   skip_on_cran()
-#   skip_on_ci()
-# 
-# 
-#   if(is_juliaconnector_prepared()){
-#     reltol <- 0.05
-#     
-#     N <- 10000
-#     seed <- 875784
-#     
-#     data1 <<- fwildclusterboot:::create_data(
-#       N = 10000,
-#       N_G1 = 20,
-#       icc1 = 0.5,
-#       N_G2 = 20,
-#       icc2 = 0.2,
-#       numb_fe1 = 10,
-#       numb_fe2 = 10,
-#       seed = 908369,
-#       weights = 1:N / N
-#     )
-#     
-#     lm_fit <- lm(proposition_vote ~ treatment + log_income,
-#                  data = data1
-#     )
-#     
-#     
-#     
-#     lm_fit_weights <- lm(
-#       proposition_vote ~ treatment + log_income,
-#       weights = data1$weights,
-#       data = data1
-#     )
-#     lm_fits <- list(ols = lm_fit, wls = lm_fit_weights)
-#     
-#     # object = lm_fit
-#     # impose_null = FALSE
-#     # type = "rademacher"
-#     # p_val_type = "two-tailed"
-#     
-#     
-#     cat("Part 1: Large B Tests", "\n")
-#     
-#     
-#     for (object in lm_fits) {
-#       cat("start ols/wls", "\n")
-#       set.seed(12391786)
-#       # type <- "rademacher"
-#       for (type in c("rademacher", "webb", "mammen", "norm")) {
-#         for (p_val_type in c("two-tailed", "equal-tailed", ">", "<")) {
-#           for (impose_null in c(TRUE, FALSE)) {
-#             cat(
-#               paste(
-#                 "type:",
-#                 type,
-#                 "p-val:",
-#                 p_val_type,
-#                 "null imposed:",
-#                 impose_null
-#               ),
-#               "\n"
-#             )
-#             
-#             cat(type, "\n")
-#             cat(p_val_type, "\n")
-#             cat(impose_null, "\n")
-#             
-#             # cat("Check 1:", "\n")
-#             if (p_val_type %in% c("two-tailed", "equal-tailed")) {
-#               boot_r <-
-#                 suppressWarnings(
-#                   boottest(
-#                     object,
-#                     clustid = "group_id1",
-#                     B = 19999,
-#                     param = "treatment",
-#                     type = type,
-#                     p_val_type = p_val_type,
-#                     impose_null = impose_null
-#                   )
-#                 )
-#               # pracma::tic()
-#               boot_jl <-
-#                 suppressWarnings(
-#                   boottest(
-#                     object,
-#                     clustid = "group_id1",
-#                     B = 19999,
-#                     param = "treatment",
-#                     type = type,
-#                     p_val_type = p_val_type,
-#                     impose_null = impose_null,
-#                     engine = "WildBootTests.jl"
-#                   )
-#                 )
-#               # pracma::toc()
-#               res <-
-#                 expect_equal(
-#                   boot_r$p_val,
-#                   boot_jl$p_val[1],
-#                   tolerance = reltol,
-#                   ignore_attr = TRUE
-#                 )
-#               # if(res == FALSE){print(res)}
-#               rm(res)
-#               res <-
-#                 expect_equal(
-#                   boot_r$t_stat,
-#                   boot_jl$t_stat[1],
-#                   tolerance = reltol,
-#                   ignore_attr = TRUE
-#                 )
-#               # if(res == FALSE){print(res)}
-#               rm(res)
-#               res <-
-#                 expect_equal(
-#                   boot_r$conf_int,
-#                   c(boot_jl$conf_int),
[…]
