### 2022-09-19 94019e7
revert to testing all julia based tests on CI
