+++ R/boottest_felm.R
@@ -531,7 +531,10 @@ boottest.felm <- function(object,
     fedfadj <- 0L
     
     if (ssc[["fixef.K"]] != "none") {
-      message(paste("Currently, boottest() only supports fixef.K = 'none'."))
+      x <- format_message(
+        "Currently, boottest() only supports fixef.K = 'none'."
+      )
+      message(x)
     }
     
     res <- boot_algo_julia(
