+++ R/boottest_fixest.R
@@ -546,7 +546,10 @@ boottest.fixest <- function(object,
     fedfadj <- 0L
     
     if (ssc[["fixef.K"]] != "none") {
-      message(paste("Currently, boottest() only supports fixef.K = 'none'."))
+      x <- format_message(
+        "Currently, boottest() only supports fixef.K = 'none'."
+        )
+      message(x)
     }
     
     res <- boot_algo_julia(
