+++ tests/testthat/test-full-enumeration.R
@@ -211,7 +211,7 @@ test_that("test full enumeration cases: r and julia", {
   skip_on_cran()
   # skip_on_ci()
   # if not skipped, codecov fails after ~6h
-  julia_prep <- is_juliaconnector_prepared()
+  julia_prep <- TRUE
   cat("julia connector prepared? ", julia_prep)
   
   if(julia_prep){
