+++ tests/testthat/test-new-bootstrap-variants.R
@@ -240,7 +240,7 @@ test_that("variants 31 R vs Julia", {
   skip_on_cran()
   skip_on_ci()
 
-  if(fwildclusterboot:::is_juliaconnector_prepared()){
+  if(TRUE){
 
     # fully enumerated - deterministic - tests
     N_G1 <- 10
