+++ NAMESPACE
@@ -25,6 +25,10 @@ S3method(tidy,mboottest)
 export(boot_aggregate)
 export(boot_ssc)
 export(boottest)
+export(format_alert)
+export(format_error)
+export(format_message)
+export(format_warning)
 export(glance)
 export(mboottest)
 export(pval)
