### 2022-09-19 15d57cf
some cmd check fixes
