### 2022-09-21 080599c
cmd check fixes
