+++ man/format_message.Rd
@@ -65,16 +65,6 @@ creates a \code{message()}. \code{format_warning()} and \code{format_error()} ch
 default type of exception to \code{warning()} and \code{stop()}, respectively.
 }
 \details{
-There is an experimental formatting feature implemented in this function.
-You can use following tags:
-\itemize{
-\item \verb{\{.b text\}} for bold formatting
-\item \verb{\{.i text\}} to use italic font style
-\item \verb{\{.url www.url.com\}} formats the string as URL (i.e., enclosing URL in
-\code{<} and \code{>}, blue color and italic font style)
-\item \verb{\{.pkg packagename\}} formats the text in blue color.
-}
-
 This features has some limitations: it's hard to detect the exact length for
 each line when the string has multiple lines (after line breaks) and the
 string contains formatting tags. Thus, it can happen that lines are wrapped at
@@ -83,35 +73,6 @@ format tag (\verb{\{.b one two three\}}), a line break might occur inside this t
 and the formatting no longer works (messing up the message-string).
 }
 \examples{
-msg <- format_message("Much too long string for just one line, I guess!",
-  line_length = 15
-)
-message(msg)
-
-msg <- format_message("Much too long string for just one line, I guess!",
-  "First new line",
-  "Second new line",
-  "(both indented)",
-  line_length = 30
-)
-message(msg)
-
-msg <- format_message("Much too long string for just one line, I guess!",
-  "First new line",
-  "Second new line",
-  "(not indented)",
-  line_length = 30,
-  indent = ""
-)
-message(msg)
-
-# Caution, experimental! See 'Details'
-msg <- format_message(
-  "This is {.i italic}, visit {.url easystats.github.io/easystats}",
-  line_length = 30
-)
-message(msg)
-
 \dontrun{
   format_alert("This is a message.")
   format_alert("This is a warning.", type = "warning")
