+++ NAMESPACE
@@ -6,6 +6,7 @@ S3method(boottest,ivreg)
 S3method(boottest,lm)
 S3method(confint,boottest)
 S3method(glance,boottest)
+S3method(glance,mboottest)
 S3method(mboottest,felm)
 S3method(mboottest,fixest)
 S3method(mboottest,lm)

[…]
