### 2022-09-22 ca42ec0
dont run mbootest.glance examples
