+++ R/methods.R
@@ -582,18 +582,21 @@ glance.mboottest <- function(x, ...) {
   #' @return A single row summary "glance" of an object of type boottest
   #'         - lists characteristics of the input regression model
   #' @examples
+  #' \dontrun{
   #' requireNamespace("fwildclusterboot")
   #' data(voters)
   #' lm_fit <- lm(
   #' proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
   #'   data = voters
   #' )
-  #' boot <- mboottest(lm_fit,
-  #'   B = 9999,
-  #'   param = "treatment",
-  #'   clustid = "group_id1"
+  #' mboot <- mboottest(
+  #'     object = lm_fit,
+  #'     clustid = "group_id1",
+  #'     B = 999,
+  #'     R = R
   #' )
   #' generics::glance(mboot)
+  #' }
   
   stopifnot(inherits(x, "mboottest"))
   broom::glance(eval(x$call$object))
