+++ man/glance.mboottest.Rd
@@ -19,16 +19,19 @@ A single row summary "glance" of an object of type boottest
 S3 method to glance at objects of class boottest
 }
 \examples{
+\dontrun{
 requireNamespace("fwildclusterboot")
 data(voters)
 lm_fit <- lm(
 proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
   data = voters
 )
-boot <- mboottest(lm_fit,
-  B = 9999,
-  param = "treatment",
-  clustid = "group_id1"
+mboot <- mboottest(
+    object = lm_fit,
+    clustid = "group_id1",
+    B = 999,
+    R = R
 )
 generics::glance(mboot)
+}
 }
