### 2022-09-23 69b11e3
ÃÃinstall julia etc in pkgcheck wfl
