+++ tests/testthat/test-global_vars.R
@@ -1,6 +1,8 @@
 test_that("global engine", {
   
   skip_on_cran()
+  skip_on_ci()
+  
 
   if(TRUE){
     data(voters)
