+++ tests/testthat/test-tstat_equivalence.R
@@ -347,6 +347,7 @@ test_that("t-stat equivalence OLS - R and R-lean", {
 test_that("t-stat equivalence OLS - WildBootTests", {
   
   skip_on_cran()
+  skip_on_ci()
 
   julia_prep <- is_juliaconnector_prepared()
   cat("julia connector prepared? ", julia_prep, "\n")
@@ -602,6 +603,7 @@ test_that("t-stat equivalence OLS - WildBootTests", {
 test_that("t-stat equivalence OLS q > 1", {
   
   skip_on_cran()
+  skip_on_ci()
   
   julia_prep <- is_juliaconnector_prepared()
   cat("julia connector prepared? ", julia_prep, "\n")
@@ -888,6 +890,7 @@ test_that("t-stat equivalence OLS q > 1", {
 test_that("t-stat equivalence IV", {
   
   skip_on_cran()
+  skip_on_ci()
 
   julia_prep <- is_juliaconnector_prepared()
   cat("julia connector prepared? ", julia_prep, "\n")
