+++ tests/testthat/test-method_equivalence.R
@@ -1,7 +1,9 @@
 test_that("Do different, but equivalent ways to specify
           linear models lead to equivalent results?", {
+  
   skip_on_cran()
-
+  skip_on_ci()
+            
   requireNamespace("fixest")
   requireNamespace("lfe")
             
