+++ tests/testthat/test_tidy.R
@@ -1,6 +1,7 @@
 test_that("test tidiers with q = 1", {
 
   skip_on_cran()
+  skip_on_ci()
   
   if(TRUE){
 
