### 2022-09-24 bfb73a1
more changes to get pkgcheck going
