+++ tests/testthat/test-error_warning.R
@@ -1463,6 +1463,8 @@ test_that("errors and warnings q = 1", {
 test_that("error warning IV/WRE and q > 1", {
   
   skip_on_cran()
+  skip_on_ci()
+  
   requireNamespace("ivreg")
   requireNamespace("fixest")
   requireNamespace("clubSandwich")
