### 2022-09-24 bc70436
further changes to accommodate pkgcheck and goodpractice
