+++ R/mboottest_lm.R
@@ -261,7 +261,8 @@ mboottest.lm <- function(object,
   clustermin <- julia_ssc$clustermin
 
   if (ssc[["fixef.K"]] != "none") {
-    format_message(paste("Currently, boottest() only supports fixef.K = 'none'."))
+    format_message(
+      paste("Currently, boottest() only supports fixef.K = 'none'."))
   }
 
   res <- boot_algo_julia(
