+++ R/boottest_ivreg.R
@@ -326,7 +326,9 @@ boottest.ivreg <- function(object,
   clustermin <- julia_ssc$clustermin
   
   if (ssc[["fixef.K"]] != "none") {
-    format_message(paste("Currently, boottest() only supports fixef.K = 'none'."))
+    format_message(
+      paste(
+        "Currently, boottest() only supports fixef.K = 'none'."))
   }
   
   res <- boot_algo_julia(
