### 2022-09-24 ad9a34a
further gp changes
