### 2022-09-25 7dbe9da
significant speedups for x1 bootstrap types
