+++ R/RcppExports.R
@@ -1,8 +1,8 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-boot_algo3_crv1 <- function(B, G, k, v, scores_mat, scores_boot, tXXinv, Ag, ssc, cores, R) {
-    .Call('_fwildclusterboot_boot_algo3_crv1', PACKAGE = 'fwildclusterboot', B, G, k, v, scores_mat, scores_boot, tXXinv, Ag, ssc, cores, R)
+boot_algo3_crv1_denom <- function(B, G, ssc, H, Cg, v, cores) {
+    .Call('_fwildclusterboot_boot_algo3_crv1_denom', PACKAGE = 'fwildclusterboot', B, G, ssc, H, Cg, v, cores)
 }
 
 boot_algo3_crv3 <- function(B, G, k, v, scores_mat, scores_boot, inv_tXX_tXgXg, cores, R, delta_b_star) {
