+++ src/boot_algo3_cpp.cpp
@@ -10,52 +10,89 @@
 
 using namespace Rcpp;
 
-// function returns CRV1 bootstrapped vcov matrix and standard errors
-
+// calculate the denominator for bootsraps of type x1 (that rely)
+// on CRV1 covariance matrices
 // [[Rcpp::export]]
-List boot_algo3_crv1(
-    const int B,
-    const int G, 
-    const int k, 
-    arma::mat v, 
-    arma::mat scores_mat, 
-    arma::mat scores_boot, 
-    arma::mat tXXinv, 
-    arma::cube Ag, 
-    double ssc, 
-    int cores, 
-    arma::mat R
-    ){
-
-  arma::vec se(B+1);
-  arma::cube boot_vcov(k, k, B+1);
+arma::vec boot_algo3_crv1_denom(int B, 
+                                int G, 
+                                double ssc, 
+                                arma::mat H, 
+                                arma::vec Cg, 
+                                arma::mat v, 
+                                int cores){
+  
+  arma::vec denom(B+1);
   
 #pragma omp parallel for num_threads(cores)
   for(int b = 0; b < (B+1); b++){
-
-    arma::mat score_hat_boot(k, k); 
     
+    //double Z_sq;
+    arma::vec Zg(G);
     for(int g = 0; g < G; g++){
-
-      arma::vec scores_g_boot = scores_mat.col(g) * v(g, b);
-      arma::mat score_hat_boot_sq = scores_g_boot - 
-        Ag.slice(g) * scores_boot.col(b);
-      score_hat_boot += score_hat_boot_sq * score_hat_boot_sq.t();
-
+      double vH = 0; 
+      for(int h = 0; h < G; h++){
+        vH += v(h, b) * H(g, h);
+      }
+      Zg(g) = Cg(g) * v(g, b) - vH;
     }
     
-    boot_vcov.slice(b) = ssc * tXXinv * score_hat_boot * tXXinv;
-    se(b) = sqrt(arma::as_scalar(R * boot_vcov.slice(b) * R.t()));
+    denom(b) = ssc * arma::as_scalar(sum(pow(Zg, 2)));
+    
   }
   
-  List res_list; 
-  res_list["boot_vcov"] = boot_vcov; 
-  res_list["se"] = se; 
+  return denom;
   
-  return res_list; 
-
-
 }
+
+// // [[Rcpp::export]]
+// List boot_algo3_crv1(
+//     const int B,
+//     const int G, 
+//     const int k, 
+//     arma::mat v, 
+//     arma::mat scores_mat, 
+//     arma::vec scores_mat2, 
+//     arma::mat scores_boot, 
+//     arma::mat tXXinv, 
+//     arma::cube Ag, 
+//     double ssc, 
+//     int cores, 
+//     arma::mat R
+// ){
+//   
+//   arma::vec se(B+1);
+//   // arma::cube boot_vcov(k, k, B+1);
+//   arma::mat RtXXinv = R * tXXinv;
+//   
+// #pragma omp parallel for num_threads(cores)
+//   for(int b = 0; b < (B+1); b++){
+//     
+//     arma::mat score_hat_boot; 
+//     
+//     for(int g = 0; g < G; g++){
+//       
+//       arma::vec scores_g_boot = scores_mat.col(g) * v(g, b);
+//       double scores_g_boot2 = scores_mat2(g) * v(g, b);
+//       
+//       arma::mat score_hat_boot_sq = pow( 
+//         (scores_g_boot2 - Ag.slice(g) * scores_boot.col(b)),2);
[…]

[…]
