+++ NEWS.md
@@ -1,4 +1,11 @@
-# fwildclusterboot 0.11.1 (development version)
+# fwildclusterboot 0.11.2 (development version)
+
++ significant speed improvements for the x1 bootstrap algorithms,
+  `bootstrap_type %in% c("11", "31")`, both for WCR and WCU
+
+
+
+# fwildclusterboot 0.11.1 
 
 ### New bootstrap algorithms following MNW (2022)
 
