### 2022-09-26 e7d6d88
delete skip_on_ci flags; pkgcheck
