+++ .github/workflows/pkgcheck.yaml
@@ -1,34 +1,0 @@
-name: pkgcheck
-
-# This will cancel running jobs once a new run is triggered
-concurrency:
-  group: ${{ github.workflow }}-${{ github.head_ref }}
-  cancel-in-progress: true
-
-on:
-  # Manually trigger the Action under Actions/pkgcheck
-  workflow_dispatch:
-  # Run on every push to main
-  push:
-    branches:
-      - main
-      - master
-
-jobs:
-  pkgcheck:
-    runs-on: ubuntu-latest
-    steps:
-      # # install julia
-      # - uses: julia-actions/setup-julia@v1
-      # # add julia to renviron
-      # - name: Create and populate .Renviron file
-      #   run: echo JULIA_BINDIR= "${{ env.juliaLocation }}" >> ~/.Renviron
-      #   shell: bash
-      # 
-      # # install WildBootTests.jl
-      # - name: install WildBootTests.jl 
-      #   run: julia -e 'using Pkg; Pkg.add(["WildBootTests", "StableRNGs"])'
-      #   # use shell bash to ensure consistent behavior across OS
-      #   shell: bash
-
-      - uses: ropensci-review-tools/pkgcheck-action@main
