### 2022-09-27 9a5f3f4
Ãadd some tests, documentation, update articles
