+++ R/boottest_fixest.R
@@ -253,6 +253,34 @@
 #' pval(boot1)
 #' confint(boot1)
 #' generics::tidy(boot1)
+#' 
+#' # run different bootstrap types following MacKinnon, Nielsen & Webb (2022):
+#' 
+#' # default: the fnw algorithm
+#' boot_fnw11 <- boottest(lm_fit,
+#'   B = 9999,
+#'   param = "treatment",
+#'   clustid = "group_id1", 
+#'   bootstrap_type = "fnw11"
+#' )
+#'
+#' # WCR 31 
+#'boot_WCR31 <- boottest(lm_fit,
+#'   B = 9999,
+#'   param = "treatment",
+#'   clustid = "group_id1",
+#'   bootstrap_type = "31"
+#' )
+#'
+#' # WCU33 
+#'boot_WCR31 <- boottest(lm_fit,
+#'   B = 9999,
+#'   param = "treatment",
+#'   clustid = "group_id1",
+#'   bootstrap_type = "33", 
+#'   impose_null = FALSE
+#' )
+#' 
 #' }
 #'
 boottest.fixest <- function(object,
@@ -374,6 +402,10 @@ boottest.fixest <- function(object,
     heteroskedastic <- FALSE
   }
   
+  check_bootstrap_types(
+    param = param, 
+    bootstrap_type = bootstrap_type
+  )
   
   R_long <- process_R(
     R = R,
@@ -476,7 +508,7 @@ boottest.fixest <- function(object,
       
       
     } else {
-    
+      
       # need some function checks here ... 
       check_boot_algo3(
         weights = stats::weights(object), 
@@ -548,7 +580,7 @@ boottest.fixest <- function(object,
     if (ssc[["fixef.K"]] != "none") {
       x <- format_message(
         "Currently, boottest() only supports fixef.K = 'none'."
-        )
+      )
       message(x)
     }
     
