+++ tests/testthat/test-new-bootstrap-variants.R
@@ -44,7 +44,7 @@ test_that("test r-fnw vs r-, stochastic", {
 
             # test the wcr
             boot1 <- boottest(object,
-                              param = "log_income",
+                              param = c("treatment", "log_income"),
                               clustid = c("group_id2"),
                               B = B,
                               impose_null = TRUE,
@@ -57,7 +57,7 @@ test_that("test r-fnw vs r-, stochastic", {
             )
     
             boot2 <- boottest(object,
-                              param = "log_income",
+                              param = c("treatment", "log_income"),
                               clustid = c("group_id2"),
                               B = B,
                               impose_null = TRUE,
@@ -149,12 +149,12 @@ test_that("new bootstrap variants II - t_stat equivalence", {
     seed = 123121,
     weights = 1:N / N
   )
-
+  
   lm_fit <- lm(
     proposition_vote ~ treatment + log_income,
     data = data
   )
-
+  
   # WCR
   wcr_algos <- wcu_algos <- c(
     "fnw11",
@@ -165,9 +165,9 @@ test_that("new bootstrap variants II - t_stat equivalence", {
   )
   p_val <- t_stat <-
     list()
-
+  
   for(x in wcr_algos){
-
+    
     cat(x)
     res <-
       suppressWarnings(
@@ -185,29 +185,29 @@ test_that("new bootstrap variants II - t_stat equivalence", {
           )
         )
       )
-
+    
     p_val[[x]] <- pval(res)
     t_stat[[x]] <- teststat(res)
-
+    
   }
-
+  
   df <- data.frame(
     "p_values" = unlist(p_val),
     "t_statistics" = unlist(t_stat)
   )
-
+  
   expect_equal(df$t_statistics[1], df$t_statistics[2])
   expect_equal(df$t_statistics[2], df$t_statistics[4])
   expect_equal(df$t_statistics[3], df$t_statistics[5])
-
-
+  
+  
   # WCU algos
-
+  
   p_val <- t_stat <-
     list()
-
+  
   for(x in wcu_algos){
-
+    
     res <-
       suppressWarnings(
         boottest(
@@ -224,37 +224,37 @@ test_that("new bootstrap variants II - t_stat equivalence", {
           )
         )
       )
-
+    
     p_val[[x]] <- pval(res)
     t_stat[[x]] <- teststat(res)
-
+    
   }
-
+  
   df <- data.frame(
     "p_values" = unlist(p_val),
     "t_statistics" = unlist(t_stat)
   )
-
+  
   expect_equal(df$t_statistics[1], df$t_statistics[2])
   expect_equal(df$t_statistics[2], df$t_statistics[4])
   expect_equal(df$t_statistics[3], df$t_statistics[5])
-
+  
 })
 
 
 
 
 test_that("variants 31 R vs Julia", {
-
+  
   skip_on_cran()
-  skip_on_ci()
-
+  
+  
   if(TRUE){
-
+    
     # fully enumerated - deterministic - tests
     N_G1 <- 10
     B <- 9999
-
+    
     data2 <- fwildclusterboot:::create_data(N = 1000,
                                             N_G1 = N_G1,
                                             icc1 = 0.8,
@@ -265,26 +265,26 @@ test_that("variants 31 R vs Julia", {
                                             seed = 41224,
                                             #seed = 123,
                                             weights = 1:N / N)
-
-
-
+    
+    
+    
     lm_fit <- lm(proposition_vote ~ treatment + ideology1 + log_income +
                    Q1_immigration,
                  data = data2
     )
-
-
-
+    
+    
+    
     # 1) test WCR31
-
+    
     boot31_jl <- boottest(lm_fit,
-                       B = 9999,
-                       param = "treatment",
-                       clustid = "group_id1",
[…]
