### 2022-09-27 9871785
add documentation, delete unneccessary files
