+++ R/arg_checks.R
@@ -126,10 +126,10 @@ check_boottest_args_plus <- function(
   
   if (((1 - sign_level) * (B + 1)) %% 1 != 0) {
     msg <- format_message(
-    "Note: The bootstrap usually performs best when the confidence", 
-     paste("level (here,", 1 - sign_level, "%)
+      "Note: The bootstrap usually performs best when the confidence", 
+      paste("level (here,", 1 - sign_level, "%)
            times the number of replications"), 
-     paste("plus 1 (", B, "+ 1 = ", B + 1, ") is an integer.")
+      paste("plus 1 (", B, "+ 1 = ", B + 1, ") is an integer.")
     )
     message(msg)
   }
