+++ R/boot_aggregate.R
@@ -209,18 +209,18 @@ boot_aggregate <- function(
     getauxweights = FALSE,
     ...){
   
-
+  
   # note: all boottest function arguments are tested in boottest()
   # therefore, only check for supported subset of features
-
+  
   check_arg(bootstrap_type, "charin(fnw11)")
   
   check_arg(full, "logical scalar")
   # => later => extend it to more than one set of vars to agg
-
+  
   dots <- list(...)
   from_summary <- isTRUE(dots$from_summary)
-
+  
   no_agg <- FALSE
   agg_rm <- NULL
   check_value_plus(agg, "match(att, period, cohort, TRUE) | scalar")
@@ -234,7 +234,7 @@ boot_aggregate <- function(
       } else if(agg == "cohort"){
         agg = c("cohort" = "::[^-].*:cohort::(.+)")
         agg_rm <- gsub("E::", "E::-?", 
-                      x$model_matrix_info$sunab$agg_att, fixed = TRUE)
+                       x$model_matrix_info$sunab$agg_att, fixed = TRUE)
       } else {
         agg <- x$model_matrix_info$sunab$agg_period
       }
@@ -243,20 +243,20 @@ boot_aggregate <- function(
   } else if(isFALSE(agg)){
     agg = c("nothing to remove" = "we want all the coefficients")
   }
-
+  
   is_name <- !is.null(names(agg))
-
+  
   if(!is_name && !grepl("(", agg, fixed = TRUE)){
     stop("Argument 'agg' must be a character in
          which the pattern to match must be in between parentheses.
          So far there are no parenthesis: please have a look at the examples.")
   }
-
+  
   coef <- x$coefficients
   cname <- names(coef)
   V <- x$cov.scaled
   
-
+  
   qui <- grepl(agg, cname, perl = TRUE)
   if(!any(qui)){
     if(from_summary){
@@ -272,11 +272,11 @@ boot_aggregate <- function(
       stop("The argument 'agg' does not match any variable.")
     }
   }
-
+  
   if(!isTRUE(x$summary)){
     x <- summary(x, ...)
   }
-
+  
   cname_select <- cname[qui]
   if(is_name){
     root <- rep(names(agg), length(cname_select))
@@ -285,56 +285,19 @@ boot_aggregate <- function(
     root <- gsub(paste0(".*", agg, ".*"), "\\1", cname_select, perl = TRUE)
     val <- gsub(paste0(".*", agg, ".*"), "\\2", cname_select, perl = TRUE)
   }
-
+  
   mm <- model.matrix(x)
-
+  
   cat("Run the wild bootstrap: this might take some time...(but 
       hopefully not too much time =) ).", "\n")
-  # pracma::tic()
-  # boot_fit <-
-  #   boottest(
-  #     object = x,
-  #     param = param,
-  #     clustid = clustid,
-  #     B = B,
-  #     impose_null = impose_null,
-  #     bootstrap_type= bootstrap_type, 
-  #     nthreads = nthreads
-  #   )
-  # pracma::toc()
-
-  # B <- boot_fit$boot_iter
-  # #V <- boot_fit$vcov_boot
-  # V_boot <- boot_fit$boot_vcov
-  # coef_boot <- boot_fit$boot_coef
-
-  # note: because feols() allows for outprojection of fixed effects, 
-  # but boottest() currently does not, V_boot and coef_boot will have
[…]
