+++ NAMESPACE
@@ -56,6 +56,7 @@ importFrom(dreamerr,validate_dots)
 importFrom(dreamerr,warn_up)
 importFrom(generics,glance)
 importFrom(generics,tidy)
+importFrom(gtools,permutations)
 importFrom(parallel,detectCores)
 importFrom(sandwich,vcovCL)
 importFrom(stats,aggregate)

[…]
