+++ R/boottest_felm.R
@@ -437,7 +437,8 @@ boottest.felm <- function(object,
     param = param,
     bootcluster = bootcluster,
     fe = fe,
-    engine = engine
+    engine = engine, 
+    bootstrap_type = bootstrap_type
   )
   
   enumerate <-
