+++ R/preprocess2.R
@@ -20,7 +20,8 @@ preprocess2.fixest <-
            param,
            fe,
            engine,
-           bootcluster) {
+           bootcluster, 
+           bootstrap_type) {
     #' preprocess data for objects of type fixest
     #'
     #' @param object an object of type fixest
@@ -34,6 +35,8 @@ preprocess2.fixest <-
     #'  "WildBootTests.jl"
     #' @param bootcluster a character string containing the name(s) of the
     #'  bootcluster variables. Alternatively, "min" or "max"
+    #' @param bootstrap_type Which bootstrap type should be run? '11', '13', '31' or 
+    #' '33'. For more infos, see the paper by MacKinnon, Nielsen & Webb (2022)
     #'
     #' @noRd
     #'
@@ -133,7 +136,9 @@ preprocess2.fixest <-
         fe = fe,
         N = N,
         has_weights = has_weights,
-        engine = engine
+        engine = engine, 
+        bootstrap_type = bootstrap_type, 
+        clustid_char = clustid
       )
 
 
@@ -239,7 +244,8 @@ preprocess2.felm <-
            param,
            fe,
            engine,
-           bootcluster) {
+           bootcluster, 
+           bootstrap_type) {
     #' preprocess data for objects of type felm
     #'
     #' @param object an object of type felm
@@ -253,6 +259,8 @@ preprocess2.felm <-
     #'  "WildBootTests.jl"
     #' @param bootcluster a character string containing the name(s) of the
     #'  bootcluster variables. Alternatively, "min" or "max"
+    #' @param bootstrap_type Which bootstrap type should be run? '11', '13', '31' or 
+    #' '33'. For more infos, see the paper by MacKinnon, Nielsen & Webb (2022)
     #'
     #' @noRd
     #'
@@ -318,7 +326,9 @@ preprocess2.felm <-
         fe = fe,
         N = N,
         has_weights = has_weights,
-        engine = engine
+        engine = engine, 
+        bootstrap_type = bootstrap_type, 
+        clustid_char = clustid
       )
       X <- get_fe$X
       Y <- get_fe$Y
@@ -379,7 +389,7 @@ preprocess2.felm <-
       R0[, 1:p] <- R
     }
     # }
-
+    
     res <- list(
       Y = Y,
       X = X,
@@ -404,13 +414,13 @@ preprocess2.felm <-
       has_fe = has_fe,
       all_c = all_c
     )
-
+    
     if (is_iv) {
       class(res) <- c("preprocess", "iv")
     } else {
       class(res) <- c("preprocess", "ols")
     }
-
+    
     res
   }
 
@@ -421,7 +431,8 @@ preprocess2.lm <-
            R,
            param,
            engine,
-           bootcluster) {
+           bootcluster, 
+           bootstrap_type = NULL) {
     #' preprocess data for objects of type lm
     #'
     #' @param object an object of type lm
@@ -437,23 +448,23 @@ preprocess2.lm <-
     #' @noRd
     #'
     #' @method preprocess2 lm
-
+    
     call <- object$call
     call_env <- environment(formula(object))
     fml <- formula(object)
     fml <- Formula::as.Formula(fml)
-
+    
     N <- nobs(object)
     # lm and felm don't drop NAs due to multicollinearity, while fixest does
     k <- length(na.omit(coef(object)))
     p <- object$p
-
+    
     is_iv <- FALSE
-
+    
     X <- model_matrix(object, collin.rm = TRUE)
[…]
