+++ R/RcppExports.R
@@ -9,14 +9,6 @@ boot_algo3_crv3 <- function(B, G, k, v, scores_mat, scores_boot, inv_tXX_tXgXg, 
     .Call('_fwildclusterboot_boot_algo3_crv3', PACKAGE = 'fwildclusterboot', B, G, k, v, scores_mat, scores_boot, inv_tXX_tXgXg, cores, R, delta_b_star)
 }
 
-get_c_all_boot_cpp <- function(coef_boot, shares, B, cores) {
-    .Call('_fwildclusterboot_get_c_all_boot_cpp', PACKAGE = 'fwildclusterboot', coef_boot, shares, B, cores)
-}
-
-get_se_all_boot_cpp <- function(V_boot, s, B, cores) {
-    .Call('_fwildclusterboot_get_se_all_boot_cpp', PACKAGE = 'fwildclusterboot', V_boot, s, B, cores)
-}
-
 #' Implementation of the heteroskedastic wild bootstrap. Computes
 #' HC robust variance estimators. For use in fwildclusterboot when no
 #' cluster variable is provided

[…]
