### 2022-10-02 6aa1371
#64 drop dependency on Matrix & Matrix.utils, replace by collapse::fsum
