### 2022-10-03 85b3fd4
add vec2mat, fix docs
