+++ R/boot_aggregate.R
@@ -141,10 +141,9 @@
 #' @examples 
 #' 
 #' \donttest{
+#' if(require(Namespace("fixest")){
 #'library(fixest)
-#'# Simple DiD example
 #'data(base_stagg)
-
 #'# The DiD estimation
 #'res_sunab = feols(y ~ x1 + sunab(year_treated, year) | id + year, base_stagg)
 #'res_sunab_3ref = feols(y ~ x1 + sunab(
@@ -169,11 +168,11 @@
 #'  boot_aggregate(
 #'    res_sunab, 
 #'    B = 99999, 
-#'    agg = TRUE
+#'    agg = TRUE,
 #'    ssc = boot_ssc(adj = FALSE, cluster.adj = FALSE)
 #'  )
 #' 
-#' 
+#' }
 #' }
 
 
