+++ NAMESPACE
@@ -39,9 +39,6 @@ export(tidy)
 importFrom(JuliaConnectoR,juliaEval)
 importFrom(JuliaConnectoR,juliaImport)
 importFrom(MASS,ginv)
-importFrom(Matrix,Diagonal)
-importFrom(Matrix,t)
-importFrom(Matrix.utils,aggregate.Matrix)
 importFrom(Rcpp,evalCpp)
 importFrom(collapse,GRP)
 importFrom(collapse,fsum)
