### 2022-10-03 0630a70
fix error in boot_aggregate examples
