+++ R/boot_aggregate.R
@@ -141,7 +141,7 @@
 #' @examples 
 #' 
 #' \donttest{
-#' if(require(Namespace("fixest")){
+#' if(requireNamespace("fixest")){
 #'library(fixest)
 #'data(base_stagg)
 #'# The DiD estimation
