### 2022-10-03 74beeb9
forgot to revert all changes regarding Matrix.utils for boot_algo2
