### 2022-10-03 aab6450
fix bug in test error-warning
