### 2022-10-03 722d5ac
further cmd check fixes
