+++ NAMESPACE
@@ -39,6 +39,7 @@ export(tidy)
 importFrom(JuliaConnectoR,juliaEval)
 importFrom(JuliaConnectoR,juliaImport)
 importFrom(MASS,ginv)
+importFrom(Matrix,Diagonal)
 importFrom(Rcpp,evalCpp)
 importFrom(collapse,GRP)
 importFrom(collapse,fsum)
