### 2022-10-04 21b6627
drop dependency on sandwich, thereby fix dimension problems with fixed effects and R * vcov
