+++ tests/testthat/test-new-bootstrap-variants.R
@@ -358,66 +358,106 @@ test_that("variants 31 R vs Julia", {
 })
 
 
-# test_that("new variants and fixed effects", {
-#   
-#   library(fixest)
-#   library(fwildclusterboot)
-#   
-#   B <- 9999
-#   
-#   data1 <<- fwildclusterboot:::create_data(
-#     N = 1000,
-#     N_G1 = 20,
-#     icc1 = 0.5,
-#     N_G2 = 20,
-#     icc2 = 0.2,
-#     numb_fe1 = 10,
-#     numb_fe2 = 10,
-#     seed = 908361239,
-#     weights = 1:N / N
-#   )
-#   
-#   feols_fit <- feols(proposition_vote ~ treatment + log_income | group_id1,
-#                data = data1
-#   )
-#   lm_fit <- lm(proposition_vote ~ treatment + log_income + as.factor(group_id1),
-#                      data = data1
-#   )  
-#   
-#   boot13_lm <- boottest(lm_fit,
-#                      B = 9999,
-#                      param = "treatment",
-#                      clustid = "group_id1",
-#                      bootstrap_type = "31"
-#   )
-#   
-#   boot13_fe <- boottest(feols_fit,
-#                      B = 9999,
-#                      param = "treatment",
-#                      clustid = "group_id1",
-#                      bootstrap_type = "31"
-#   )
-# 
-#   expect_error(
-#    boot13 <- boottest(feols_fit,
-#                           B = 9999,
-#                           param = "treatment",
-#                           clustid = "group_id1",
-#                           bootstrap_type = "13", 
-#                           fe= "group_id1"
-#     )
-#   )
-#   
-#   boot13fe <- boottest(feols_fit,
-#                      B = 999,
-#                      param = "treatment",
-#                      clustid = "group_id1",
-#                      bootstrap_type = "13"#, 
-#                      #fe = "group_id1"
-#   )
-#   
-#     
-#   
-# })
-# 
-# 
+test_that("new variants and fixed effects", {
+
+  library(fixest)
+  library(fwildclusterboot)
+
+  B <- 9999
+
+  data1 <<- fwildclusterboot:::create_data(
+    N = 1000,
+    N_G1 = 20,
+    icc1 = 0.5,
+    N_G2 = 20,
+    icc2 = 0.2,
+    numb_fe1 = 10,
+    numb_fe2 = 10,
+    seed = 961239,
+    weights = 1:N / N
+  )
+
+  feols_fit <- feols(proposition_vote ~ treatment + log_income | group_id1,
+               data = data1
+  )
+  
+  lm_fit <- lm(proposition_vote ~ treatment + log_income + as.factor(group_id1),
+                     data = data1
+  )
+
+  # x1 variants
+  
+  boot31_lm <- boottest(lm_fit,
+                     B = 9999,
+                     param = "treatment",
+                     clustid = "group_id1",
+                     bootstrap_type = "31" , 
+                     seed = 123, 
+                     ssc = boot_ssc(adj = FALSE, cluster.adj = FALSE)
+  )
+
+  boot31_fe <- boottest(feols_fit,
+                     B = 9999,
+                     param = "treatment",
+                     clustid = "group_id1",
+                     bootstrap_type = "31", 
+                     seed = 123, 
+                     ssc = boot_ssc(adj = FALSE, cluster.adj = FALSE)
+  )
+  
+  expect_equal(
+    pval(boot31_lm),
+    pval(boot31_fe)
+  )
+  
+  expect_equal(
+    teststat(boot31_lm),
+    teststat(boot31_fe)
+  )
+  
+  expect_equal(
+    boot31_lm$t_boot,
+    boot31_fe$t_boot
+  )
+  
+  
[…]
