+++ NAMESPACE
@@ -56,7 +56,6 @@ importFrom(generics,glance)
 importFrom(generics,tidy)
 importFrom(gtools,permutations)
 importFrom(parallel,detectCores)
-importFrom(sandwich,vcovCL)
 importFrom(stats,aggregate)
 importFrom(stats,as.formula)
 importFrom(stats,coef)
