+++ man/mboottest.fixest.Rd
@@ -134,8 +134,6 @@ regression coefficients.}
 \item{teststat_boot}{All bootstrap t-statistics.}
 \item{regression}{The regression object used in boottest.}
 \item{call}{Function call of boottest.}
-\item{engine}{The employed bootstrap algorithm.}
-\item{nthreads}{The number of threads employed.}
 \item{internal_seed}{The integer value -inherited from set.seed() - used
 within boottest() to set the random seed in either R or Julia. If NULL,
 no internal seed was created.}
@@ -150,18 +148,7 @@ in Roodman et al., 2019.
 
 To guarantee reproducibility, you can either use \verb{boottest()'s} \code{seed}
 function argument, or
-set a global random seed via
-\itemize{
-\item \code{set.seed()} when using
-\enumerate{
-\item the lean algorithm (via \code{engine = "R-lean"}) including the
-heteroskedastic wild bootstrap
-\item the wild cluster bootstrap via \code{engine = "R"} with Mammen weights or
-\item \code{engine = "WildBootTests.jl"}
-}
-\item \code{dqrng::dqset.seed()} when using \code{engine = "R"} for Rademacher, Webb
-or Normal weights
-}
+set a global random seed via\code{set.seed()}
 }
 
 \examples{
