+++ DESCRIPTION
@@ -83,7 +83,6 @@ Suggests:
     tibble, 
     did2s, 
     sandwich
-Remotes: s3alfisc/summclust
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
