+++ src/wildboottest.cpp
@@ -3,12 +3,12 @@
 #include <RcppEigen.h>
 #ifdef _OPENMP
 #include <omp.h>
-// [[Rcpp::plugins(openmp)]]
 #else
 #define omp_get_max_threads() 0
 #define EIGEN_DONT_PARALLELIZE
 #endif
 
+// [[Rcpp::plugins(openmp)]]
 // [[Rcpp::depends(RcppArmadillo, RcppEigen)]]
 
 using namespace Rcpp;
