### 2022-10-06 d28cc85
improve performance of heteroskedastic bootstrap
