+++ R/boottest_felm.R
@@ -369,6 +369,15 @@ boottest.felm <- function(object,
     )
   }
   
+  if(bootstrap_type != "fnw11"){
+    if(engine == "R"){
+      if(conf_int){
+        message("Confidence Intervals are currently only supported for 
+                the R engine with 'bootstrap_type = 'fnw11' '.")
+      }
+    }
+  }
+  
   if (!is.null(beta0)) {
     stop(
       "The function argument 'beta0' is deprecated. Please use the
