### 2022-10-06 76b9ab5
some bug fixes for mboottest methods
