### 2022-10-07 29f9844
some updates to cran release notes
