### 2022-10-08 8bf233b
add support for WCR31 for heteroskedastic bootstrap
