### 2022-10-08 1bc2676
add bug for heterosk bootstrap with btype Ã= fnw11
