### 2022-10-09 0eed8c1
fix more bugs with 12 and 13 heteroskedastic bootstrap
