+++ R/boottest_felm.R
@@ -561,6 +561,10 @@ boottest.felm <- function(object,
       impose_null = impose_null
     )
     
+    if(bootstrap_type == "fnw11"){
+      bootstrap_type <- "11"
+    }
+    
     res <- boot_algo1(
       preprocessed_object = preprocess,
       boot_iter = B,
