+++ R/boottest_fixest.R
@@ -566,6 +566,10 @@ boottest.fixest <- function(object,
       impose_null = impose_null
     )
     
+    if(bootstrap_type == "fnw11"){
+      bootstrap_type <- "11"
+    }
+    
     res <- boot_algo1(
       preprocessed_object = preprocess,
       boot_iter = B,
