+++ cran-comments.md
@@ -1,484 +1,49 @@
-## Re-Submission of 0.12 
+## Test environment
 
-The package did not pass the initial CRAN checks. I believe that the received notes are mostly false positives. On Friday, I therefore replied by email to CRAN-submissions@r-project.org, but have not yet received a reply. As I am not sure if the email received you, I am resubmitting, and have attached the content of my email. Further, note that `fwildclusterboot` is currently archived due to `Matrix.utils` being archived, a dependency that I have dropped in the submission at hand. 
+- local test (windows), R 4.2.1
+- [github actions](https://github.com/s3alfisc/fwildclusterboot/actions)
+- [win-devel](https://win-builder.r-project.org/eCVI5mKh15EF/)
+- [rhub]()
 
-"Dear CRAN team,
 
-The pre-tests produce the following notes on windows and Debian:
+## Check Results
 
-- Possibly misspelled words in DESCRIPTION:
-  MacKinnon (35:50, 38:70)
-  multiway (40:23)
+- rhub failed with prep errors for the following setups:  
+  - Debian Linux, R-devel, GCC ASAN/UBSAN
+  - Ubuntu Linux 20.04.1 LTS, R-release, GCC
 
-- * checking top-level files ... NOTE
-Non-standard file/directory found at top level:  ‘codemeta.json’
+There were no errors or warnings. There were multiple notes:
 
-I believe both to be false positives - the words are spelled correctly and a 'codemeta.json' appeared to be allowed in the past.
+- Fedora Linux (R-hub): 
+  - Found the following (possibly) invalid URLs:
+  URL: https://www.econ.queensu.ca/sites/econ.queensu.ca/files/wpaper/qed_wp_1485.pdf. This is a valid URL.
+  - checking installed package size ... NOTE installed size is  5.1Mb sub-directories of 1Mb or more:
+    libs   4.0Mb
+  - Non-standard file/directory found at top level:
+  ‘codemeta.json’
+  - checking HTML version of manual ... NOTE
+Skipping checking HTML validation: no command 'tidy' found.  I cannot change that Tidy is not on the path, or update Tidy on the external Fedora Linux server.
+  - spelling errors = false positives
 
-On windows, the following note further is a false positive - I have checked the URL prior to submission, and it works.
-
-- Found the following (possibly) invalid URLs:
-  URL: https://www.econ.queensu.ca/sites/econ.queensu.ca/files/wpaper/qed_wp_1485.pdf
-    From: README.md
-          NEWS.md
-    Status: Error
-    Message: server certificate verification failed. CAfile: none CRLfile: none
-              (Status without verification: OK)
-
-Two notes remain that I do not understand, both of which are related to the html manual:
-
-- On Windows:  checking HTML version of manual ... [1s/1s] NOTE
+- Windows (R-hub): 
+  - checking HTML version of manual ... NOTE
 Found the following HTML validation problems:
-setBoottest_engine.html:32:1 (setBoottest_engine.Rd:19): Warning: <div> isn't allowed in <h2> elements
-- On Debian: * checking HTML version of manual ... [8s] NOTE
+setBoottest_engine.html:32:1: Warning: <div> isn't allowed in <h2> elements. Don't know what this means. 
+  - checking for detritus in the temp directory ... NOTE
+  'lastMiKTeXException'. 
+  - spelling errors = false positives
+
+- Windows-devel:
+  - spelling errors -> false positives
+  - Non-standard file/directory found at top level:
+  'codemeta.json'
+  - checking HTML version of manual ... [7s] NOTE
 Encountered the following conversion/validation errors:
 missing value where TRUE/FALSE needed
 
-All the best, and thanks for all your efforts!
-
-Alex"
-
-
[…]
