### 2022-10-10 89fe942
install julia in test workflow
