### 2022-10-11 c1ccb1d
minor cleanups
