+++ R/mboottest_lm.R
@@ -25,8 +25,8 @@
 #'        Other options include "min", where bootstrapping is clustered by the
 #'        cluster variable with the fewest clusters.
 #'        Further, the subcluster bootstrap (MacKinnon & Webb, 2018) is
-#'        supported - see the \code{vignette("fwildclusterboot",
-#'         package = "fwildclusterboot")} for details.
+#'        supported - see the `vignette("fwildclusterboot",
+#'         package = "fwildclusterboot")` for details.
 #' @param seed An integer. Allows to set a random seed. For details,
 #'  see below.
 #' @param R Hypothesis Vector or Matrix giving linear combinations of
@@ -65,19 +65,19 @@
 #' @param bootstrapc Logical scalar, FALSE by default. TRUE  to request
 #' bootstrap-c instead of bootstrap-t
 #' @param ssc An object of class `boot_ssc.type` obtained with the function
-#'  \code{\link[fwildclusterboot]{boot_ssc}}. Represents how the small sample
+#'  [fwildclusterboot::boot_ssc()]. Represents how the small sample
 #'   adjustments are computed. The defaults are `adj = TRUE, fixef.K = "none",
 #'    cluster.adj = "TRUE", cluster.df = "conventional"`.
 #'             You can find more details in the help file for `boot_ssc()`.
 #'             The function is purposefully designed to mimic fixest's
-#'             \code{\link[fixest]{ssc}} function.
+#'             [fixest::ssc()] function.
 #' @param ... Further arguments passed to or from other methods.
 #'
 #' @importFrom dreamerr check_arg validate_dots
 #' @importFrom stats as.formula coef model.matrix model.response model.weights
 #' residuals rlnorm rnorm update
 #'
-#' @return An object of class \code{mboottest}
+#' @return An object of class `mboottest`
 #'
 #' \item{p_val}{The bootstrap p-value.}
 #' \item{N}{Sample size. Might differ from the regression sample size if
@@ -110,7 +110,7 @@
 #'
 #' @references Roodman et al., 2019, "Fast and wild: Bootstrap inference in
 #' STATA using boottest", The STATA Journal.
-#' (\url{https://ideas.repec.org/p/qed/wpaper/1406.html})
+#' (<https://ideas.repec.org/p/qed/wpaper/1406.html>)
 #' @references Cameron, A. Colin, Jonah B. Gelbach, and Douglas L. Miller.
 #'  "Bootstrap-based improvements for inference with clustered errors."
 #'  The Review of Economics and Statistics 90.3 (2008): 414-427.
