### 2022-10-13 f8c68cb
changes for passing initial cran checks
