+++ DESCRIPTION
@@ -31,19 +31,19 @@ Authors@R:
 	         role = c("ctb"))
 	)
 Description: Implementation of fast algorithms for wild cluster bootstrap 
-             inference developed in Roodman et al (2019, STATA Journal,
-             <doi:10.1177/1536867X19830877>) and MacKinnon et al (2022), 
+             inference developed in 'Roodman et al' (2019, 'STATA' Journal,
+             <doi:10.1177/1536867X19830877>) and 'MacKinnon et al' (2022), 
              which makes it feasible to quickly calculate bootstrap test 
              statistics based on a large number of bootstrap draws even for 
-             large samples. Multiple bootstrap types as described in MacKinnon, 
-             Nielsen & Webb (2022) are supported. 
-             Further, multiway clustering, regression weights, 
-             bootstrap weights, fixed effects and subcluster bootstrapping
-             are supported. Further, both restricted (WCR) and unrestricted
-             (WCU) bootstrap are supported. Methods are provided for a variety 
+             large samples. Multiple bootstrap types as described in 'MacKinnon, 
+             Nielsen & Webb' (2022) are supported. 
+             Further, 'multiway' clustering, regression weights, 
+             bootstrap weights, fixed effects and 'subcluster' bootstrapping
+             are supported. Further, both restricted ('WCR') and unrestricted
+             ('WCU') bootstrap are supported. Methods are provided for a variety 
              of fitted models, including 'lm()', 'feols()' 
              (from package 'fixest') and 'felm()' (from package 'lfe'). 
-             Additionally implements a heteroskedasticity-robust (HC1) wild 
+             Additionally implements a 'heteroskedasticity-robust' ('HC1') wild 
              bootstrap.
              Further, the package provides an R binding to 'WildBootTests.jl',
              which provides additional speed gains and functionality, 
