+++ src/wildboottest.cpp
@@ -182,6 +182,112 @@ List wildboottestHC(const arma::vec & y,
 }
 
 
+
+// [[Rcpp::export]]
+List wildboottestHC_v_provided(
+    const arma::vec & y,
+    const arma::mat & X,
+    const arma::mat & R,
+    const double & r,
+    const int & B,
+    const int & N_G_bootcluster,
+    const int & cores,
+    const int & type, 
+    const double & small_sample_correction, 
+    const int bootstrap_type, 
+    arma::mat weights) {
+  
+  // function implements wild cluster bootstrap,
+  // imposing the null. 
+  // same as wildboottestHC, but weights matrix v is provided
+  
+  
+  int N = X.n_rows;
+  
+  // impose null on the bootstrap dgp
+  arma::mat XXinv = (X.t() * X ).i();
+  arma::mat beta = XXinv * (X.t() * y);
+  
+  arma::mat XXinvX = XXinv * X.t();
+  arma::mat RXXinvX = (R.t() * XXinv) * X.t(); //1 x N
+  arma::mat RXXinvX_squared = (arma::pow(RXXinvX, 2));
+  
+  // calculate beta constrained
+  arma::vec beta_r = beta - XXinv * R * (R.t() * XXinv * R).i() *
+    (R.t() * beta - r);
+  arma::vec yhat_r = X * beta_r;
+  arma::vec resid_r = y - yhat_r;
+  
+  // NumericVector s(2);
+  // NumericVector prob = NumericVector::create();
+  // s[0] = 1;
+  // s[1] = -1;
+  
+  // compute 1/(1-h_ii)'s -> resid_multiplier
+  arma::vec resid_multiplier(N);
+  // resid_multiplier.ones();
+  
+  if(bootstrap_type == 1){
+    
+    resid_multiplier.ones();
+    
+  } else {
+    
+    arma::mat hatmat = X * XXinv * X.t(); 
+    arma::vec diag_hatmat(N);
+    
+    for(int i = 0; i < N; i++){
+      diag_hatmat(i) = hatmat(i,i);
+    }
+    
+    if(bootstrap_type == 2){
+      resid_multiplier = 1 / arma::sqrt(1-diag_hatmat);    
+    } else if(bootstrap_type == 3){
+      resid_multiplier = 1 / (1-diag_hatmat);    
+    }
+    
+  }
+  
+  arma::vec t_boot(B + 1);
+  
+#pragma omp parallel for num_threads(cores)
+  for(int b = 1; b < B + 1; b++){
+    
+    // create bootstrap sample
+    //arma::vec weights = RcppArmadillo::sample(s, N_G_bootcluster, true, prob);
+    //arma::vec weights = sample_weights(N_G_bootcluster, type);
+    arma::vec y_boot = yhat_r + resid_multiplier % resid_r % weights.col(b);
+    // get bootstrapped coefs and resids
+    arma::vec coef_boot = XXinvX * y_boot ; // k x 1 
+    arma::vec resid_boot = y_boot - X * coef_boot;
+    arma::mat boot_var = RXXinvX_squared * pow(resid_boot, 2);
+    
+    // calculate t-stats
+    t_boot(b) = arma::as_scalar((R.t() * coef_boot - r) / 
+      // arma::sqrt(small_sample_correction * (R.t() * sigma_boot * R)));
+      arma::sqrt(small_sample_correction * boot_var));
+  }
+  
+  // for b = 0 - always HC1
+  arma::vec y_boot = yhat_r + resid_r;
+  // get bootstrapped coefs and resids
+  arma::vec coef_boot = XXinvX * y_boot ; // k x 1 
+  arma::vec resid_boot = y_boot - X * coef_boot;
[…]
