+++ tests/testthat/test-new-bootstrap-variants.R
@@ -1,5 +1,8 @@
-test_that("test r-fnw vs r-, stochastic", {
+test_that("test r-fnw vs r-new, non-stochastic (as same seed)", {
 
+  
+      skip_on_cran()
+  
       reltol <- 0.05
       B <- 9999
 
@@ -26,12 +29,11 @@ test_that("test r-fnw vs r-, stochastic", {
         data = data1
       )
       lm_fits <- list(
-        ols = lm_fit#,
-        #  wls = lm_fit_weights
+        ols = lm_fit
       )
-# 
+
       # object <- lm_fit
-      # type <- "mammen"
+      # type <- "rademacher"
       # p_val_type = "two-tailed"
       
       
@@ -141,6 +143,8 @@ test_that("test r-fnw vs r-, stochastic", {
 
 
 test_that("new bootstrap variants II - t_stat equivalence", {
+  
+  skip_on_cran()
   
   N <- 1000
   N_G1 <- 17
@@ -360,6 +364,8 @@ test_that("variants 31 R vs Julia", {
 
 test_that("new variants and fixed effects", {
 
+  skip_on_cran()
+  
   library(fixest)
   library(fwildclusterboot)
 
