+++ tests/testthat/test-new-bootstrap-variants.R
@@ -1,5 +1,7 @@
 test_that("test r-fnw vs r-, stochastic", {
 
+      skip_on_cran()
+  
       reltol <- 0.05
       B <- 9999
 
@@ -141,6 +143,8 @@ test_that("test r-fnw vs r-, stochastic", {
 
 
 test_that("new bootstrap variants II - t_stat equivalence", {
+  
+  skip_on_cran()
   
   N <- 1000
   N_G1 <- 17
@@ -360,6 +364,8 @@ test_that("variants 31 R vs Julia", {
 
 test_that("new variants and fixed effects", {
 
+  skip_on_cran()
+  
   library(fixest)
   library(fwildclusterboot)
 
