### 2022-10-29 990f6c5
update cran notes
