### 2022-11-02 c7112f4
remove CRAN installation instructions
