### 2022-11-08 e778388
address cran comments
