### 2022-11-15 a335081
change juliaconnector link to cran #77
