### 2022-11-15 13b4105
add link to Julia's pkg manager to description #77
