### 2022-11-19 45f809a
notes for cran resubmission
