### 2022-11-20 7ff7036
update JuliaConnectoR CRAN link
