### 2022-12-17 440f570
minor documentation updates
