+++ R/srr-stats-standards.R
@@ -10,81 +10,105 @@
 #' @srrstatsVerbose TRUE
 #'
 #' @srrstats {G1.0} *`boottest()` links to multiple published papers.*
+#' 
 #' @srrstats {G1.1} *The help files of all boottest methods document that the
 #' "fast and wild" algorithm is already implemented in the STATA boottest 
-#' package. The readme and descriptions also point to the WildBootTests.jl 
-#' implementation. The WRE13 and WRE33 are, to the best of my knowledge, 
-#' only available through fwildclusterboot*.
+#' package. Additional information in the boottest() documentation points to 
+#' Julia, Python and Stata implementations. 
+#' 
 #' @srrstats {G1.2} *Life Cycle "Maturing".*
+#' 
 #' @srrstats {G1.3} *All statistical terminology should be clarified and
 #'  unambiguously defined.* This is done in an extra vignette.
+#' 
 #' @srrstats {G1.4} *`roxygen2` is used throughout the package to document
 #' all functions.*
+#' 
 #' @srrstats {G1.4a} *All internal (non-exported) functions should also be
 #' documented in standard [`roxygen2`](https://roxygen2.r-lib.org/) format,
 #'  along with a final `@noRd` tag to suppress automatic generation of `.Rd`
 #'   files.* Done.
+#' 
 #' @srrstats {G1.6} *Software should include code necessary to compare
 #'  performance claims with alternative implementations in other R packages.*
 #'  Is a link to this blog post sufficient?
 #'  ://s3alfisc.github.io/blog/post/1000x-faster-wild-cluster-bootstrap
 #'  -inference-in-r-with-fwildclusterboot/
+#' 
 #' @srrstats {G2.0} *Function argument checks are implemented throughout
 #' with via the `dreamerr` package.* Function values checked via dreamerr.
+#' 
 #' @srrstats {G2.0a} Provide explicit secondary documentation of any
 #' expectations on lengths of inputs
+#' 
 #' @srrstats {G2.1} *Implement assertions on types of inputs (see the initial
 #' point on nomenclature above).*Function values checked via dreamerr.
+#' 
 #' @srrstats {G2.1a} *Provide explicit secondary documentation of expectations
 #'  on data types of all vector inputs.*Function values checked via dreamerr.
+#' 
 #' @srrstats {G2.2} *Appropriately prohibit or restrict submission of
 #'  multivariate input to parameters expected to be univariate.*Function
 #'  values checked via dreamerr.
+#' 
 #' @srrstats {G2.3} *For univariate character input:*Function values checked
 #' via dreamerr.
+#' 
 #' @srrstats {G2.3a} *Use `match.arg()` or equivalent where applicable to only
 #' permit expected values.*Function values checked via dreamerr.
+#' 
 #' @srrstats {G2.3b} *Either: use `tolower()` or equivalent to ensure input of
 #'  character parameters is not case dependent; or explicitly document that
 #'  parameters are strictly case-sensitive.*Function values checked via
 #'  dreamerr.
+#' 
 #' @srrstats {G2.4} *Provide appropriate mechanisms to convert between
 #' different
 #'  data types, potentially including:*Function values checked via dreamerr.
+#' 
 #' @srrstats {G2.4a} *explicit conversion to `integer` via `as.integer()`*
 #' use custom function `as_integer()`
+#' 
 #' @srrstats {G2.4c} *explicit conversion to character via `as.character()`
 #' (and not `paste` or `paste0`)* Done
+#' 
 #' @srrstats {G2.4d} *explicit conversion to factor via `as.factor()`* Yes.
+#' 
 #' @srrstats {G2.6} *Software which accepts one-dimensional input should ensure
[…]
