+++ man/boottest.lm.Rd
@@ -246,12 +246,11 @@ stops the procedure after 10 iterations.
 
 The fast wild cluster bootstrap algorithms are further implemented in the
 following software packages:
-\item{
-Stata: \href{https://github.com/droodman/boottest}{boottest} }
-\item{
-Julia: \href{https://github.com/droodman/WildBootTests.jl}{WildBootTests.jl} }
-\item{
-Python: \href{https://github.com/s3alfisc/wildboottest}{wildboottest} }
+\itemize{
+\item Stata: \href{https://github.com/droodman/boottest}{boottest}
+\item Julia: \href{https://github.com/droodman/WildBootTests.jl}{WildBootTests.jl}
+\item Python: \href{https://github.com/s3alfisc/wildboottest}{wildboottest}
+}
 }
 
 \examples{
