+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   Multiple-Estimations-with-fixest: Multiple-Estimations-with-fixest.html
   WildBootTests: WildBootTests.html
   fwildclusterboot: fwildclusterboot.html
-last_built: 2022-12-17T13:12Z
+last_built: 2022-12-17T18:16Z
 urls:
   reference: https://s3alfisc.github.io/fwildclusterboot/reference
   article: https://s3alfisc.github.io/fwildclusterboot/articles
