+++ pkgdown.yml
@@ -4,9 +4,10 @@ pkgdown_sha: ~
 articles:
   Different-Variants-of-the-Wild-Cluster-Bootstrap: Different-Variants-of-the-Wild-Cluster-Bootstrap.html
   Multiple-Estimations-with-fixest: Multiple-Estimations-with-fixest.html
+  Wild-Cluster-Bootstrap-Inference-in-Stata-Julia-and-Python: Wild-Cluster-Bootstrap-Inference-in-Stata-Julia-and-Python.html
   WildBootTests: WildBootTests.html
   fwildclusterboot: fwildclusterboot.html
-last_built: 2022-12-17T18:16Z
+last_built: 2022-12-17T18:29Z
 urls:
   reference: https://s3alfisc.github.io/fwildclusterboot/reference
   article: https://s3alfisc.github.io/fwildclusterboot/articles
