+++ .github/workflows/pkgcheck.yaml
@@ -12,7 +12,7 @@ on:
   push:
     branches:
       - main
-      - pkgchecks
+      - pkgcheck
 
 jobs:
   pkgcheck:
