+++ vignettes/articles/Different-Variants-of-the-Wild-Cluster-Bootstrap.Rmd
@@ -69,8 +69,8 @@ msummary(
 
 Given the same `seed = 123`, both implementations of the `11` algorithm produce exactly the same p-values. The p-values produced by all other algorithms differ slightly, but are overall very close to each other. Confidence intervals are currently only implemented for the `fnw11` algorithm.
 
-Now, which bootstrap type should you run? MNW argue in favor of the "31" or "S" type. In the future, I hope that I'll provide a more thorough discussion. 
+Now, which bootstrap type should you run? MNW argue in favor of the "31" or "S" type. In the future, I hope that I'll provide a more thorough discussion down here =)  
 
 ## Literature 
 
-[MacKinnon, Nielsen and Webb(2022)- Fast and Reliable Bootstrap and Jacknife Infernece (2023)](https://www.econ.queensu.ca/sites/econ.queensu.ca/files/wpaper/qed_wp_1485.pdf)
+[MacKinnon, Nielsen and Webb(2022)- Fast and Reliable Bootstrap and Jacknife Inference (2023)](https://www.econ.queensu.ca/sites/econ.queensu.ca/files/wpaper/qed_wp_1485.pdf)
