### 2023-01-23 7c28618
update readme installation instructions
