+++ README.md
@@ -49,8 +49,9 @@ with $B = 100.000$ iterations in less than a second!
 -   The wild cluster bootstrap for OLS (Cameron, Gelbach & Miller 2008,
     Roodman et al, 2019).
 -   Multiple new versions of the wild cluster bootstrap as described in
-    MacKinnon, Nielsen & Webb (2022), including the WCR13, WCR31, WCR33,
-    WCU13, WCU31 and WCU33.
+    MacKinnon, Nielsen & Webb (2022), including the WCR13 (WCR-V), WCR31
+    (WCR-S), WCR33 (WCR-B), WCU13 (WCU-V), WCU31 (WCU-S) and WCU33
+    (WCU-B).
 -   The subcluster bootstrap (MacKinnon and Webb 2018).
 -   Confidence intervals formed by inverting the test and iteratively
     searching for bounds.
@@ -73,10 +74,14 @@ Additional features are provided through `WildBootTests.jl`:
 
 ### Installation
 
-You can install compiled versions of`{fwildclusterboot}` from R-universe
-(compiled) or github by following one of the steps below:
+You can install compiled versions of`{fwildclusterboot}` from CRAN
+(compiled), R-universe (compiled) or github by following one of the
+steps below:
 
 ``` r
+# from CRAN 
+install.packages("fwildclusterboot")
+
 # from r-universe (windows & mac, compiled R > 4.0 required)
 install.packages('fwildclusterboot', repos ='https://s3alfisc.r-universe.dev')
 # dev version from github
