### 2023-02-01 #546 @maurolepore
Dear @s3alfisc

Today starts my rotation as EiC and I'm reviewing the status of current open reviews.

Thanks a lot for your efforts in meeting our standards. Where are we standing now? Do you feel ready for me to resume the review process? The next step would be to re-run checks and then look for a handling editor.
