### 2023-02-01 #546 @s3alfisc
Hi @maurolepore - thanks for reaching out - I was actually about to do the same this weekend! Unfortunately, I never got the `pkgcheck` to run properly, neither on my local machine nor on github actions. But I have not really tried in a while (I had given up at some point) - so I'll try once again and if I cannot get it to run, I suppose that I would have to withdraw the submission?
