### 2023-02-01 #546 @s3alfisc
Actually, `pkgcheck` is now running on my local machine - so I'll respond to it's feedback, update all standards based on a few new features, and will notify you once everything is done!
