### 2023-02-01 0be62dc
switch pkgcheck to windows, as ubuntu latest fails
