+++ .github/workflows/pkgcheck.yaml
@@ -16,6 +16,6 @@ on:
 
 jobs:
   pkgcheck:
-    runs-on: ubuntu-latest
+    runs-on: windows-latest
     steps:
       - uses: ropensci-review-tools/pkgcheck-action@main
