+++ R/methods.R
@@ -53,9 +53,9 @@
 #' The fast wild cluster bootstrap algorithms are further implemented in the 
 #' following software packages: 
 #' \itemize{
-#' \item Stata: [boottest](https://github.com/droodman/boottest) 
-#' \item Julia: [WildBootTests.jl](https://github.com/droodman/WildBootTests.jl) 
-#' \item Python: [wildboottest](https://github.com/s3alfisc/wildboottest) 
+#' \item Stata:[boottest](https://github.com/droodman/boottest) 
+#' \item Julia:[WildBootTests.jl](https://github.com/droodman/WildBootTests.jl) 
+#' \item Python:[wildboottest](https://github.com/s3alfisc/wildboottest) 
 #' }
 #' 
 #' @references Roodman et al., 2019, "Fast and wild: Bootstrap inference in
