+++ man/boottest.Rd
@@ -42,9 +42,9 @@ or Normal weights
 The fast wild cluster bootstrap algorithms are further implemented in the
 following software packages:
 \itemize{
-\item Stata: \href{https://github.com/droodman/boottest}{boottest}
-\item Julia: \href{https://github.com/droodman/WildBootTests.jl}{WildBootTests.jl}
-\item Python: \href{https://github.com/s3alfisc/wildboottest}{wildboottest}
+\item Stata:\href{https://github.com/droodman/boottest}{boottest}
+\item Julia:\href{https://github.com/droodman/WildBootTests.jl}{WildBootTests.jl}
+\item Python:\href{https://github.com/s3alfisc/wildboottest}{wildboottest}
 }
 }
 
