+++ .github/workflows/pkgcheck.yaml
@@ -16,6 +16,6 @@ on:
 
 jobs:
   pkgcheck:
-    runs-on: windows-latest
+    runs-on: ubuntu-latest
     steps:
       - uses: ropensci-review-tools/pkgcheck-action@main
