### 2023-02-01 #546 @s3alfisc
I've run the pkgcheck action, with the following [results](https://github.com/s3alfisc/fwildclusterboot/issues/91#issuecomment-1412694651):

- the pkgcheck coverage test fails, as `pkgcheck` does not run some tests that my local github actions CI runs (coverage is at 89% running all tests, [codecov link](https://app.codecov.io/github/s3alfisc/fwildclusterboot/commit/b92803442d37f5d2bdfb6a5c0a5795e4f726e9cb))
- installation on ubuntu-lastest fails. No problems on older versions of ubuntu via [github actions](https://github.com/s3alfisc/fwildclusterboot/actions/runs/4068569451).
- The following functions are used in other packages: pval, error_message - which I think should be fine?

I'll start the bot to confirm all of this later today & then you can decide if you think everything is good enough to proceed, @maurolepore ?
