### 2023-02-02 #546 @maurolepore
Thanks @s3alfisc for following up.

It's looking good!

* Addressing the duplicated names is optional.
- [ ] Could you avoid the test-coverage problem with [covr exclusions](https://github.com/r-lib/covr#exclusions)?
- [ ] Can you paste here the results from R CMD check? I'm curious what message you get.
