### 2023-02-02 #546 @mpadge
@maurolepore You may ignore that error from our systems. It arises through tests which use `JuliaConnectorR` to connect with julia code, and specifically:
```
Error: Evaluation in Julia failed.
Original Julia error message:
LoadError: ArgumentError: Package StableRNGs not found in current path.
```
@s3alfisc does that install via [GitHub action workflow](https://github.com/s3alfisc/fwildclusterboot/blob/488cc58f19b429c5a10b028f7a4403c2a9334704/.github/workflows/R-CMD-check.yaml#L83-L87), but we don't yet have a way to custom-install required Julia packages. Sorry for any inconvenience.
