### 2023-02-02 #546 @maurolepore
Thanks @mpadge.
Then I'm inclined to move forward and start the search for a handling editor. While you're here and given this is a stats submission, can I ask you to give me :+1: if you agree or explain if you dissagree?
