### 2023-02-02 #546 @s3alfisc
Hi @mpadge , thanks for chiming in =) At some point you suggested that I should disable all tests based on WildBootTests.jl when I do not find a Julia installation - do you have a suggestion on how I could implement that? I unfortunately never managed to figure this out myself.

@maurolepore , I unfortunately cannot disable these test via `covr`, as I am comparing output produced via fwildclusterboot's central function (boottest()) with a Julia implementation.
