### 2023-02-02 #546 @adamhsparks
I’ve not followed too closely, but do have an interest in mixing R and Julia. Have you tried something like, `tryCatch(system2(command="juli1a", args="--version"), error = print("Julia not found"))`?

From: https://discourse.julialang.org/t/test-if-julia-is-installed/83703/4
