### 2023-02-04 #546 @s3alfisc
Hi @adamhsparks, thanks so much for reminding me of this thread! Completely forgot about it :D I did not get it to work back in the day - it turns out I had not added Julia to my path. Now I've added the function below to check if Julia (or any programming language) is installed - it returns a logical indicating if Julia is installed or not. I can probably tweak it a bit and also get out the Julia version number.

```
find_proglang <- function(lang){

  #' Check if julia or python are installed /
  #' can be found on the PATH
  #'
  #' @param lang which language to check. Either 'julia' or 'python'
  #'
  #' @return logical. TRUE if lang is found on path, FALSE if not

  dreamerr::check_arg(lang, "charin(julia, python)")

  # x == 0L if no error is found
  x =
    tryCatch(
      system2(
        command= lang, args="--version"
      ),
      error = function(e) e
    )

  language_found <- ifelse(x == 0L, TRUE, FALSE)

  language_found

}

```
