### 2023-02-04 af265cc
check in tests if julia is installed, skip if not
