+++ tests/testthat/test-r-vs-julia.R
@@ -3,7 +3,10 @@ test_that("test r against Julia I: stochastic tests", {
 
   skip_on_cran()
   skip_on_ci()
-
+  skip_if_not(
+    find_proglang("julia"), 
+    message = "skip test as julia installation not found."
+  )
 
   if(TRUE){
     reltol <- 0.05
