+++ tests/testthat/test_tidy.R
@@ -1,7 +1,11 @@
 test_that("test tidiers with q = 1", {
 
   skip_on_cran()
-
+  skip_if_not(
+    find_proglang("julia"), 
+    message = "skip test as julia installation not found."
+  )
+  
   if(TRUE){
 
     lm_fit <<-
@@ -62,6 +66,10 @@ test_that("test tidiers with q > 1", {
   
   skip_on_cran()
   skip_on_ci()
+  skip_if_not(
+    find_proglang("julia"), 
+    message = "skip test as julia installation not found."
+  )
 
   if(TRUE){
     
