### 2023-02-04 #546 @maurolepore
@s3alfisc if I understand correctly one elegant alternative is to combine `nzchar()` with `Sys.which()`:

``` r
Sys.which("python")
#> python
#>     ""
Sys.which("julia")
#>            julia
#> "/usr/bin/julia"

nzchar("")
#> [1] FALSE
nzchar("not/an/empty/string")
#> [1] TRUE

nzchar(Sys.which("python"))
#> [1] FALSE
nzchar(Sys.which("julia"))
#> [1] TRUE
```
