+++ .github/workflows/R-CMD-check.yaml
@@ -4,7 +4,7 @@ on:
   push:
     branches:
       - master
-      - ropensci
+      - dev
   pull_request:
     branches:
       - master
