### 2023-02-04 f2485aa
add nzchar(Sys.which()) solution to check for Julia installation
