### 2023-02-04 #546 @s3alfisc
Thanks @maurolepore - I've switched to your solution in the [dev version ](https://github.com/s3alfisc/fwildclusterboot/commit/f2485aa7c580f51741484dbf05b4d875c57ed1ed)=)
