### 2023-02-04 032fde6
add examples to find_proglang
