### 2023-02-05 #546 @adamhsparks
So that should be possible to use to check before running tests, no?
