### 2023-02-05 #546 @s3alfisc
Yes, exactly, e.g. with `testthat::skip_if_not(nzchar(Sys.which('julia')))` on top of the test file =)
