+++ R/boottest_lm.R
@@ -485,7 +485,6 @@ boottest.lm <- function(object,
         r = r,
         sign_level = sign_level,
         param = param,
-        # seed = seed,
         p_val_type = p_val_type,
         nthreads = nthreads,
         type = type,
@@ -521,7 +520,6 @@ boottest.lm <- function(object,
         type = type,
         full_enumeration = full_enumeration,
         small_sample_correction = small_sample_correction,
-        seed = internal_seed, 
         object = object, 
         impose_null = impose_null
       )
@@ -556,8 +554,7 @@ boottest.lm <- function(object,
       type = type,
       full_enumeration = full_enumeration,
       small_sample_correction = small_sample_correction,
-      heteroskedastic = heteroskedastic,
-      seed = internal_seed
+      heteroskedastic = heteroskedastic
     )
     conf_int <- p_grid_vals <- grid_vals <- FALSE
   } else if (engine == "WildBootTests.jl") {
