### 2023-02-11 060e4c5
restructure bootstrap algorithms, drop seed argument, add sampling argument to boottest()
