+++ R/boottest_felm.R
@@ -61,7 +61,6 @@
 #' wild cluster bootstrap via the algorithm outlined in "fast and wild" 
 #' (Roodman et al (2019)). "11" is the default for the heteroskedastic 
 #' bootstrap.
-#' @param seed An integer. Allows to set a random seed. For details, see below.
 #' @param R Hypothesis Vector giving linear combinations of coefficients.
 #'  Must be either NULL or a vector of the same length as `param`. If NULL,
 #'   a vector of ones of length param.
@@ -131,6 +130,12 @@
 #' @param bootstrapc Logical scalar, FALSE by default. TRUE  to request
 #' bootstrap-c instead of bootstrap-t. Only relevant when 'engine =
 #' "WildBootTests.jl"'
+#' @param sampling 'dqrng' or 'standard'. If 'dqrng', the 'dqrng' package is
+#' used for random number generation (when available). If 'standard', 
+#' functions from the 'stats' package are used when available. 
+#' This argument is mostly a convenience to control random number generation in 
+#' a wrapper package around `fwildclusterboot`, `wildrwolf`. 
+#' I recommend to use the fast' option. 
 #' @param ... Further arguments passed to or from other methods.
 #'
 #' @importFrom dreamerr check_arg validate_dots
@@ -166,8 +171,7 @@
 #' @method boottest felm
 #'
 #' @section Setting Seeds:
-#' To guarantee reproducibility, you can either use `boottest()'s` `seed`
-#' function argument, or
+#' To guarantee reproducibility, you need to 
 #' set a global random seed via
 #' + `set.seed()` when using
 #'    1) the lean algorithm (via `engine = "R-lean"`) including the
@@ -258,7 +262,6 @@
 #'     clustid = c("group_id1", "group_id2"),
 #'     fe = "Q1_immigration",
 #'     sign_level = 0.2,
-#'     seed = 8,
 #'     r = 2
 #'   )
 #'   # test treatment + ideology1 = 2
@@ -313,7 +316,6 @@ boottest.felm <- function(object,
                           bootcluster = "max",
                           fe = NULL,
                           conf_int = TRUE,
-                          seed = NULL,
                           R = NULL,
                           r = 0,
                           beta0 = NULL,
@@ -324,6 +326,7 @@ boottest.felm <- function(object,
                           p_val_type = "two-tailed",
                           tol = 1e-6,
                           maxiter = 10,
+                          sampling = "dqrng",
                           nthreads = getBoottest_nthreads(),
                           ssc = boot_ssc(
                             adj = TRUE,
@@ -354,8 +357,6 @@ boottest.felm <- function(object,
   check_arg(p_val_type, "charin(two-tailed, equal-tailed,>, <)")
   
   check_arg(conf_int, "logical scalar")
-  check_arg(seed, "scalar integer | NULL")
-  check_arg(internal_seed, "scalar logical")
   check_arg(R, "NULL| scalar numeric | numeric vector")
   check_arg(r, "numeric scalar | NULL")
   check_arg(fe, "character scalar | NULL |formula ")
@@ -368,6 +369,9 @@ boottest.felm <- function(object,
   check_arg(floattype, "charin(Float32, Float64)")
   check_arg(maxmatsize, "scalar integer | NULL")
   check_arg(bootstrapc, "scalar logical")
+  
+  check_arg(sampling, "charin('dqrng', 'standard')")
+  
   
   if(bootstrap_type %in% c("33", "13")){
     stop(
@@ -411,12 +415,6 @@ boottest.felm <- function(object,
   if (inherits(fe, "formula")) {
     fe <- attr(terms(fe), "term.labels")
   }
-  
[…]
