+++ tests/testthat/test-seed.R
@@ -6,12 +6,14 @@ test_that("seed works for OLS", {
     message = "skip test as julia installation not found."
   )
   
+  set.seed(12312)
+  dqrng::dqset.seed(9786)
+  
   requireNamespace("fixest")
   requireNamespace("dqrng")
   
-  if(TRUE){
-    data1 <<-
-      fwildclusterboot:::create_data(
+  data1 <<-
+    fwildclusterboot:::create_data(
         N = 5000,
         N_G1 = 40,
         icc1 = 0.01,
@@ -58,110 +60,6 @@ test_that("seed works for OLS", {
   
       expect_true(boot_lm_s1$p_val != boot_lm_s2$p_val)
   
-  
-      # Case 2: same internal seed
-  
-      boot_lm_s1 <- suppressMessages(
-        boottest(
-          object = lm_fit,
-          clustid = "group_id1",
-          #  full enumeration (N_G = 10, 2^10 = 1024)
-          B = 999,
-          seed = 1,
-          param = "treatment",
-          type = "rademacher",
-          conf_int = FALSE,
-          engine = engine
-        )
-      )
-  
-      boot_lm_s2 <- suppressMessages(
-        boottest(
-          object = lm_fit,
-          clustid = "group_id1",
-          # no full enumeration (N_G = 10, 2^10 = 1024)
-          B = 999,
-          seed = 1,
-          param = "treatment",
-          type = "rademacher",
-          conf_int = FALSE,
-          engine = engine
-        )
-      )
-  
-      expect_equal(boot_lm_s1$p_val, boot_lm_s2$p_val)
-
-  
-      # Case 3: seed outside and within
-  
-      set.seed(9)
-      dqrng::dqset.seed(9)
-  
-      boot_lm_s1 <- suppressMessages(
-        boottest(
-          object = lm_fit,
-          clustid = "group_id1",
-          #  full enumeration (N_G = 10, 2^10 = 1024)
-          B = 999,
-          param = "treatment",
-          type = "rademacher",
-          conf_int = FALSE,
-          engine = engine
-        )
-      )
-  
-      boot_lm_s2 <- suppressMessages(
-        boottest(
-          object = lm_fit,
-          clustid = "group_id1",
-          # no full enumeration (N_G = 10, 2^10 = 1024)
-          B = 999,
-          seed = 9,
-          param = "treatment",
-          type = "rademacher",
-          conf_int = FALSE,
-          engine = engine
-        )
-      )
-  
-      expect_equal(boot_lm_s1$p_val, boot_lm_s2$p_val)
-  
-  
-      # Case 4 different seed outside & within
-  
-      set.seed(9)
-      dqrng::dqset.seed(9)
-      
-      boot_lm_s1 <- suppressMessages(
-        boottest(
-          object = lm_fit,
-          clustid = "group_id1",
-          #  full enumeration (N_G = 10, 2^10 = 1024)
-          B = 999,
-          param = "treatment",
-          type = "rademacher",
-          conf_int = FALSE,
-          engine = engine
-        )
-      )
-  
-      boot_lm_s2 <- suppressMessages(
-        boottest(
-          object = lm_fit,
-          clustid = "group_id1",
-          # no full enumeration (N_G = 10, 2^10 = 1024)
-          B = 999,
-          seed = 2,
-          param = "treatment",
-          type = "rademacher",
-          conf_int = FALSE,
-          engine = engine
-        )
-      )
-  
-      expect_true(boot_lm_s1$p_val != boot_lm_s2$p_val)
-  
-  
       # Case 5: different starting seeds
   
       set.seed(9)
@@ -194,39 +92,4 @@ test_that("seed works for OLS", {
   
[…]

[…]
