### 2023-02-11 230f5e2
drop code in boot_algo1
