### 2023-02-11 09d1bb2
minor cleanups to boot_algo_fnr.R
