### 2023-02-11 bb3acd6
re-fix minor error, pass Julia rng object
