+++ tests/testthat/test-convergence.R
@@ -2,6 +2,7 @@ test_that("algorithm performance test", {
   
   skip_on_cran()
   skip_on_ci()
+  
   skip_if_not(
     find_proglang("julia"), 
     message = "skip test as julia installation not found."
