+++ R/boottest_felm.R
@@ -370,7 +370,7 @@ boottest.felm <- function(object,
   check_arg(maxmatsize, "scalar integer | NULL")
   check_arg(bootstrapc, "scalar logical")
   
-  check_arg(sampling, "charin('dqrng', 'standard')")
+  check_arg(sampling, "charin(dqrng, standard)")
   
   
   if(bootstrap_type %in% c("33", "13")){
