+++ R/boottest_lm.R
@@ -342,8 +342,8 @@ boottest.lm <- function(object,
   check_arg(maxmatsize, "scalar integer | NULL")
   check_arg(bootstrapc, "scalar logical")
   
-  check_arg(sampling, "charin('dqrng', 'standard')")
-            
+  check_arg(sampling, "charin(dqrng, standard)")
+  
   if(bootstrap_type != "fnw11"){
     if(engine == "R"){
       if(conf_int){
