### 2023-02-11 acad5f7
fix bug with lean clustered algo and full enumeration
