### 2023-02-11 bbe4fbf
set seeds for new variants tests
