+++ R/boot_aggregate.R
@@ -222,6 +222,12 @@ boot_aggregate <- function(
   check_arg(full, "logical scalar")
   # => later => extend it to more than one set of vars to agg
   
+  # remind packages users to set a global seed
+  inform_seed(
+    frequency_id = "seed-reminder-bootagg", 
+    engine = engine
+  )
+  
   dots <- list(...)
   from_summary <- isTRUE(dots$from_summary)
   
@@ -251,9 +257,12 @@ boot_aggregate <- function(
   is_name <- !is.null(names(agg))
   
   if(!is_name && !grepl("(", agg, fixed = TRUE)){
-    stop("Argument 'agg' must be a character in
+    rlang::abort(
+        c("Argument 'agg' must be a character in
          which the pattern to match must be in between parentheses.
-         So far there are no parenthesis: please have a look at the examples.")
+         So far there are no parenthesis: please have a look at the examples."), 
+        use_cli_format = TRUE
+    )
   }
   
   coef <- x$coefficients
@@ -273,7 +282,10 @@ boot_aggregate <- function(
       x <- summary(x, agg = FALSE, ...)
       return(x$coeftable)
     } else {
-      stop("The argument 'agg' does not match any variable.")
+      rlang::abort(
+        "The argument 'agg' does not match any variable.", 
+        use_cli_format = TRUE
+      )
     }
   }
   
