+++ R/boottest_felm.R
@@ -372,31 +372,46 @@ boottest.felm <- function(object,
   
   check_arg(sampling, "charin(dqrng, standard)")
   
+  # remind packages users to set a global seed
+  inform_seed(
+    frequency_id = "seed-reminder-boot-felm", 
+    engine = engine
+  )  
   
   if(bootstrap_type %in% c("33", "13")){
-    stop(
-      paste(
+    rlang::abort(
+      c(
         "The bootstrap types '13' and '33' are not yet supported for objects
         of type felm. Support will be added in the future. Until then, you 
         can run the '13' and '33' bootstrap types with objects of type 'lm' 
         and 'fixest'."
-      )
+      ), 
+      use_cli_format = TRUE
     )
   }
   
   if(bootstrap_type != "fnw11"){
     if(engine == "R"){
       if(conf_int){
-        message("Confidence Intervals are currently only supported for 
-                the R engine with 'bootstrap_type = 'fnw11' '.")
+        rlang::inform(
+        "Confidence Intervals are currently only supported for 
+                the R engine with 'bootstrap_type = 'fnw11' '.", 
+        use_cli_format = TRUE, 
+        .frequency = "regularly", 
+        .frequency_id = "CI only for fnw algo."
+        )
       }
     }
   }
   
   if (!is.null(beta0)) {
-    stop(
-      "The function argument 'beta0' is deprecated. Please use the
-      function argument 'r' instead, by which it is replaced."
+    rlang::abort(
+      c(
+        "The function argument 'beta0' is deprecated.
+         Please use the function argument 'r' instead,
+         by which it is replaced."
+      ),
+      use_cli_format = TRUE
     )
   }
   
