### 2023-02-12 0fc8bad
return bootstrapped t-stats as vector #99
