+++ README.md
@@ -108,10 +108,12 @@ data(voters)
 # fit the model via fixest::feols(), lfe::felm() or stats::lm()
 lm_fit <- lm(proposition_vote ~ treatment  + log_income + as.factor(Q1_immigration) + as.factor(Q2_defense), data = voters)
 # bootstrap inference via boottest()
-lm_boot <- boottest(lm_fit, clustid = c("group_id1"), B = 9999, param = "treatment", seed = 1)
+lm_boot <- boottest(lm_fit, clustid = c("group_id1"), B = 9999, param = "treatment")
+#> Too guarantee reproducibility, don't forget to set a global random seed
+#> **both** via `set.seed()` and `dqrng::dqset.seed()`.
+#> This message is displayed once every 8 hours.
 summary(lm_boot)
-#> boottest.lm(object = lm_fit, param = "treatment", B = 9999, clustid = c("group_id1"), 
-#>     seed = 1)
+#> boottest.lm(object = lm_fit, param = "treatment", B = 9999, clustid = c("group_id1"))
 #>  
 #>  Hypothesis: 1*treatment = 0
 #>  Observations: 300
@@ -121,7 +123,7 @@ summary(lm_boot)
 #>  Number of Clusters: 40
 #> 
 #>              term estimate statistic p.value conf.low conf.high
-#> 1 1*treatment = 0    0.079     3.983       0     0.04     0.118
+#> 1 1*treatment = 0    0.079     3.983   0.001    0.039     0.119
 ```
 
 ## Citation
@@ -141,7 +143,7 @@ citation("fwildclusterboot")
 #> A BibTeX entry for LaTeX users is
 #> 
 #>   @Misc{,
-#>     title = {fwildclusterboot: Fast Wild Cluster Bootstrap Inference for Linear Regression Models (Version 0.12)},
+#>     title = {fwildclusterboot: Fast Wild Cluster Bootstrap Inference for Linear Regression Models (Version 0.12.4.3)},
 #>     author = {Alexander Fischer and David Roodman},
 #>     year = {2021},
 #>     url = {https://cran.r-project.org/package=fwildclusterboot},
