### 2023-02-12 1220369
bump WildBootTests.jl version requirement to 0.9
