+++ NEWS.md
@@ -2,40 +2,44 @@
 
 ## Potentially Breaking Changes: 
 
-* `boottest()`, `mboottest()` and `boot_aggregate()`no longer have a dedicated `seed` argument. Reproducibility of results can only be controlled by setting a **global seed** via `drqng::dqset.seed()` and `set.seed()`. For more context, see the discussion below. 
+* `boottest()`, `mboottest()` and `boot_aggregate()`no longer have a dedicated `seed` argument. From version 0.13, reproducibility of results can only be controlled by setting a **global seed** via `drqng::dqset.seed()` and `set.seed()`. For more context, see the discussion below. As a consequence, results produced via old versions of `fwildlcusterboot` are no longer exactly reproducible. 
 
-In consequence, results produced via older versions of `fwildclusterboot` might no longer be exactly reproducible.   
+* When the bootstrap is run via `engine = "WildBootTests.jl"`, the bootstrapped t-statistics and the original t-statistic are now returned as vectors (to align with the results from other `enginges`). Previously, they were returned as matrices. 
 
 ## Other Changes: 
 
 * `boottest()` receives a new argument, `sampling`, which controls if random numbers are drawn via functions from `base` or the `dqrng` package. 
-* Some code refactoring. All bootstrap algorithms and their associated files have been renamed (e.g. `boot_algo2.R` is not called `boot_algo_fastnwild.R`).
-* Much nicer error message formatting, via `rlang::abort()`, `warn()` and `inform()`.
+* Some code refactoring. All bootstrap algorithms and their associated files have been renamed (e.g. `boot_algo2.R` is not called `boot_algo_fastnwild.R`, etc.).
+* Much nicer error and message formatting, via `rlang::abort()`, `warn()` and `inform()`. `rlang` is added as a dependency.
 
 ## Background on the Change to Seeding
 
-Prior to the changes introduced with `v0.13`, if the `seed` argument was specified in `boottest()`, `boottest()` would call `set.seed()` or `dqrng::dqset.seed()` internally, to guarantee that calling `boottest()` twice under the same seed would lead to exactly identical inferences. But because both functions are not 'scoped' within `boottest()`, they would reset the **global** seed. The next random number generated in a given R session would then relate to the new seed state set by the call to `boottest()`. In other words, code like this 
+Prior to the changes introduced in `v0.13`, `boottest()` will always call `set.seed()` or `dqrng::dqset.seed()` internally, regardless of whether the `seed` argument is specified or not (in the ladder case, it will create an internal seed by randomly drawing from a large set of integers). I considered this harmless, as setting seeds inside `boottest()` in this way does not affect the reproducibility of scripts run end-to-end.
+
+However, I have learned that is generally considered bad practice to overwrite global variables without notification - for example, the authors of [numpy](https://numpy.org/doc/stable/reference/random/generated/numpy.random.seed.html) have deprecated their `np.random.seed()` function for this reason. 
+
+Here is a quick example on what happens if a function "reseeds": it affects the future chain of random draws. 
 
 ```r
-# Example 1
-set.seed(1)
-dqrng.dqset.seed(2)
-boottest(..., seed = NULL)
-rnorm(1)
-```
-and 
-```r
-# Example 2
-set.seed(1)
-dqrng.dqset.seed(2)
-boottest(..., seed = 3)
-rnorm(1)
-```
-would leed to different inferences, because the the two calls of `rnorm(1)` would be based on different global seed states. 
+fn_reseed <- function(x){set.seed(x)}
 
[…]
