### 2023-02-15 123a7e7
add session startup message on seed changes
