+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: fwildclusterboot
 Title: Fast Wild Cluster Bootstrap Inference for Linear Models
-Version: 0.12.4.3
+Version: 0.13
 Authors@R: 
     c(
     person(given = "Alexander",
