+++ cran-comments.md
@@ -2,7 +2,14 @@
 
 I have tested the package on: 
 - rhub
+- win devel
 - github actions
+
+and received the following notes: 
+
+* checking for detritus in the temp directory ... NOTE
+Found the following files/directories:
+  'lastMiKTeXException'
 
 
 ## Resubmission of 0.12.1 
