+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   Wild-Cluster-Bootstrap-Inference-in-Stata-Julia-and-Python: Wild-Cluster-Bootstrap-Inference-in-Stata-Julia-and-Python.html
   WildBootTests: WildBootTests.html
   fwildclusterboot: fwildclusterboot.html
-last_built: 2022-12-19T22:00Z
+last_built: 2023-02-26T14:58Z
 urls:
   reference: https://s3alfisc.github.io/fwildclusterboot/reference
   article: https://s3alfisc.github.io/fwildclusterboot/articles
