### 2023-03-16 #546 @maurolepore
Thanks @s3alfisc for the reminder.
I was confused -- I expected the stats team to take over. My apologies!

I'm reviewing the EiC duties for stats submissions and it seems that I need to call `@ropensci-review-bot check srr`. So here I go.
