### 2023-03-16 #546 @maurolepore
@mpadge can you please guide me? I don't want to cause additional delay.

In the [stats guide on EiC responsabilities](https://stats-devguide.ropensci.org/pkgsubmission.html#editor-in-chief) I see:

> Call @ropensci-review-bot check srr to generate a summary of compliance with our “srr” (Software Review Roclets) system for documenting standards compliance.
