### 2023-03-16 #546 @mpadge
@maurolepore Yeah, `check package` command includes all the information contained in `check srr`. But it's still probably good practice to call `check srr` first - it only spits out a few lines to tell whether you should proceed to full check.
