### 2023-03-27 #546 @helske
Hi @s3alfisc, I reran the the automatic checks which still show some issues, but based on the discussion above it seems that the these are mainly are caused by problems in our side so I will proceed with seeking for reviewers.

Regarding the items marked with ✖️:

- The package is missing is the `codemeta.json` file, could you please create this for example with the codemetar (https://github.com/ropensci/codemetar) package by running `codemetar::write_codemeta()` when in the package root directory.
- When checking the coverage from the repo directly, it shows very nice coverage of 88%, so no problems there.
- Also looking at the github actions history the R CMD check seems to work (at least for platforms other than MacOS).
