### 2023-03-27 #546 @helske
---

Thanks again for your submission @s3alfisc!. In addition to the notes above, few more comments:
- You could run `desc::desc_normalize()` for your package which reformats the DESCRIPTION fields in standard way (e.g., orders the packages in `Imports` field alphabetically).
- Some of the test files in `testthat` contain several consecutive empty lines; I suggest using only one to make the files bit more readable. The [`styler`](https://styler.r-lib.org/) package might be helpful here.
- You might want to use "proper" heading for the README file instead of the default fwildclusterboot (which I think is also automatically used in the github.io pages), e.g. the same heading as in CRAN.
- You use `requireNamespace` in tests, but given that those packages are already in declared in `DESCRIPTION`, you probably don't need this. Alternatively, you could replace those calls with `testthat` packages `skip_if_not_installed` which should properly indicate you that the test was skipped.

---
