### 2023-03-27 #546 @s3alfisc
Hi @helske - thanks for being the editor of this package, and for your first feedback. I will try to address all of your initial comments within the next days. Best, Alex
