### 2023-03-27 b4bd7d1
clean up description file using desc_normalize
