+++ DESCRIPTION
@@ -1,97 +1,80 @@
 Package: fwildclusterboot
 Title: Fast Wild Cluster Bootstrap Inference for Linear Models
 Version: 0.13.0
-Authors@R: 
-    c(
-    person(given = "Alexander",
-           family = "Fischer",
-           role = c("aut", "cre"),
-           email = "alexander-fischer1801@t-online.de"),
-    person(given = "David", 
-           family = "Roodman", 
-           role = "aut"), 
-    person(given = "Achim", 
-           family = "Zeileis",
-           role = c("ctb"),
-           comment = "Author of included sandwich fragments"),
-	person(given = "Nathaniel",
-	       family = "Graham",
-	       role = c("ctb"), 
-           comment = "Contributor to included sandwich fragments"),
-	person(given = "Susanne",
-	       family = "Koell", 
-	       role = c("ctb"), 
-           comment = "Contributor to included sandwich fragments"),
-	person("Laurent", 
-	        "Berge",
-	        role = c("ctb"), 
-            comment = "Author of included fixest fragments"),
-	 person("Sebastian", 
-	        "Krantz", 
-	         role = c("ctb"))
-	)
-Description: Implementation of fast algorithms for wild cluster bootstrap 
-             inference developed in 'Roodman et al' (2019, 'STATA' Journal,
-             <doi:10.1177/1536867X19830877>) and 'MacKinnon et al' (2022), 
-             which makes it feasible to quickly calculate bootstrap test 
-             statistics based on a large number of bootstrap draws even for 
-             large samples. Multiple bootstrap types as described in 'MacKinnon, 
-             Nielsen & Webb' (2022) are supported. 
-             Further, 'multiway' clustering, regression weights, 
-             bootstrap weights, fixed effects and 'subcluster' bootstrapping
-             are supported. Further, both restricted ('WCR') and unrestricted
-             ('WCU') bootstrap are supported. Methods are provided for a variety 
-             of fitted models, including 'lm()', 'feols()' 
-             (from package 'fixest') and 'felm()' (from package 'lfe'). 
-             Additionally implements a 'heteroskedasticity-robust' ('HC1') wild 
-             bootstrap.
-             Last, the package provides an R binding to 'WildBootTests.jl',
-             which provides additional speed gains and functionality, 
-             including the 'WRE' bootstrap for instrumental variable models 
-             (based on models of type 'ivreg()' from package 'ivreg')
-             and hypotheses with q > 1.
+Authors@R (parsed):
+    * Alexander Fischer <alexander-fischer1801@t-online.de> [aut, cre]
+    * David Roodman [aut]
+    * Achim Zeileis [ctb] (Author of included sandwich fragments)
+    * Nathaniel Graham [ctb] (Contributor to included sandwich fragments)
+    * Susanne Koell [ctb] (Contributor to included sandwich fragments)
+    * Laurent Berge [ctb] (Author of included fixest fragments)
+    * Sebastian Krantz [ctb]
+Description: Implementation of fast algorithms for wild cluster
+    bootstrap inference developed in 'Roodman et al' (2019, 'STATA'
+    Journal, <doi:10.1177/1536867X19830877>) and 'MacKinnon et al' (2022),
+    which makes it feasible to quickly calculate bootstrap test statistics
+    based on a large number of bootstrap draws even for large samples.
[…]
