### 2023-03-27 07339f4
drop requireNamespace from test
