+++ tests/testthat/test-method_equivalence.R
@@ -10,9 +10,7 @@ test_that("Do different, but equivalent ways to specify
   set.seed(2351)
   dqrng::dqset.seed(2351)
   
-  requireNamespace("fixest")
-  requireNamespace("lfe")
-            
+
   print_results <- FALSE
 
     data1 <<-
@@ -1435,9 +1433,7 @@ test_that("Do different, but equivalent ways to specify
 
 test_that("clustid can be fe", {
   
-  requireNamespace("fixest")
-  requireNamespace("lfe")
-    
+
   print_results <- FALSE
   
   data1 <<-
