+++ tests/testthat/test-r-lean.R
@@ -1,8 +1,7 @@
 test_that("test lean cpp boottest", {
-  
   set.seed(96578)
   dqrng::dqset.seed(9568)
-  
+
   data1 <<- fwildclusterboot:::create_data(
     N = 1000,
     N_G1 = 1000,
@@ -16,21 +15,21 @@ test_that("test lean cpp boottest", {
   lm_fit <- lm(proposition_vote ~ treatment + ideology1 + log_income,
     data = data1
   )
-  feols_fit <-fixest::feols(
+  feols_fit <- fixest::feols(
     proposition_vote ~ treatment + ideology1 + log_income,
     data = data1
   )
-  felm_fit <-lfe::felm(proposition_vote ~ treatment + ideology1 + log_income,
+  felm_fit <- lfe::felm(proposition_vote ~ treatment + ideology1 + log_income,
     data = data1
   )
 
 
   # Test 1: heteroskedadestic wild bootstrap
-  
-  
+
+
   set.seed(96578)
   dqrng::dqset.seed(9568)
-  
+
   boot_lm <- boottest(lm_fit,
     param = "treatment",
     B = 999,
@@ -38,10 +37,10 @@ test_that("test lean cpp boottest", {
     nthreads = 1,
     type = "webb"
   )
-  
+
   set.seed(96578)
   dqrng::dqset.seed(9568)
-  
+
   boot_lm2 <- boottest(lm_fit,
     param = "treatment",
     B = 999,
@@ -52,17 +51,17 @@ test_that("test lean cpp boottest", {
 
   set.seed(96578)
   dqrng::dqset.seed(9568)
-  
+
   boot_feols <- boottest(feols_fit,
     param = "treatment",
     B = 9999,
     ssc = boot_ssc(adj = FALSE, cluster.adj = FALSE),
     nthreads = 1
   )
-  
+
   set.seed(96578)
   dqrng::dqset.seed(9568)
-  
+
   boot_felm <- boottest(felm_fit,
     param = "treatment",
     B = 9999,
@@ -134,7 +133,7 @@ test_that("test lean cpp boottest", {
 
 
   # test oneway clustering
-  
+
   boot_lm1 <- boottest(feols_fit,
     param = "treatment",
     clustid = "group_id1",
@@ -182,74 +181,63 @@ test_that("test lean cpp boottest", {
 
   expect_equal(pval(boot_lm1), pval(boot_lm2), tolerance = 0.05)
   expect_equal(teststat(boot_lm1), teststat(boot_lm2))
-  
-  
+
+
   # test for non-standard hypotheses
-  
 })
 
 
 test_that("r-lean multi-param tests", {
-  
-  
+  N <- 2000
+  seed <- 7896
 
-    N <- 2000
-    seed <- 7896
-    
-    data1 <<- fwildclusterboot:::create_data(
-      N = N,
-      N_G1 = 8,
-      icc1 = 0.5,
-      N_G2 = 20,
-      icc2 = 0.2,
-      numb_fe1 = 10,
-      numb_fe2 = 10,
-      seed = seed
-      ,
-      weights = 1:N / N
+  data1 <<- fwildclusterboot:::create_data(
+    N = N,
+    N_G1 = 8,
+    icc1 = 0.5,
+    N_G2 = 20,
+    icc2 = 0.2,
+    numb_fe1 = 10,
+    numb_fe2 = 10,
+    seed = seed,
+    weights = 1:N / N
+  )
+
+  lm_fit <- lm(proposition_vote ~ treatment + log_income,
+    data = data1
+  )
+
+  boot <- suppressWarnings(
+    boottest(
+      lm_fit,
+      B = 999,
[…]
