+++ tests/testthat/test_fixest_sugar.R
@@ -1,6 +1,4 @@
 test_that("test fixest formula sugar", {
-  
-
   voters <- fwildclusterboot:::create_data(
     N = 10000,
     N_G1 = 20,
@@ -12,7 +10,7 @@ test_that("test fixest formula sugar", {
     seed = 908369,
     weights = 1:N / N
   )
-  
+
   fit <-
     fixest::feols(c(income, proposition_vote) ~ treatment, data = voters)
   res <-
@@ -25,7 +23,7 @@ test_that("test fixest formula sugar", {
         clustid = "group_id1"
       )
     )
-  
+
   fit <-
     fixest::feols(proposition_vote ~ csw(treatment, ideology1), data = voters)
   res <-
@@ -38,28 +36,28 @@ test_that("test fixest formula sugar", {
         clustid = "group_id1"
       )
     )
-  
+
   fit <-
     fixest::feols(
-      proposition_vote ~ fixest::i(treatment, ideology1), 
+      proposition_vote ~ fixest::i(treatment, ideology1),
       data = voters
     )
-  
+
   res <-
     boottest(fit,
-             B = 999,
-             param = "fixest::treatment::0:ideology1",
-             clustid = "group_id1"
+      B = 999,
+      param = "fixest::treatment::0:ideology1",
+      clustid = "group_id1"
     )
-  
-  
+
+
   fit <- fixest::feols(
     proposition_vote ~ treatment,
     split = ~Q2_defense,
     cluster = ~group_id1,
     data = voters
   )
-  
+
   res <- lapply(
     fit,
     \(x) boottest(
@@ -69,15 +67,14 @@ test_that("test fixest formula sugar", {
       clustid = "group_id1"
     )
   )
-  
-  expect_equal(unlist(lapply(res, function(x) {
-    x$t_stat
-  })),
-  unlist(lapply(fit, function(x) {
-    fixest::tstat(x)["treatment"]
-  })),
-  ignore_attr = TRUE
+
+  expect_equal(
+    unlist(lapply(res, function(x) {
+      x$t_stat
+    })),
+    unlist(lapply(fit, function(x) {
+      fixest::tstat(x)["treatment"]
+    })),
+    ignore_attr = TRUE
   )
-  
-  
 })
