### 2023-04-01 #546 @meghapsimatrix
Yes please send the draft. Thanks!

On Sat, Apr 1, 2023 at 6:39 AM Alexander Fischer ***@***.***>
wrote:

> @mpadge <https://github.com/mpadge> asked me to clarify that the review
> should be based on the "master" branch - the repo used to have a dedicated
> "ropensci" branch - to which this submission was linked - which I have
> deprecated a while back as it had gotten stale.
>
> Besides, I have a paper draft for fwildclusterboot, which is not yet in a
> state to be shared widely, but which I think could be helpful to reviewers,
> as it gives details on some of the few small deviations of the algorithm
> implemented in fwildclusterboot vs the fast and wild paper. Should I send
> you this draft @meghapsimatrix <https://github.com/meghapsimatrix>?
>
