### 2023-04-03 #546 @helske
Thanks Alex, looking good. Regarding the readme.Rmd, this line defines the title for the readme as well as for the pkgdown site:
https://github.com/s3alfisc/fwildclusterboot/blob/1783a83a39aedb38e1f25983b3a6ccadcd183436/README.Rmd#L18
i.e. these:
![image](https://user-images.githubusercontent.com/1560448/229415095-b5491dee-47b4-428d-a31b-1325162a8f01.png)
and
![image](https://user-images.githubusercontent.com/1560448/229415263-012de2a2-7c8d-4b29-b808-779789f0b04b.png)
