### 2023-04-13 #546 @meghapsimatrix
-   [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

This is a great package with super fast implementation of cluster wild bootstrap. Thanks for the opportunity to review.

**R CMD check**

1 NOTE

    ❯ checking top-level files ... NOTE
      Non-standard file/directory found at top level:
        ‘codemeta.json’

**Comments on vignettes and documentation and comments on trying to use the functions (organized based on sections of the pkgdown website <https://s3alfisc.github.io/fwildclusterboot/index.html>)**

-   README

    -   The hex logo appears twice

-   Get Started Vignette

    -   The vignette could go over why it is important to run many bootstrap replications. And, what does many mean? Also, is there a point where the number of bootstraps is too excessive?

    -   The vignette could also provide a high level overview of the algorithm that makes cluster wild bootstrapping run fast, especially in R. The fastness seems to be the core idea behind the package so some detail on that somewhere on the vignette might be good.

    -   Under the The `boottest()` function section, the text says that you are creating a random dataset but the code shows that you are using the voters dataset.

    -   The vignette could also go over when to use which weights. For example, if we have a small number of clusters, Rademacher weights may not be appropriate.

-   Function Reference

    -   The functions could be grouped together in some logical way. Right now, the list of functions is organized alphabetically.

    -   I like the `tidy()` implementation a lot. Small improvement would be to change the column names to snake_case and add which types of weights were used to bootstrap (e.g., Rademacher).

    -   I am a bit confused about interpreting the plot from the `plot()` function. A bit more explanation of what the blue and red lines mean could help :)

    -   A bit more explanation on the different `bootstrap_type` options would help too.

    -   In the documentation page for `boottest` it says we can use either `boottest`'s seed argument or set the seed globally using `set.seed()` and `dqrng::dqset.seed()`. I tried using the seed argument but was not able to run it. Adding the seed argument directly to the `boottest` function would be an improvement and make the set up more use-friendly rather than having to set seed using both `set.seed()`and `dqrng::dqset.seed()`.

    ``` r
    library(fwildclusterboot)

    dqrng::dqset.seed(2352342)

    data(voters)

    # estimate the regression model via lm
    lm_fit <- lm(
      proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration ,
      data = voters
    )

    boot_lm <- boottest(
      lm_fit,
      clustid = "group_id1",
      param = "treatment",
      B = 9999,
      seed = 20230412
    )
    ```

    Error: in boottest.lm(lm_fit, clustid = "group_id1", param ...:
     'seed' is not a valid argument of function boottest.lm.
    Please note that the seeding behavior for random number generation for `boottest()` has changed with `fwildclusterboot`
    version 0.13.

    It will no longer be possible to exactly reproduce results produced by versions lower than 0.13.

    If your prior results were produced under sufficiently many bootstrap iterations, none of your conclusions will change.  For
    more details about this change, please read the notes in
    [news.md](https://cran.r-project.org/web/packages/fwildclusterboot/news/news.html).
    This warning is displayed once per session.
    Called from: stop_up(my_call, res)

-   Vignettes fwildclusterboot and fixest

    -   This vignette could use more explanation.
