### 2023-04-14 #546 @helske
Thanks for the review @meghapsimatrix! I apologize that unfortunately it seems that the review template link in my email was incorrect, as your review used the basic template used for non-statistical-software packages. Would you mind filling the correct one, which can be found here: https://stats-devguide.ropensci.org/pkgreview.html#pkgrev-template?

@s3alfisc I am still seeking for another reviewer, you might want to hold on with non-trivial software updates and replies until we've gotten a second review as well.
