### 2023-04-14 #546 @meghapsimatrix
Sorry about that! Will fill out the correct one!

On Fri, Apr 14, 2023 at 12:47 AM Jouni Helske ***@***.***>
wrote:

> Thanks for the review @meghapsimatrix <https://github.com/meghapsimatrix>!
> I apologize that unfortunately it seems that the review template link in my
> email was incorrect, as your review used the basic template used for
> non-statistical-software packages. Would you mind filling the correct one,
> which can be found here:
> https://stats-devguide.ropensci.org/pkgreview.html#pkgrev-template?
>
> @s3alfisc <https://github.com/s3alfisc> I am still seeking for another
> reviewer, you might want to hold on with non-trivial software updates and
> replies until we've gotten a second review as well.
>
