### 2023-04-14 9efd141
add model_matrix.plm method
