+++ R/preprocess2.R
@@ -13,6 +13,191 @@ preprocess2 <- function(object, ...) {
 }
 
 
+preprocess2.plm <-
+  function(object,
+           clustid,
+           R,
+           param,
+           fe,
+           engine,
+           bootcluster,
+           bootstrap_type) {
+    #' preprocess data for objects of type plm
+    #'
+    #' @param object an object of type fixest
+    #' @param clustid a character string containing the name(s) of the
+    #'  cluster variables
+    #' @param R Hypothesis Vector giving linear combinations of coefficients.
+    #' @param fe character vector. name of the fixed effect to be projected
+    #'  out in the bootstrap
+    #' @param param character vector. names of the parameter(s) to test
+    #' @param engine The bootstrap algorithm to run. Either "R" or
+    #'  "WildBootTests.jl"
+    #' @param bootcluster a character string containing the name(s) of the
+    #'  bootcluster variables. Alternatively, "min" or "max"
+    #' @param bootstrap_type Which bootstrap type should be run?
+    #' '11', '13', '31' or 33'. For more infos, see the paper
+    #' by MacKinnon, Nielsen & Webb (2022)
+    #'
+    #' @noRd
+    #'
+    #' @method preprocess2 fixest
+
+
+    call <- object$call
+    call_env <- object$call_env
+    fml <- formula(object)
+    fml <- Formula::as.Formula(fml)
+
+    fml_full <- formula(fml, collapse = TRUE)
+
+    N <- nobs(object)
+    # lm and felm don't drop NAs due to multicollinearity, while fixest does
+    k <- length(na.omit(coef(object)))
+
+    # plm specific checks
+    if(object$args$model != "within"){
+      rlang::abort("boottest() only supports estimation of 'within' models
+                  via the plm::plm().",
+                  use_cli_format = TRUE
+      )
+    }
+
+    is_iv <- FALSE
+    has_fe <- FALSE
+
+    if(object$args$model == "within"){
+      has_fe <- TRUE
+    }
+
+    mf <- model.frame(object)
+    Y <- model_frame[,1]
+    X <- model_frame[,2:(k+1)]
+
+    weights <- weights(object)
+    if (is.null(weights)) {
+      has_weights <- FALSE
+      weights <- rep(1, N)
+    } else {
+      has_weights <- TRUE
+    }
+
+    if (has_weights) {
+      if (!is.null(fe)) {
+        rlang::abort(
+          "boottest() unfortunately currently does not support WLS and fixed
+          effects. Please set fe = NULL to run a bootstrap with WLS.",
+          use_cli_format = TRUE
+        )
+      }
+    }
+
+    fixed_effect <- NULL
+    k2 <- 0
+    W <- n_fe <- NULL
+
+
+    if (has_fe) {
+      # if(!is.null(fe)){
+      get_fe <- transform_fe(
+        object = object,
+        X = X,
+        Y = Y,
+        fe = fe,
+        N = N,
+        has_weights = has_weights,
+        engine = engine,
+        bootstrap_type = bootstrap_type,
+        clustid_char = clustid
+      )
+
+
+      X <- get_fe$X
+      Y <- get_fe$Y
+      fixed_effect <- get_fe$fixed_effect
+      W <- get_fe$W
+      n_fe <- get_fe$n_fe
+      k2 <- get_fe$k2
+
+    }
+
+
+    # get cluster variable
+    if (!is.null(clustid)) {
+      clustid_list <- get_cluster(
+        object = object,
+        clustid_char = clustid,
+        N = N,
+        bootcluster = bootcluster,
[…]
