+++ R/boottest_plm.r
@@ -1,7 +1,582 @@
-boottest.plm <- function(){
+#' Fast wild cluster bootstrap inference for object of class plm
+#'
+#' `boottest.plm` is a S3 method that allows for fast wild cluster
+#' bootstrap inference for objects of class plm by  implementing
+#' fast wild bootstrap algorithms as developed in Roodman et al., 2019 
+#' and MacKinnon, Nielsen & Webb (2022).
+#'
+#' @param object An object of class plm and estimated via `plm::feols()`.
+#' Non-linear models are not supported.
+#' @param clustid A character vector or rhs formula containing the names of the
+#' cluster variables. If NULL,
+#'        a heteroskedasticity-robust (HC1) wild bootstrap is run.
+#' @param param A character vector or rhs formula. The name of the regression
+#'        coefficient(s) for which the hypothesis is to be tested
+#' @param B Integer. The number of bootstrap iterations. When the number of
+#'  clusters is low,
+#'        increasing B adds little additional runtime.
+#' @param bootcluster A character vector or rhs formula of length 1. Specifies
+#' the bootstrap clustering variable or variables. If more
+#'        than one variable is specified, then bootstrapping is clustered by the
+#'         intersections of
+#'        clustering implied by the listed variables. To mimic the behavior of
+#'        stata's boottest command,
+#'        the default is to cluster by the intersection of all the variables
+#'        specified via the `clustid` argument,
+#'        even though that is not necessarily recommended (see the paper by
+#'         Roodman et al cited below, section 4.2).
+#'        Other options include "min", where bootstrapping is clustered by
+#'        the cluster variable with the fewest clusters.
+#'        Further, the subcluster bootstrap (MacKinnon & Webb, 2018) is
+#'         supported - see the `vignette("fwildclusterboot", package =
+#'          "fwildclusterboot")` for details.
+#' @param fe A character vector or rhs formula of length one which contains
+#' the name of the fixed effect to be projected
+#'        out in the bootstrap. Note: if regression weights are used, fe
+#'        needs to be NULL.
+#' @param sign_level A numeric between 0 and 1 which sets the significance level
+#'        of the inference procedure. E.g. sign_level = 0.05
+#'        returns 0.95% confidence intervals. By default, sign_level = 0.05.
+#' @param conf_int A logical vector. If TRUE, boottest computes confidence
+#'        intervals by test inversion. If FALSE, only the p-value is returned.
+#' @param engine Character scalar. Either "R", "R-lean" or "WildBootTests.jl".
+#'  Controls if `boottest()` should run via its native R implementation 
+#'  or `WildBootTests.jl`.
+#'  "R" is the default and implements the cluster bootstrap
+#'  as in Roodman (2019). "WildBootTests.jl" executes the
+#'  wild cluster bootstrap via the WildBootTests.jl
+#'  package. For it to run, Julia and WildBootTests.jl need
+#'  to be installed.
+#'  The "R-lean" algorithm is a memory friendly, but less
+#'  performant rcpp-armadillo based implementation of the wild
+#'  cluster bootstrap.
+#'  Note that if no cluster is provided, boottest() always
+#'  defaults to the "lean" algorithm. You can set the employed
+#'  algorithm globally by using the
+#'  `setBoottest_engine()` function.
+#' @param bootstrap_type Determines which wild cluster bootstrap type should be 
+#' run. Options are "fnw11","11", "13", "31" and "33" for the wild cluster 
[…]
