+++ man/boottest.plm.Rd
@@ -1,0 +1,381 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/boottest_plm.r
+\name{boottest.plm}
+\alias{boottest.plm}
+\title{Fast wild cluster bootstrap inference for object of class plm}
+\usage{
+\method{boottest}{plm}(
+  object,
+  param,
+  B,
+  clustid = NULL,
+  bootcluster = "max",
+  fe = NULL,
+  sign_level = 0.05,
+  conf_int = TRUE,
+  R = NULL,
+  r = 0,
+  beta0 = NULL,
+  type = "rademacher",
+  impose_null = TRUE,
+  bootstrap_type = "fnw11",
+  p_val_type = "two-tailed",
+  tol = 1e-06,
+  maxiter = 10,
+  sampling = "dqrng",
+  nthreads = getBoottest_nthreads(),
+  ssc = boot_ssc(adj = TRUE, fixef.K = "none", cluster.adj = TRUE, cluster.df =
+    "conventional"),
+  engine = getBoottest_engine(),
+  floattype = "Float64",
+  maxmatsize = FALSE,
+  bootstrapc = FALSE,
+  getauxweights = FALSE,
+  ...
+)
+}
+\arguments{
+\item{object}{An object of class plm and estimated via \code{plm::feols()}.
+Non-linear models are not supported.}
+
+\item{param}{A character vector or rhs formula. The name of the regression
+coefficient(s) for which the hypothesis is to be tested}
+
+\item{B}{Integer. The number of bootstrap iterations. When the number of
+clusters is low,
+increasing B adds little additional runtime.}
+
+\item{clustid}{A character vector or rhs formula containing the names of the
+cluster variables. If NULL,
+a heteroskedasticity-robust (HC1) wild bootstrap is run.}
+
+\item{bootcluster}{A character vector or rhs formula of length 1. Specifies
+the bootstrap clustering variable or variables. If more
+than one variable is specified, then bootstrapping is clustered by the
+intersections of
+clustering implied by the listed variables. To mimic the behavior of
+stata's boottest command,
+the default is to cluster by the intersection of all the variables
+specified via the \code{clustid} argument,
+even though that is not necessarily recommended (see the paper by
+Roodman et al cited below, section 4.2).
+Other options include "min", where bootstrapping is clustered by
+the cluster variable with the fewest clusters.
+Further, the subcluster bootstrap (MacKinnon & Webb, 2018) is
+supported - see the \code{vignette("fwildclusterboot", package = "fwildclusterboot")} for details.}
+
+\item{fe}{A character vector or rhs formula of length one which contains
+the name of the fixed effect to be projected
+out in the bootstrap. Note: if regression weights are used, fe
+needs to be NULL.}
+
+\item{sign_level}{A numeric between 0 and 1 which sets the significance level
+of the inference procedure. E.g. sign_level = 0.05
+returns 0.95\% confidence intervals. By default, sign_level = 0.05.}
+
+\item{conf_int}{A logical vector. If TRUE, boottest computes confidence
+intervals by test inversion. If FALSE, only the p-value is returned.}
+
+\item{R}{Hypothesis Vector giving linear combinations of coefficients.
+Must be either NULL or a vector of the same length as \code{param}. If NULL,
+a vector of ones of length param.}
+
+\item{r}{A numeric. Shifts the null hypothesis
+H0: param = r vs H1: param != r}
+
+\item{beta0}{Deprecated function argument. Replaced by function argument 'r'.}
+
+\item{type}{character or function. The character string specifies the type
+of boostrap to use: One of "rademacher", "mammen", "norm"
+and "webb". Alternatively, type can be a function(n) for drawing
[…]
