+++ NAMESPACE
@@ -4,6 +4,7 @@ S3method(boottest,felm)
 S3method(boottest,fixest)
 S3method(boottest,ivreg)
 S3method(boottest,lm)
+S3method(boottest,plm)
 S3method(confint,boottest)
 S3method(glance,boottest)
 S3method(glance,mboottest)
