### 2023-04-15 b7f9322
only allow models pooling and within + add fe checks
