### 2023-04-15 d1537bc
add literature vignette on wild cluster bootstrapping
