+++ vignettes/Literature.Rmd
@@ -1,0 +1,43 @@
+---
+title: "Literature on the Wild Bootstrap and Clustered Inference in Regression Models"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Literature on the Wild Bootstrap and Clustered Inference in Regression Models}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+
++ [Roodman, MacKinnon, Nielsen & Webb - Fast and Wild Bootstrap Inference (Stata Journal), 2019](https://journals.sagepub.com/doi/pdf/10.1177/1536867X19830877)
+
++ [MacKinnon, Nielsen & Webb - Fast and Reliable Jackknife and Bootstrap
+Methods for Cluster-Robust Inference (Journal of Applied Econometrics), 2023](https://arxiv.org/pdf/2301.04527.pdf)
+
++ [MacKinnon - Fast cluster bootstrap methods for linear regression models (Econometrics and Statistics, 2021)](https://www.econstor.eu/bitstream/10419/247206/1/qed-wp-1465.pdf)
+
++ [MacKinnon, Nielsen & Webb - Cluster-robust inference: A guide to empirical practice (Journal of Econometrics), 2023](https://www.sciencedirect.com/science/article/pii/S0304407622000781)
+
++ [MacKinnon & Webb -  The wild bootstrap for few (treated) clusters (Econometrics Journal), 2018](https://academic.oup.com/ectj/article-abstract/21/2/114/5078969)
+
++ [MacKinnon - Wild Cluster Bootstrap Confidence Intervals (L'Actualité économique), 2015](https://www.erudit.org/en/journals/ae/1900-v1-n1-ae02507/1036912ar.pdf)
+
++ [Davidson & MacKinnon - Wild bootstrap tests for IV regression (Journal of Economic and Business Statistics), 2010](https://www.tandfonline.com/doi/abs/10.1198/jbes.2009.07221)
+
++ [Cameron, Gelbach & Miller - Bootstrap-based improvements for inference with clustered errors (Review of Economics and Statistics)](https://www.nber.org/system/files/working_papers/t0344/t0344.pdf)
+
++ [Flachaire - Bootstrapping heteroskedastic regression models: wild bootstrap vs. pairs bootstrap (Computational Statistics & Data Analysis), 2005](https://shs.hal.science/file/index/docid/175910/filename/Flachaire_03a.pdf)
+
++ [Webb - Reworking wild bootstrap based inference for clustered errors (forthcoming at Canadian Journal of Economics)](https://ideas.repec.org/p/qed/wpaper/1315.html)
+
++ [Kline & Santos - A Score Based Approach to Wild Bootstrap Inference (Journal of Econometric Methods), 2012](https://www.nber.org/system/files/working_papers/w16127/w16127.pdf)
+
+
+
+
