### 2023-04-15 8e129cb
add vignette with a brief overview on the literature
