### 2023-04-17 s3alfisc/fwildclusterboot#112 @s3alfisc
Vignette Improvements
From @meghapsimatrix [review](https://github.com/ropensci/software-review/issues/546), I agree with all points.

 - [ ] The vignette could go over why it is important to run many bootstrap replications. And, what does many mean? Also, is there a point where the number of bootstraps is too excessive?
- [ ] The vignette could also provide a high level overview of the algorithm that makes cluster wild bootstrapping run fast, especially in R. The fastness seems to be the core idea behind the package so some detail on that somewhere on the vignette might be good.
- [ ] Under the The boottest() function section, the text says that you are creating a random dataset but the code shows that you are using the voters dataset.
- [ ] The vignette could also go over when to use which weights. For example, if we have a small number of clusters, Rademacher weights may not be appropriate.

In general, I should consider adding more guidelines on when wild cluster bootstrapping is required, which one to use, etc.

I have started by adding a brief [literature section](https://s3alfisc.github.io/fwildclusterboot/articles/Literature.html).
