### 2023-04-17 6fb464d
fix ssr-stats-standards #111
