### 2023-04-19 #546 @meghapsimatrix
Here is the correct template filled out :) @s3alfisc thanks for addressing my comments :) Great job on this super useful package :)

-   [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (If you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

### Compliance with Standards

-   [x] This package complies with a sufficient number of standards for a silver badge
-   [x] This grade of badge is the same as what the authors wanted to achieve

The following standards currently deemed non-applicable (through tags of `@srrstatsNA`) could potentially be applied to future versions of this software: (Please specify)

Please also comment on any standards which you consider either particularly well, or insufficiently, documented.

_-   G1.0 is well documented. The authors have made clear connections to academic literature._

_-   G.1.1 is also well documented. The authors provide citations to the novel-ness of the algorithm in general and in R._

_-   G.1.3 Some statistical terminology could be expanded upon more. For example, vignettes use terms like restricted and unrestricted bootstraps which could use bit more explanation :)_

_-   RE4.18 and 4.3 to 4.5 Summary output and tidy outputs are well done._

For packages aiming for silver or gold badges:

-   [x] This package extends beyond minimal compliance with standards in the following ways: (please describe)

_- Compliance with a good number of standards beyond those identified as minimally necessary. Numbers of standards: G : 50 / 68; RE : 20 / 48; Total : 70 / 116_

_- Demonstrating excellence in compliance with multiple standards from at least two broad sub-categories.The package demonstrates excellence in compliance with G.1.0, G.1.1, RE4.18 and RE4.3 to 4.5_

_- Generality of usage beyond one single envisioned use case. The package supports ordinary least squares (OLS), fixed effects, and instrumental variables (IV) regression models. Subcluster bootstrapping, multiway bootstrapping, confidence intervals calculations are also supported._

_- Internal aspects of package structure and design. Algorithms are written in a way that is efficient, flexible, generalizable, and accurate. Right now the package works for OLS and IV models. More types of models, like multilevel models, could be incorporated._

### General Review

-   [x] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING
-   [x] The documentation is sufficient to enable general use of the package beyond one specific use case

Estimated hours spent reviewing: _10 hours_

### Further Comments

I already provided narrative comments which have been addressed :) Thanks!!
