### 2023-04-29 #546 @meghapsimatrix
Thank you!

On Sat, Apr 29, 2023 at 7:43 AM Alexander Fischer ***@***.***>
wrote:

> Hi Megha, I have opened a (yet incomplete PR) on the dev branch
> <https://github.com/s3alfisc/fwildclusterboot/pull/114>to incorporate
> your feedback. The idea is to add a "FAQ" vignette in which I will try to
> explain some key concepts without too much technical jargon. I have also
> added you as a package contributor!
>
