### 2023-05-15 s3alfisc/fwildclusterboot#115 @95goo
boottest fails when a fixed effect variable is of type "date"
Hi, thank you very much for this amazing package, It's been great to dig into.
Below is some sample code. I am clustering by c and also using it as a fixed effect in my model. Note that b is also a categorical variable.
`
model <- feols(Y ~ X |b + c, cluster = "c")
`
I run boottest as follows:
`
boottest(model, clustid = "c", param = "X",  B = 9999)
`
The error I receive is: "error in solve.default(crossprod(weights_sq * X)) : system is computationally singular: reciprocal condition number"

Any advices at all will be much appreciated. Thank you very much.
