### 2023-05-15 s3alfisc/fwildclusterboot#115 @s3alfisc
Hi, thanks for the feedback and the nice words!

Unfortunatley (or fortunately 😄 ) I cannot reproduce this error. Which version of the package are you running? What happens if you install the most up to date version from CRAN? If you still run into these issues, can you maybe provide data so that I can reproduce it?

Here is the code that I run:

```R
library(fwildclusterboot)
library(fixest)
data(voters)

sapply(
  voters[,
         c(
           "proposition_vote",
           "treatment",
           "Q1_immigration",
           "Q2_defense"
           )],
  class
)

# proposition_vote
# "integer"
# treatment
# "integer"
# Q1_immigration
# "factor"
# Q2_defense
# "factor"

fit <- feols(
  proposition_vote ~ treatment |
    Q1_immigration + Q2_defense,
  data = voters
)

boot <- boottest(
  fit,
  param = "treatment",
  clustid = "Q1_immigration",
  B = 999
)
summary(boot)
# boottest.fixest(object = fit, param = "treatment", B = 999, clustid = "Q1_immigration")
#
# Hypothesis: 1*treatment = 0
# Observations: 300
# Bootstr. Type: rademacher
# Clustering: 1-way
# Confidence Sets: 95%
# Number of Clusters: 10
#
# term estimate statistic p.value conf.low conf.high
# 1 1*treatment = 0    0.077     1.462   0.108   -0.003     0.213

boot <- boottest(
  fit,
  param = "treatment",
  clustid = "Q2_defense",
  B = 999
)
summary(boot)
# boottest.fixest(object = fit, param = "treatment", B = 999, clustid = "Q2_defense")
#
# Hypothesis: 1*treatment = 0
# Observations: 300
# Bootstr. Type: rademacher
# Clustering: 1-way
# Confidence Sets: 95%
# Number of Clusters: 10
#
# term estimate statistic p.value conf.low conf.high
# 1 1*treatment = 0    0.077     2.758   0.033     0.01     0.159
```
