### 2023-05-15 s3alfisc/fwildclusterboot#115 @95goo
Thanks very much. I updated the package as you recommended.
I tried to run with only one fixed effect(categorical variable) to test now:
`
model <- feols(date = reg_data, Y ~ X | c, cluster = "c")
boottest(model, clustid = "c", param = "X",  B = 200)
`

This is the error I receive: "Error in 1 | c : operations are possible for numeric, logical, or complex types"
If you have any recommendation further please let me know, otherwise I will provide sample data for assistance
