### 2023-05-16 s3alfisc/fwildclusterboot#115 @s3alfisc
What is the exact type of your date and year variables? I.e. what do you get when run running sapply(data, class)?
