### 2023-05-16 s3alfisc/fwildclusterboot#115 @95goo
The year is formatted as 'numeric' while the week is formatted as 'date' object. Is there an issue with this?
