### 2023-05-16 s3alfisc/fwildclusterboot#115 @s3alfisc
Great, maybe the date type causes the problem here - I will check this later today. What happens if you convert the date to a plain factor?
