### 2023-05-16 s3alfisc/fwildclusterboot#115 @s3alfisc
In your first example, were either b or c date variables?
