### 2023-05-17 s3alfisc/fwildclusterboot#115 @95goo
Thank you very much sir. Converting the week variable to the plain factor using "factor" makes the code below work for me!!
`
model <- feols(date = reg_data, Y ~ X | week, cluster = "c")
`
hooray!!!

In my first example b and c were not date variables however. =( I get the following error for the following code where c is a categorical variable (10 distinct values)
`
model <- feols(date = reg_data, Y ~ X | c, cluster = "c")
`
`
boottest(model, clustid = "c", param = "X",  B = 200)
`

This is the error I receive: "Error in 1 | c : operations are possible for numeric, logical, or complex types"
