### 2023-05-17 s3alfisc/fwildclusterboot#115 @95goo
HOWEVER, changing variable "c" from `character` to `factor` using the "factor" function... the boottest function works... I do not totally understand this and would love to learn more about why this is the case
`
reg_data <-reg_data %>% mutate(c = factor(c))
`
`
model <- feols(date = reg_data, Y ~ X | c, cluster = "c")
`
`
boottest(model, clustid = "c", param = "X",  B = 200)
`
