### 2023-05-19 s3alfisc/fwildclusterboot#115 @s3alfisc
I cannot reproduce the error that you observe with character variables. Can you provide me an example with simulated date where the error occurs? All fixed effects are transformed into factors internally, hence it should not matter if you provide c as a character or factor.

Note that in general, it is dangerous to call variables "c", because c is also a base function. Lots of things that could go wrong there. ChatGPT says the following: "It is generally not recommended to use the name "c" for a variable in R because "c" is a commonly used base function in R for combining or concatenating objects. If you assign a value to the variable "c", you will overwrite the default behavior of the base function, leading to potential confusion and errors in your code."

With the date variable, you have indeed discovered a bug - something fails in the fixed effects preprocessing pipeline. I'll try to fix that asap =)
