### 2023-05-21 0c69619
initial sparsification of algos #116
