+++ R/boot_algo_fastnreliable.R
@@ -29,6 +29,7 @@
 #' @param impose_null logical scalar. Should the null be imposed on the
 #' bootstrap dgp or not?
 #' @importFrom MASS ginv
+#' @importFrom Matrix crossprod tcrossprod t Matrix
 #' @importFrom summclust vcov_CR3J
 #' @return A list of bootstrap results. 
 #' @noRd
@@ -63,6 +64,8 @@ boot_algo_fastnreliable <- function(
   }
 
   X <- preprocessed_object$X
+  # convert to sparse matrix
+  X <- Matrix::Matrix(X)
   y <- preprocessed_object$Y
   R <- preprocessed_object$R0
   cluster_df <- preprocessed_object$clustid
@@ -97,12 +100,12 @@ boot_algo_fastnreliable <- function(
   # precompute a range of other objects
   tXgXg <- lapply(
     seq_along(1:G),
-    function(g) crossprod(X_list[[g]])
+    function(g) Matrix::crossprod(X_list[[g]])
   )
 
   tXgyg <- lapply(
     seq_along(1:G),
-    function(g) t(X_list[[g]]) %*% y_list[[g]]
+    function(g) Matrix::t(X_list[[g]]) %*% y_list[[g]]
   )
 
 
@@ -126,22 +129,22 @@ boot_algo_fastnreliable <- function(
 
     tX1gX1g <- lapply(
       seq_along(1:G),
-      function(g) crossprod(X1_list[[g]])
+      function(g) Matrix::crossprod(X1_list[[g]])
     )
 
     tX1gyg <- lapply(
       seq_along(1:G),
-      function(g) t(X1_list[[g]]) %*% y_list[[g]]
+      function(g) Matrix::t(X1_list[[g]]) %*% y_list[[g]]
     )
 
     tXgX1g <- lapply(
       seq_along(1:G),
-      function(g) t(X_list[[g]]) %*% X1_list[[g]]
+      function(g) Matrix::t(X_list[[g]]) %*% X1_list[[g]]
     )
 
     tX1X1 <- Reduce("+", tX1gX1g) # crossprod(X1)
     tX1y <- Reduce("+", tX1gyg) #t(X1) %*% y
-    tX1X1inv <- solve(tX1X1)
+    tX1X1inv <- Matrix::solve(tX1X1)
 
   }
 
@@ -159,19 +162,19 @@ boot_algo_fastnreliable <- function(
 
     inv_tXX_tXgXg <- lapply(
       1:G,
-      function(x) MASS::ginv(tXX - tXgXg[[x]])
+      function(x) solve(tXX - tXgXg[[x]])
     )
 
     beta_1g_tilde <- lapply(
       1:G,
-      function(g) MASS::ginv(tX1X1 - tX1gX1g[[g]]) %*% (tX1y - tX1gyg[[g]])
+      function(g) solve(tX1X1 - tX1gX1g[[g]]) %*% (tX1y - tX1gyg[[g]])
     )
 
   } else if(bootstrap_type == "WCU3x"){
 
     beta_g_hat <- lapply(
       1:G,
-      function(g) MASS::ginv(tXX - tXgXg[[g]]) %*% (tXy - tXgyg[[g]])
+      function(g) solve(tXX - tXgXg[[g]]) %*% (tXy - tXgyg[[g]])
     )
 
   }
@@ -186,7 +189,7 @@ boot_algo_fastnreliable <- function(
     if(is.null(inv_tXX_tXgXg)){
       inv_tXX_tXgXg <- lapply(
         1:G,
-        function(x) MASS::ginv(tXX - tXgXg[[x]])
+        function(x) solve(tXX - tXgXg[[x]])
       )
     }
   }
@@ -219,25 +222,26 @@ boot_algo_fastnreliable <- function(
[…]
