### 2023-05-28 #546 @markean
Hi @s3alfisc. Thank you for submitting your interesting package. I have attached my review below, which is based on the latest GitHub version. Hope it is helpful!

---

### Compliance with Standards

- [x] This package complies with a sufficient number of standards for a (bronze/silver/gold) badge
- [x] This grade of badge is the same as what the authors wanted to achieve

The following standards currently deemed non-applicable (through tags of `@srrstatsNA`) could potentially be applied to future versions of this software: (Please specify)

Please also comment on any standards which you consider either particularly well, or insufficiently, documented.

I have checked the standards via the `srr` package by running `srr::srr_report()`. The author has put all the tags in a single file and just stated that many standards have been applied (e.g., Done, Yes, Checked, etc.), which makes it rather difficult to see how each standard has been reflected in the code. It would be easier to review if the tags were placed in relevant locations of the package. Below I have some comments on the compliance with standards.

- G2.3b: It seems many character inputs are case sensitive.
For example, see below.
```
library(fwildclusterboot)
lm_fit <- lm(
  proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
  data = voters
)
boot <- boottest(lm_fit,
  B = 1000, param = "treatment", clustid = "group_id1",
  type = "NORM"
)
```
```
Error: in boottest.lm(lm_fit, B = 1000, param = "...:
 Argument 'type' must be a
single character equal to
'rademacher', 'mammen',
'norm', 'gamma' or 'webb'.
Problem: the value 'NORM'
doesn't match any choice
(note that no partial
matching is performed).
```

- RE1.0: Which parameter can be specified as a formula? Do you have any example?

- RE1.1: Should it be tagged as 'srrstatsNA'?

- RE3.0, RE3.1: Apart from the code quality, can the error/warning message be suppressed? This is not explicitly stated.

- RE5.0, RE6.1, RE6.2, RE6.3, RE7.0: No response can be found in the `srr-stats-standards.R` file. Are they all not applicable?

- G5.8a, G5.8b: Why are they not applicable? Are they tested?

- RE4.8 ~ RE7.4 (at the bottom of the `srr-stats-standards.R` file): No response can be found.

For packages aiming for silver or gold badges:

- [x] This package extends beyond minimal compliance with standards in the following ways: (please describe)

### General Review
- [x] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING
- [ ] The documentation is sufficient to enable general use of the package beyond one specific use case

I think adding 'Details' sections in the generic `boottest()` function would improve general accessability. Some high-level descriptions, such as what each method does depending on the object class or what assumptions are made, would be helpful. See, for example, https://r-pkgs.org/man.html#title-description-details. Although references are introduced in the documentation, many users would not be willing to search for the articles. Additionally, the reference section could adopt a coherent bibliographical style.

#### Algorithms
It seems there is no way to gracefully interrupt or abort the execution of the `boottest()` function in an R session.
This can be an issue especially in the bootstrap context where the iteration number could be very large.
It can be difficult address this issue when interfacing with other languages or APIs in R, but I was wondering if this has been considered in the package design.

#### Visualisation (where appropriate)
In the `plot()` method, it needs to be explained what the blue and red lines indicate either in the plot itself or in the documentation. Additionally, since it is a generic method, it would be good if the method supports other arguments. It does not accept some common arguments, which contradicts the usage of `...` in the documentation.
See below.
```
library(fwildclusterboot)
lm_fit <- lm(
  proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
  data = voters
)
boot <- boottest(lm_fit,
  B = 9999,
  param = "treatment",
  clustid = "group_id1"
)
plot(boot, main = "My Title")
```
```
Error: in plot.boottest(boot, main = "Title"):
 'main' is not a valid argument of function plot.boottest.
```

#### Package Design
- The argument `R` in the `boottest()` function currently seems to only accepts a numeric vector. Many existing R packages for hypothesis testing (e.g., the `car` package with the `hypothesis.matrix` argument in the `linearHypothesis()` function or the `multcomp` package with the `linfct` argument in the `glht()` function) allow users to specify a hypothesis symbolically (or at least using a character vector). Can this feature be added in the package? I believe new users to this package would expect this functionality.

- Can the seed mechanism be unified using only the `set.seed()` function? Many users would expect reproducibility by simply using `set.seed()`, not `dqrng::dqset.seed()`.

#### Others
The use of `rlang` for producing informative messages to the console is very good. However, unexpected line breaks occur for lengthy messages. This seems to be caused by manually applied or typed margin columns. Although it is aesthetically pleasing in the source code, I would recommend fixing this issue for general users. See below for example.
```
library(fwildclusterboot)
lm_fit <- lm(
  proposition_vote ~ treatment + ideology1 + log_income +
    Q1_immigration,
  data = voters
)
boot <- boottest(lm_fit,
  B = 9999,
  clustid = c("group_id1", "group_id2"),
  param = c("treatment", "ideology1"),
  R = c(1, 1, 1),
  r = 2
)
```
```
Error in `process_R()`:
! The constraints vector must either be NULL or a numeric of
           the same length as the `param` input vector.
```

---
