### 2023-05-29 #546 @s3alfisc
Hi @markean,

Thanks a lot for your thorough and very helpful review! One key takeaway from your and @meghapsimatrix's review is that I should assume less specialized knowledge from my package's users and explain multiple concepts in more detail. I will definitely work on that!

### Statistical Software Standards

I am sorry about not placing the standard tags in the code - I tried this at some point, but it soon appeared to me that my code would be clustered with tags if I handled it that way. Maybe I should have just done so anyways, and am happy to do it if it helps you assess how the package complies with standards, @markean.

Regarding the standards that you mentioned in more detail:

- RE1.0: Which parameter can be specified as a formula? Do you have any example?
   The `param` and `clustid` arguments can be specified as formula. E.g. you can run `boottest()` as
   ```r
   library(fwildclusterboot)
   data(voters)
   fit <- lm(proposition_vote ~ treatment, data = voters)
   boottest(fit, param = ~treatment, clustid = ~group_id1, B = 999)
   ```
- RE1.1: Should it be tagged as 'srrstatsNA'?
   Yes, you are right!
- RE3.0, RE3.1: Apart from the code quality, can the error/warning message be suppressed? This is not explicitly stated.
  No, this is currently not possible, and I will put a `verbose` argument on my to do list. Great suggestion, thanks!
- RE5.0, RE6.1, RE6.2, RE6.3, RE7.0: No response can be found in the srr-stats-standards.R file. Are they all not applicable? Yes, they are either not applicable, or I did not fully understand what they are asking for... 😅 I will make sure to add them to the non-applicable section and add a comment.

- G5.8a, G5.8b: Why are they not applicable? Are they tested?
   Regarding the zero length data, I thought it would be not applicable, as data is practically never provided to `boottest()`. For all other non-supported data types, I assumed that I should allow anything that passes through `lm()` / `model.matrix()`. The key bottleneck is for models with fixed effect estimation, where I internally transform all model fixed effects to a factor. I have currently [one open bug](https://github.com/s3alfisc/fwildclusterboot/issues/115), which occurs when there is a `date` variable specified in the model. In this case, the creation of the `model.frame()` of the clustering variable fails. I will have to fix this bug, and add a particular test case. Are there any other special variable types you think I should be testing?
- RE4.8 ~ RE7.4 (at the bottom of the srr-stats-standards.R file): No response can be found. Yes - for some cases, I did not even try to provide a response, as they appeared too "out of scope". I.e. for
  "RE6.3 Where a model object is used to generate a forecast (for example, through a [predict()](https://rdrr.io/r/stats/predict.html)   method), the default plot method should provide clear visual distinction between modelled (interpolated) and forecast (extrapolated) values."
  I did not provide an explanation for applicability or non-applicability, because prediction is clearly so much out of scope. For reasons of housekeeping, it would probably have been better if I had added a "out of scope" or "no forecasting functionality" comment. I will make sure to add these comments.

### Other comments

> I think adding 'Details' sections in the generic boottest() function would improve general accessability.

This is a great suggestion, I will make sure to add this.

> It seems there is no way to gracefully interrupt or abort the execution of the boottest() function in an R session.

In fact I currently don't know how to do this. The "main" algorithm is fully vectorized. The c++ code that cannot be interrupted is either a matrix multiplication via rcpp::eigen or the creation of the G X B bootstrap weights matrix. For the heteroskedastic bootstrap, there are some c++ for loops, and I think that rcpp provides options to (more or less, or even fully) gracefully abort. I will read up on this!
> In the plot() method, it needs to be explained what the blue and red lines indicate either in the plot itself or in the documentation. Additionally, since it is a generic method, it would be good if the method supports other arguments. It does not accept some common arguments, which contradicts the usage of ... in the documentation.

All great points, I'll work on that.

- Regarding multiple contrast tests - I can implement wild bootstrapped F-tests, and so far mostly have not done so because there are other R packages that handle this (e.g. @meghapsimatrix `wildmeta`) and because the algorithm in the fast and wild paper is quite cumbersome :D But based on the algorithms in the new-ish MacKinnon et al papers, it should be one day of work to implement a Wald test (note that I am always too optimistic with these things). Btw, note that a Wald test can be run via the `mboottest()` functions, but only through `WildBootTests.jl`.

> Can the seed mechanism be unified using only the set.seed() function? Many users would expect reproducibility by simply using set.seed(), not dqrng::dqset.seed().

Unfortunately, I don't think so. Generation of rademacher, webb and normal weights via `dqrng::dqset.seed()` is significantly faster than via e.g. `stats::sample()` (see the benchmark I have once posted [here](https://stackoverflow.com/questions/5458271/fast-sampling-in-r/69801770#69801770)). But unfortunately, the `dqrng` package does not support Mammen weights. though I think this option will be added soon. So I think that I could fully switch to `dqrng` for the random sampling for the cluster bootstrap. For the heteroskedastic bootstrap, I sample "directly in rcpp", which does not allow me to call drqrng (at least not without jumping through some hoops and calling it's c/c++ code directly, which I am afraid might be beyond my abilities). So for now, I don't think I can change this, or at least not without significant time investment. One solution could be the following: I could introduce a `fwildclusterboot::set_seed()` function, that calls `set.seed()` and `dqrng::dqset.seed()` and explicitly tells users that both functions are being called. This + very explicit documentation might be a good option?
- On `rlang` error message line breaks - I think I introduced these to pass the `goodpractice` line length checks 😅 There's probably a way to stay within the 80 character limit and not have these annoying breaks in the console. I'll do some googling / ask chatGPT =)

So, this was a very long response - sorry about that 😄. Thanks so much for your review @markean! I will try to address all your (and @meghapsimatrix's) feedback swiftly (I plan to work on it over the next weekend) and will ping you once there are updates / once I am finished.

Best, Alex
