### 2023-05-30 #546 @mpadge
@s3alfisc Perspective from the rOpenSci stats team is that standards documentation should definitely be done *inline*, at each point in the code at which compliance actually occurs. This enables both easy understanding of compliance, and far easier long-term maintenance for you, as you yourself then don't have to constantly switch back-and-forth between two separate locations for code and compliance documentation. We have also updated our check system to ensure that documentation of stats compliance is not concentrated in a single file, so your package may well fail that check if run again now? (If you want to see, it's probably best to set up our [`pkgcheck-action` workflow](https://github.com/ropensci-review-tools/pkgcheck-action) in your repo, rather than cluttering the review here with extra checks.) Hope that helps :smiley:
