+++ DESCRIPTION
@@ -9,6 +9,11 @@ Authors@R: c(
             comment = "Megha reviewed the package (v. 0.13) for ropensci 
                       , see <https://github.com/ropensci/software-review/issues/546>"
           ),
+    person("Eunseop", "Kim",
+            role = "rev", 
+            comment = "Eunseop reviewed the package (v. 0.13) for ropensci 
+                      , see <https://github.com/ropensci/software-review/issues/546>"
+          )
     person("Achim", "Zeileis", role = "ctb",
            comment = "Author of included sandwich fragments"),
     person("Nathaniel", "Graham", role = "ctb",
