### 2023-06-08 a98e711
add sparse matrix support for MNW bootstrap algo #116
