+++ R/boot_algo_fastnreliable.R
@@ -1,5 +1,5 @@
 #' WRE13, WRE33, WRU13 and WRU33 bootstraps as in MNW (2022) "Fast and
-#' reliable" and MacKinnon "Fast Cluster Bootstraps" 
+#' reliable" and MacKinnon "Fast Cluster Bootstraps"
 #' (Econometrics & Statistics, 2021)
 #'
 #' @param preprocessed_object A list: output of the preprocess2 function.
@@ -31,7 +31,7 @@
 #' @importFrom MASS ginv
 #' @importFrom Matrix crossprod tcrossprod t Matrix
 #' @importFrom summclust vcov_CR3J
-#' @return A list of bootstrap results. 
+#' @return A list of bootstrap results.
 #' @noRd
 
 
@@ -49,7 +49,7 @@ boot_algo_fastnreliable <- function(
   full_enumeration,
   small_sample_correction,
   object,
-  impose_null, 
+  impose_null,
   sampling){
 
 
@@ -64,13 +64,17 @@ boot_algo_fastnreliable <- function(
   }
 
   X <- preprocessed_object$X
+  y <- preprocessed_object$Y
+  
   # convert to sparse matrix
   X <- Matrix::Matrix(X)
-  y <- preprocessed_object$Y
+  y <- Matrix::Matrix(y)
+  
   R <- preprocessed_object$R0
   cluster_df <- preprocessed_object$clustid
   clustid <- names(cluster_df)
   fe <- preprocessed_object$fe
+  #' @srrstats {G2.4d} *explicit conversion to factor via `as.factor()`*
   cluster <- as.factor(cluster_df[,1])
   bootcluster <- preprocessed_object$bootcluster
   G <- N_G_bootcluster <- length(unique(bootcluster[[1]]))
@@ -89,14 +93,14 @@ boot_algo_fastnreliable <- function(
     type = type,
     full_enumeration = full_enumeration,
     N_G_bootcluster = N_G_bootcluster,
-    boot_iter = B, 
+    boot_iter = B,
     sampling = sampling
   )
 
   # create X_g's, X1_g's, y_g's etc
   X_list <- matrix_split(X, cluster, "row")
   y_list <- split(y, cluster, drop = FALSE)
-  
+
   # precompute a range of other objects
   tXgXg <- lapply(
     seq_along(1:G),
@@ -120,7 +124,7 @@ boot_algo_fastnreliable <- function(
   beta_g_hat <- NULL
   beta_1g_tilde <- NULL
   inv_tXX_tXgXg <- NULL
-  
+
 
   if(bootstrap_type %in% c("WCR3x", "WCU3x")){
     # X1: X without parameter beta for which hypothesis beta = 0 is tested
@@ -162,25 +166,25 @@ boot_algo_fastnreliable <- function(
 
     inv_tXX_tXgXg <- lapply(
       1:G,
-      function(x) solve(tXX - tXgXg[[x]])
+      function(x) MASS::ginv(as.matrix(tXX - tXgXg[[x]]))
     )
 
     beta_1g_tilde <- lapply(
       1:G,
-      function(g) solve(tX1X1 - tX1gX1g[[g]]) %*% (tX1y - tX1gyg[[g]])
+      function(g) MASS::ginv(as.matrix(tX1X1 - tX1gX1g[[g]])) %*% (tX1y - tX1gyg[[g]])
     )
 
   } else if(bootstrap_type == "WCU3x"){
 
     beta_g_hat <- lapply(
       1:G,
-      function(g) solve(tXX - tXgXg[[g]]) %*% (tXy - tXgyg[[g]])
+      function(g) MASS::ginv(as.matrix(tXX - tXgXg[[g]])) %*% (tXy - tXgyg[[g]])
     )
 
[…]
