### 2023-06-08 d940db1
more sparse matrix support for MNW algos #118
