+++ R/preprocess2.R
@@ -21,7 +21,7 @@ preprocess2.fixest <-
            param,
            fe,
            engine,
-           bootcluster, 
+           bootcluster,
            bootstrap_type) {
     #' preprocess data for objects of type fixest
     #'
@@ -36,7 +36,7 @@ preprocess2.fixest <-
     #'  "WildBootTests.jl"
     #' @param bootcluster a character string containing the name(s) of the
     #'  bootcluster variables. Alternatively, "min" or "max"
-    #' @param bootstrap_type Which bootstrap type should be run? 
+    #' @param bootstrap_type Which bootstrap type should be run?
     #' '11', '13', '31' or 33'. For more infos, see the paper
     #' by MacKinnon, Nielsen & Webb (2022)
     #'
@@ -61,11 +61,11 @@ preprocess2.fixest <-
       rlang::abort(
         "boottest() only supports OLS estimation via fixest::feols() - it
         does not support non-linear models computed via e.g. fixest::fepois()
-        or fixest::feglm.", 
+        or fixest::feglm.",
         use_cli_format = TRUE
       )
     }
-    
+
     # if (!is.null(object$is_sunab)) {
     #   if(object$is_sunab == TRUE){
     #     rlang::abort(
@@ -120,7 +120,7 @@ preprocess2.fixest <-
       if (!is.null(fe)) {
         rlang::abort(
           "boottest() unfortunately currently does not support WLS and fixed
-          effects. Please set fe = NULL to run a bootstrap with WLS.", 
+          effects. Please set fe = NULL to run a bootstrap with WLS.",
           use_cli_format = TRUE
         )
       }
@@ -140,8 +140,8 @@ preprocess2.fixest <-
         fe = fe,
         N = N,
         has_weights = has_weights,
-        engine = engine, 
-        bootstrap_type = bootstrap_type, 
+        engine = engine,
+        bootstrap_type = bootstrap_type,
         clustid_char = clustid
       )
 
@@ -248,7 +248,7 @@ preprocess2.felm <-
            param,
            fe,
            engine,
-           bootcluster, 
+           bootcluster,
            bootstrap_type) {
     #' preprocess data for objects of type felm
     #'
@@ -263,8 +263,8 @@ preprocess2.felm <-
     #'  "WildBootTests.jl"
     #' @param bootcluster a character string containing the name(s) of the
     #'  bootcluster variables. Alternatively, "min" or "max"
-    #' @param bootstrap_type Which bootstrap type should be run? 
-    #' '11', '13', '31' or  33'. For more infos, 
+    #' @param bootstrap_type Which bootstrap type should be run?
+    #' '11', '13', '31' or  33'. For more infos,
     #' see the paper by MacKinnon, Nielsen & Webb (2022)
     #'
     #' @noRd
@@ -297,7 +297,7 @@ preprocess2.felm <-
       rlang::abort(
         "IV regression is currently not supported by boottest() for
         objects of type 'felm'. You can either use 'fixest::feols()'
-        or 'ivreg::ivreg' for IV-regression.", 
+        or 'ivreg::ivreg' for IV-regression.",
         use_cli_format = TRUE
       )
       is_iv <- TRUE
@@ -333,8 +333,8 @@ preprocess2.felm <-
         fe = fe,
         N = N,
         has_weights = has_weights,
-        engine = engine, 
-        bootstrap_type = bootstrap_type, 
+        engine = engine,
+        bootstrap_type = bootstrap_type,
         clustid_char = clustid
       )
       X <- get_fe$X
@@ -396,7 +396,7 @@ preprocess2.felm <-
       R0[, 1:p] <- R
     }
     # }
-    
+
     res <- list(
       Y = Y,
       X = X,
[…]
