### 2023-06-09 b69217b
fix bug in plot method when no conf int is computed
