### 2023-06-10 7a00a08
faster Moore Penrose inverses via pinv
