+++ NAMESPACE
@@ -36,6 +36,11 @@ importFrom(JuliaConnectoR,juliaEval)
 importFrom(JuliaConnectoR,juliaImport)
 importFrom(MASS,ginv)
 importFrom(Matrix,Diagonal)
+importFrom(Matrix,Matrix)
+importFrom(Matrix,crossprod)
+importFrom(Matrix,sparse.model.matrix)
+importFrom(Matrix,t)
+importFrom(Matrix,tcrossprod)
 importFrom(Rcpp,evalCpp)
 importFrom(collapse,GRP)
 importFrom(collapse,fsum)
