### 2023-06-10 108af9d
allow to interrupt cpp loops #119
