+++ src/boot_algo3_cpp.cpp
@@ -26,6 +26,8 @@ arma::vec boot_algo3_crv1_denom(int B,
 #pragma omp parallel for num_threads(cores)
   for(int b = 0; b < (B+1); b++){
     
+    Rcpp::checkUserInterrupt()
+    
     //double Z_sq;
     arma::vec Zg(G);
     for(int g = 0; g < G; g++){
@@ -115,6 +117,8 @@ List boot_algo3_crv3( const int B,
   
 #pragma omp parallel for num_threads(cores)
   for(int b = 0; b < B +1; b++){
+    
+    Rcpp::checkUserInterrupt()
     
     arma::mat delta_diff(G,k);
     
