+++ src/wildboottest.cpp
@@ -147,7 +147,7 @@ List wildboottestHC(const arma::vec & y,
 #pragma omp parallel for num_threads(cores)
   for(int b = 1; b < B + 1; b++){
     
-    Rcpp::checkUserInterrupt()
+    Rcpp::checkUserInterrupt();
     
     // create bootstrap sample
     //arma::vec weights = RcppArmadillo::sample(s, N_G_bootcluster, true, prob);
@@ -235,7 +235,7 @@ List wildboottestCL(const arma::vec & y,
 #pragma omp parallel for num_threads(cores)
   for(int b = 1; b < B + 1; b++){
     
-    Rcpp::checkUserInterrupt()
+    Rcpp::checkUserInterrupt();
     
     // create bootstrap sample
     arma::vec y_boot(n);
@@ -358,7 +358,7 @@ List wildboottestCL_enum(const arma::vec & y,
 #pragma omp parallel for num_threads(cores)
   for(int b = 1; b < B + 1; b++){
     
-    Rcpp::checkUserInterrupt()
+    Rcpp::checkUserInterrupt();
     
     // create bootstrap sample
     arma::vec y_boot(n);
