### 2023-06-11 ac43c69
add eigen_pinv support (large speed gains)
