### 2023-06-17 e0360be
delete print method
