### 2023-06-17 a74de20
fix eigen_pinv bug
