### 2023-06-17 a2af254
drop support for print.mboottest
