+++ R/methods.R
@@ -750,26 +750,26 @@ nobs.mboottest <- function(object, ...) {
 #' )
 #' print(boot)
 #'
-print.boottest <- function(x, ..., digits = 4) {
-  stopifnot(inherits(x, "boottest"))
-
-  print(x$call)
-  cat("", "\n")
-
-  vals <- lapply(
-    c("p_val", "conf_int", "t_stat"),
-    function(y) {
-      ifelse(is.null(x[[y]]), 
-        'not computed',
-        round(x[[y]], digits = digits)
-      )
-    }
-  )
-
-  cat("p value:", vals[[1]], "\n")
-  cat("confidence interval:", vals[[2]], "\n")
-  cat("test statistic", vals[[3]], "\n")
-}
+# print.boottest <- function(x, ..., digits = 4) {
+#   stopifnot(inherits(x, "boottest"))
+# 
+#   print(x$call)
+#   cat("", "\n")
+# 
+#   vals <- lapply(
+#     c("p_val", "conf_int", "t_stat"),
+#     function(y) {
+#       ifelse(is.null(x[[y]]), 
+#         'not computed',
+#         round(x[[y]], digits = digits)
+#       )
+#     }
+#   )
+# 
+#   cat("p value:", vals[[1]], "\n")
+#   cat("confidence interval:", vals[[2]], "\n")
+#   cat("test statistic", vals[[3]], "\n")
+# }
 
 
 #' S3 method to print key information for objects of type `mboottest`
@@ -794,22 +794,22 @@ print.boottest <- function(x, ..., digits = 4) {
 #' print(wboottest)
 #' }
 #'
-print.mboottest <- function(x, ..., digits = 4) {
-  stopifnot(inherits(x, "mboottest"))
-
-  print(x$call)
-  cat("", "\n")
-
-  vals <- lapply(
-    c("p_val", "teststat"),
-    function(y) {
-      round(x[[y]], digits = digits)
-    }
-  )
-
-  cat("p value:", vals[[1]], "\n")
-  cat("test statistic", vals[[2]], "\n")
-}
+# print.mboottest <- function(x, ..., digits = 4) {
+#   stopifnot(inherits(x, "mboottest"))
+# 
+#   print(x$call)
+#   cat("", "\n")
+# 
+#   vals <- lapply(
+#     c("p_val", "teststat"),
+#     function(y) {
+#       round(x[[y]], digits = digits)
+#     }
+#   )
+# 
+#   cat("p value:", vals[[1]], "\n")
+#   cat("test statistic", vals[[2]], "\n")
+# }
 
 
 tidy.mboottest <- function(x, ...) {
