+++ NAMESPACE
@@ -14,7 +14,6 @@ S3method(nobs,boottest)
 S3method(nobs,mboottest)
 S3method(plot,boottest)
 S3method(print,boottest)
-S3method(print,mboottest)
 S3method(pval,boottest)
 S3method(pval,mboottest)
 S3method(summary,boottest)
@@ -22,7 +21,6 @@ S3method(summary,mboottest)
 S3method(teststat,boottest)
 S3method(teststat,mboottest)
 S3method(tidy,boottest)
-S3method(tidy,mboottest)
 export(boot_aggregate)
 export(boot_ssc)
 export(boottest)
