+++ R/mboottest_fixest.R
@@ -105,8 +105,10 @@
 #'
 #' @section Setting Seeds:
 #' To guarantee reproducibility, you need to
-#' set a global random seed via`set.seed()` 
-#'
+#' set a global random seed via`set.seed()`
+#' @section Multiple Fixed Effects:
+#' If your feols() model contains fixed effects, boottest() will internally convert all fixed
+#' effects but the one specified via the `fe` argument to dummy variables.
 #' @references Roodman et al., 2019, "Fast and wild: Bootstrap inference in
 #' STATA using boottest", The STATA Journal.
 #' (<https://ideas.repec.org/p/qed/wpaper/1406.html>)
@@ -173,56 +175,56 @@ mboottest.fixest <- function(object,
                              ),
                              ...) {
   call <- match.call()
-  
+  type <- tolower(type)
   dreamerr::validate_dots(stop = TRUE)
-  
+
   # Step 1: check arguments of feols call
   check_arg(object, "MBT class(fixest)")
   check_arg(clustid, "character scalar | character vector | formula")
   check_arg(B, "MBT scalar integer")
   check_arg(R, "MBT numeric matrix")
-  
+
   check_arg(type, "charin(rademacher, mammen, norm, gamma, webb)")
   check_arg(p_val_type, "charin(two-tailed, equal-tailed,>, <)")
-  
+
   check_arg(r, "numeric vector | NULL")
   check_arg(fe, "character scalar | NULL | formula")
   check_arg(bootcluster, "character vector | formula")
   check_arg(tol, "numeric scalar GT{0}")
   # check_arg(maxiter, "scalar integer")
   check_arg(boot_ssc, "class(ssc) | class(boot_ssc)")
-  
+
   check_arg(floattype, "charin(Float32, Float64)")
   check_arg(maxmatsize, "scalar integer | NULL")
   check_arg(bootstrapc, "scalar logical")
-  
+
   inform_seed(
-    frequency_id = "seed-reminder-m-fixest", 
+    frequency_id = "seed-reminder-m-fixest",
     engine = "WildBootTests.jl"
   )
-  
+
   if (inherits(clustid, "formula")) {
     clustid <- attr(terms(clustid), "term.labels")
   }
-  
+
   if (inherits(bootcluster, "formula")) {
     bootcluster <- attr(terms(bootcluster), "term.labels")
   }
-  
+
   if (inherits(fe, "formula")) {
     fe <- attr(terms(fe), "term.labels")
   }
-  
+
   # fixest specific checks
   if (object$method != "feols") {
     rlang::abort(
       c("mboottest() only supports OLS estimation via fixest::feols() - it
          does not support non-linear models computed via e.g.
-         fixest::fepois() or fixest::feglm."), 
+         fixest::fepois() or fixest::feglm."),
       use_cli_format = TRUE
     )
   }
-  
+
   if (!is.null(object$fixef_removed)) {
     rlang::abort(
       paste(
@@ -231,20 +233,20 @@ mboottest.fixest <- function(object,
                internal pre-processing does not account for this deletion.
                Therefore, please exclude such fixed effects prior
                to estimation with feols(). You can find them listed under
-               '$fixef_removed' of your fixest object."), 
+               '$fixef_removed' of your fixest object."),
         use_cli_format = TRUE
     )
   }
-  
+
   fedfadj <- 0L
-  
+
   check_mboottest_args_plus(
     object = object,
     R = R,
     r = r,
     fe = fe
   )
-  
+
   preprocess <- preprocess2.fixest(
     object = object,
[…]
