+++ R/boottest_lm.R
@@ -3,7 +3,7 @@
 
 #' `boottest.lm` is a S3 method that allows for fast wild cluster
 #' bootstrap inference for objects of class lm by  implementing
-#'fast wild bootstrap algorithms as developed in Roodman et al., 2019 
+#'fast wild bootstrap algorithms as developed in Roodman et al., 2019
 #' and MacKinnon, Nielsen & Webb (2022).
 #'
 #' @param object An object of class lm
@@ -36,7 +36,7 @@
 #' @param conf_int A logical vector. If TRUE, boottest computes confidence
 #'        intervals by test inversion. If FALSE, only the p-value is returned.
 #' @param engine Character scalar. Either "R", "R-lean" or "WildBootTests.jl".
-#'  Controls if `boottest()` should run via its native R implementation 
+#'  Controls if `boottest()` should run via its native R implementation
 #'  or `WildBootTests.jl`.
 #'  "R" is the default and implements the cluster bootstrap
 #'  as in Roodman (2019). "WildBootTests.jl" executes the
@@ -50,13 +50,13 @@
 #'  defaults to the "lean" algorithm. You can set the employed
 #'  algorithm globally by using the
 #'  `setBoottest_engine()` function.
-#' @param bootstrap_type Determines which wild cluster bootstrap type should be 
-#' run. Options are "fnw11","11", "13", "31" and "33" for the wild cluster 
+#' @param bootstrap_type Determines which wild cluster bootstrap type should be
+#' run. Options are "fnw11","11", "13", "31" and "33" for the wild cluster
 #' bootstrap and "11" and "31" for the heteroskedastic bootstrap.
-#' For more information, see the details section. "fnw11" is the default for 
-#' the cluster bootstrap, which runs a "11" type 
-#' wild cluster bootstrap via the algorithm outlined in "fast and wild" 
-#' (Roodman et al (2019)). "11" is the default for the heteroskedastic 
+#' For more information, see the details section. "fnw11" is the default for
+#' the cluster bootstrap, which runs a "11" type
+#' wild cluster bootstrap via the algorithm outlined in "fast and wild"
+#' (Roodman et al (2019)). "11" is the default for the heteroskedastic
 #' bootstrap.
 #' @param R Hypothesis Vector giving linear combinations of coefficients.
 #' Must be either NULL or a vector of the same length as `param`. If NULL,
@@ -66,7 +66,8 @@
 #' @param beta0 Deprecated function argument. Replaced by function argument 'r'.
 #' @param type character or function. The character string specifies the type
 #'        of boostrap to use: One of "rademacher", "mammen", "norm"
-#'        and "webb". Alternatively, type can be a function(n) for drawing
+#'        and "webb".
+#'        Alternatively, type can be a function(n) for drawing
 #'        wild bootstrap factors. "rademacher" by default.
 #'        For the Rademacher distribution, if the number of replications B
 #'        exceeds the number of possible draw ombinations, 2^(#number
@@ -108,11 +109,11 @@
 #' bootstrap-c instead of bootstrap-t. Only relevant when 'engine =
 #' "WildBootTests.jl"'
 #' @param sampling 'dqrng' or 'standard'. If 'dqrng', the 'dqrng' package is
-#' used for random number generation (when available). If 'standard', 
-#' functions from the 'stats' package are used when available. 
-#' This argument is mostly a convenience to control random number generation in 
[…]
