### 2023-06-17 2b9eeb6
update plot documentation
