+++ tests/testthat/test-new-bootstrap-variants.R
@@ -1,4 +1,5 @@
 test_that("test r-fnw vs r-, stochastic", {
+  
   skip_on_cran()
   skip_if_not(
     fwildclusterboot:::find_proglang("julia"),
