+++ R/boottest_felm.R
@@ -200,6 +200,14 @@
 #' @section Multiple Fixed Effects:
 #' If your felm() model contains fixed effects, boottest() will internally convert all fixed
 #' effects but the one specified via the `fe` argument to dummy variables.
+#' @section Run `boottest` quietly:
+#' You can suppress all warning and error messages by setting the following global
+#' options:
+#' `options(rlib_warning_verbosity = "quiet")`
+#' `options(rlib_message_verbosity = "quiet")`
+#' Not that this will turn off all warnings (messages) produced via `rlang::warn()` and
+#' `rlang::inform()`, which might not be desirable if you use other software build on
+#' `rlang`, as e.g. the `tidyverse`.
 #' @section Stata, Julia and Python Implementations:
 #' The fast wild cluster bootstrap algorithms are further implemented in the
 #' following software packages:
