+++ R/boottest_ivreg.R
@@ -115,7 +115,14 @@
 #' @section Setting Seeds:
 #' To guarantee reproducibility, you need to
 #' set a global random seed via `set.seed()`
-#'
+#' @section Run `boottest` quietly:
+#' You can suppress all warning and error messages by setting the following global
+#' options:
+#' `options(rlib_warning_verbosity = "quiet")`
+#' `options(rlib_message_verbosity = "quiet")`
+#' Not that this will turn off all warnings (messages) produced via `rlang::warn()` and
+#' `rlang::inform()`, which might not be desirable if you use other software build on
+#' `rlang`, as e.g. the `tidyverse`.
 #' @references Roodman et al., 2019, "Fast and wild: Bootstrap inference in
 #'    STATA using boottest", The STATA Journal.
 #'    (<https://ideas.repec.org/p/qed/wpaper/1406.html>)
