### 2023-06-17 cf7be51
update section on plot method in vignette #122
