### 2023-06-17 d63856a
enforce pvalues, tstats to be a a flat vector
