### 2023-06-17 s3alfisc/fwildclusterboot#115 @s3alfisc
It looks like `Formula` and `expand.model.frame` do not handle date variables in the second part of the formula.

```r
        suppressWarnings(
          expand.model.frame(
            model =
              manipulate_object(object),
            extras = clustid_fml,
            na.expand = FALSE,
            envir = call_env
          )
        )
```
The error I receive is

```r
Error in Ops.Date(1, date) : | not defined for "Date" objects
Called from: Ops.Date(1, date)
```.

`sandwich` fails in the same context as well:

```r
library(sandwich)
library(fwildclusterboot)

data(voters)
date <- sample(1:7, nrow(voters), TRUE)
voters$date <- as.Date(date,origin = "1970-01-01")

sapply(voters, class)

fit <- feols(proposition_vote ~ treatment | date, data = voters)
sandwich::vcovCL(fit, ~date)
# Error in Ops.Date(treatment, date) : | not defined for "Date" objects

fwildclusterboot::boottest(fit, param = "treatment", clustid = "group_id", B = 999)
# Error in Ops.Date(1, date) : | not defined for "Date" objects
```
Tagging @zeileis here for awareness (in case you are not aware already).

For now, I will label this as "won't fix".
