+++ NAMESPACE
@@ -41,8 +41,10 @@ importFrom(Matrix,t)
 importFrom(Matrix,tcrossprod)
 importFrom(Rcpp,evalCpp)
 importFrom(collapse,GRP)
+importFrom(collapse,fmean)
 importFrom(collapse,fsum)
 importFrom(collapse,qF)
+importFrom(collapse,qtab)
 importFrom(dqrng,dqsample)
 importFrom(dqrng,dqset.seed)
 importFrom(dreamerr,check_arg)
