### 2023-06-17 b7dc2a6
faster computation of H #113
