### 2023-06-17 s3alfisc/fwildclusterboot#115 @zeileis
I think `Formula` is not involved here, is it? The error message would indicate that standard `formula` processing (as opposed to `Formula`) is used. To `expand.model.frame` the model specification looks like a standard `formula` and hence treats `|` in the basic way and not for separating a model part. I would recommend to be explicit and use `... | factor(date)` instead. Then `expand.model.frame()` seems to work again.
