### 2023-06-18 8e6b8ce
revert use of collapse::qtab, introduces test failures #117
