+++ vignettes/articles/bibliography.bib
@@ -1,0 +1,540 @@
+@article{arel2022modelsummary,
+  title={modelsummary: Data and Model Summaries in R},
+  author={Arel-Bundock, Vincent},
+  journal={Journal of Statistical Software},
+  volume={103},
+  pages={1--23},
+  year={2022}
+}
+
+@book{CameronTrivedi2013,
+  author    = {A. Colin Cameron and Pravin K. Trivedi},
+  title     = {Regression Analysis of Count Data},
+  year      = {2013},
+  edition   = {2nd},
+  publisher = {Cambridge University Press},
+  address   = {Cambridge}
+}
+
+@book{ChambersHastie1992,
+  editor    = {John M. Chambers and Trevor J. Hastie},
+  title     = {Statistical Models in S},
+  publisher = {Chapman \& Hall},
+  year      = {1992},
+  address   = {London}
+}
+
+@manual{Jackman2015,
+  title  = {pscl: Classes and Methods for R Developed in the Political Science Computational Laboratory, Stanford University},
+  author = {Simon Jackman},
+  year   = {2015},
+  note   = {R package version 1.4.9},
+  url    = {https://CRAN.R-project.org/package=pscl}
+}
+
+@article{Mullahy1986,
+  author  = {John Mullahy},
+  title   = {Specification and Testing of Some Modified Count Data Models},
+  year    = {1986},
+  journal = {Journal of Econometrics},
+  volume  = {33},
+  number  = {3},
+  pages   = {341--365},
+  doi     = {10.1016/0304-4076(86)90002-3}
+}
+
+@book{McCullaghNelder1989,
+  author    = {Peter McCullagh and John A. Nelder},
+  title     = {Generalized Linear Models},
+  edition   = {2nd},
+  year      = {1989},
+  publisher = {Chapman \& Hall},
+  address   = {London},
+  doi       = {10.1007/978-1-4899-3242-6}
+}
+
+@manual{R,
+  title        = {R: {A} Language and Environment for Statistical Computing},
+  author       = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address      = {Vienna, Austria},
+  year         = {2017},
+  url          = {https://www.R-project.org/}
+}
+
+@article{StasinopoulosRigby2007,
+  author  = {D. Mikis Stasinopoulos and Robert A. Rigby},
+  title   = {Generalized Additive Models for Location Scale and Shape ({GAMLSS}) in R},
+  journal = {Journal of Statistical Software},
+  year    = {2007},
+  volume  = {23},
+  number  = {7},
+  pages   = {1--46},
+  doi     = {10.18637/jss.v023.i07}
+}
+
+@book{VenablesRipley2002,
+  author    = {William N. Venables and Brian D. Ripley},
+  title     = {Modern Applied Statistics with S},
+  edition   = {4th},
+  year      = {2002},
+  pages     = {495},
+  publisher = {Springer-Verlag},
+  address   = {New York},
+  doi       = {10.1007/978-0-387-21706-2}
+}
+
+@book{Wood2006,
+  author    = {Simon N. Wood},
+  title     = {Generalized Additive Models: An Introduction with R},
+  year      = {2006},
+  publisher = {Chapman \& Hall/CRC},
+  address   = {Boca Raton}
+}
+
+@article{Yee2009,
+  author  = {Thomas W. Yee},
+  title   = {The VGAM Package for Categorical Data Analysis},
+  journal = {Journal of Statistical Software},
+  year    = {2010},
+  volume  = {32},
+  number  = {10},
+  pages   = {1--34},
+  doi     = {10.18637/jss.v032.i10}
+}
+
+@article{ZeileisKleiberJackman2008,
[…]
