+++ vignettes/articles/wild_bootstrap.Rmd
@@ -1,0 +1,365 @@
+---
+title: "The Wild (Cluster) Bootstrap"
+bibliography: bibliography.bib
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+
+### Wild Cluster Bootstrap 101
+
+We are interested in testing a linear hypothesis $H_0: \beta_{j} = 0$ against $H_1: \beta_{j} \neq 0$
+
+for a linear model of interest
+
+\begin{align}
+y &= X \beta + u \\
+  &=
+\begin{bmatrix}
+y_{1} \\
+y_{2} \\
+...\\
+y_{G}
+\end{bmatrix}
+=
+\begin{bmatrix}
+X_{1} \\
+X_{2} \\
+...\\
+X_{G}
+\end{bmatrix} \beta +
+\begin{bmatrix}
+u_{1} \\
+u_{2} \\
+...\\
+u_{G}
+\end{bmatrix},
+\end{align}
+
+with $E(u|X) = 0$, where group $g$ contains $N_{g}$ observations so that $N = \sum_{g = 1}^{G} N_{g}$. The regression residuals $u$ are allowed to be correlated within clusters, but are assumed to be uncorrelated across clusters. In consequence, the models' covariance matrix is block diagonal. For each cluster, we denote $E(u_{g} u_{g}') =\Sigma_{g}$ and $E(u u') =\Sigma$.
+
+A generic wild bootstrap test then proceeds in the following steps: 
+
+::: callout
+* Step 1: Either
+  - ... regress $y$ on $X$; obtain point estimates $\hat{\beta}$ and estimated residuals $\hat{u}$; This type of bootstrap is known as the "WCU" (wild cluster unrestricted)
+   - ... or regress $Y$ on $X$ imposing the null (via restricted least squares); obtain point estimates $\tilde{\beta}$ and estimated residuals $\tilde{u}$. This type of bootstrap is known as the "WCR" (wild cluster restricted)
+   - to unify notation, we will denote $\ddot{\beta}$,      with $\ddot{\beta} = \hat{\beta}$ for the WCU and      $\ddot{\beta} = \tilde{\beta}$ for the WCR
+
+* Step 2:
+  - calculate the (cluster) robust t-statistic for the null hypothesis of interest, i.e.
+  $$
+    t_{j} = \frac{\hat{\beta_{j}}}{\sqrt{\hat{\Sigma}_{jj,x}}}
+  $$
+The subscript $x$ for the variance-covariance estimate $\hat{\Sigma}_{jj,x}$ denotes the type $x$ of the employed cluster robust variance estimator, and the $jj$ subscribt selects the jth column and row of the variance covariance matrix.
+
+* Step 3:
+  - For $b = 1, ..., B$ bootstrap iterations, create
+     - a set of bootstrap error terms $v_{b}^{*} \in \mathbb{R}^{G}$
+     - a new bootstrap sample $y_{b}^{*} = X \ddot{\beta} + f_{y}(\hat{u}) \times v_{b}^{*}$, where for each observation from the same cluster g, the transformed predicted error $f_{y}(\hat{u})$ is multiplied with the same weight $v_{b,g}^{*}$
+     - a bootstrapped t-statistic
+      $$
+        t_{j,b}^{*} = \frac{\hat{\beta_{j,b}^{*}} - \ddot{\beta_{j}}}{\sqrt{\hat{\Sigma}_{jj,x,b}^{*}}}
+      $$
[…]
