### 2023-06-18 b7cda39
cleanups and update readme
