+++ NEWS.md
@@ -1,3 +1,51 @@
+# fwildclusterboot 0.14
+
+## Performance
+
+- sparsifies MNW bootstraps - bootstrap types 31, 33, 13 (large speed gains for problems with high dimensional fixed effects)
+
+Here is a speed comparison for a problem with $k=100$ dummies from a 
+high dimensional fixed effect: 
+```r
+library(fwildclusterboot)
+
+voters <-
+  fwildclusterboot:::create_data(
+    N = 100000,
+    N_G1 = 20,
+    icc1 = 0.01,
+    N_G2 = 10,
+    icc2 = 0.01,
+    numb_fe1 = 100,
+    numb_fe2 = 10,
+    seed = 71986045
+  )
+
+lm_fit <- lm(proposition_vote ~ treatment + ideology1 + log_income +
+               Q1_immigration,
+             data = voters
+)
+
+microbenchmark::microbenchmark(
+
+)
+```
+
+- computes pinv via rcpp eigen (large speed gains)
+- it is now possible to interrupt rcpp loops
+- unlocks parallelization (nthreads was internally set to 1 for some reason)
+
+## rOpenSci Review feedback 
+
+- update docs:
+  - add a vignette on wild bootstrap concepts
+  - better explanation of plot method in docs and vignette 
+  - some guidelines on how to turn messages and warnings off
+- reorganization of ropensci ssr tags into code
+
+
+
+
 # fwildclusterboot 0.13
 
 ## Potentially Breaking Changes: 
