### 2023-06-18 cbaa84d
only use pinv when standard inversion fails
