### 2023-06-18 s3alfisc/fwildclusterboot#115 @s3alfisc
Thanks for your feedback, Achim. Indeed you are right that `Formula` is not involved =) I'll close this issue, as I don't think I will fix this in the nearer future. Hope this is ok with you @95goo ?
