### 2023-06-19 520628d
allow to project out cluster fixed effects for 11, 31 bootstrap types #124
