### 2023-06-19 87cf286
move computation of H to rcpp
