### 2023-06-25 1b28140
clear error message when clustid != bootcluster for fastnreliable algos
