### 2023-06-25 80be119
update news, wild bootstrap 101 vignette
