### 2023-06-27 9ffad72
fix bug in boot_aggregate with HC errors
