### 2023-07-02 347b2db
move fixest preprocessing to fixest::fetch_data
