### 2023-07-02 5e6b539
add error when running boottest() with varying slopes
