+++ R/arg_checks.R
@@ -117,6 +117,25 @@ check_boottest_args_plus <- function(
       rlang::abort("Advanced formula notation in fixest / fixest via ^ to interact
           fixed effects is currently not supported in boottest().")
     }
+    
+    if (
+      # '^' illegal in fixef argument, but legal in main formula -
+      # e.g. fml = y ~ x1 + I(x2^2) shold be possible
+      #' @srrstats {G2.4c} *explicit conversion to character via `as.character()`
+      #' (and not `paste` or `paste0`)* Done
+      
+      ("fixef_vars" %in% names(object) &&
+       grepl("[",
+             Reduce(paste, as.character(as.formula(object$fml_all$fixef))),
+             fixed = TRUE
+       ))
+      # note: whitespace ~ - for IV
+      # grepl("~", deparse_fml, fixed = TRUE)
+    ) {
+      rlang::abort("Varying slopes in fixest / fixest via [] to interact
+          fixed effects is currently not supported in boottest().")
+    }
+    
 
 
     if (!is.null(fe)) {
