### 2023-07-02 d5bf18a
revert fetch_data for fixest
