+++ R/preprocess2.R
@@ -11,14 +11,9 @@
 #'  routines as part of initial pre-processing to ensure that all other
 #'  sub-functions of a package receive inputs of a single defined class or
 #'  type.* Similar preprocess method + run_bootstrap function.
-
-
 preprocess2 <- function(object, ...) {
   UseMethod("preprocess2")
 }
-
-
-
 preprocess2.fixest <-
   function(object,
            clustid,
@@ -48,18 +43,14 @@ preprocess2.fixest <-
     #' @noRd
     #'
     #' @method preprocess2 fixest
-
     call <- object$call
     call_env <- object$call_env
     fml <- formula(object)
     fml <- Formula::as.Formula(fml)
-
     fml_full <- formula(fml, collapse = TRUE)
-
     N <- nobs(object)
     # lm and felm don't drop NAs due to multicollinearity, while fixest does
     k <- length(na.omit(coef(object)))
-
     method <- object$family
     # fixest specific checks
     if (object$method != "feols") {
@@ -70,7 +61,6 @@ preprocess2.fixest <-
         use_cli_format = TRUE
       )
     }
-
     # if (!is.null(object$is_sunab)) {
     #   if(object$is_sunab == TRUE){
     #     rlang::abort(
@@ -79,40 +69,36 @@ preprocess2.fixest <-
     #     )
     #   }
     # }
-
     is_iv <- ifelse(!is.null(object$fml_all$iv), TRUE, FALSE)
     has_fe <- ifelse(!is.null(object$fml_all$fixef), TRUE, FALSE)
-
     if (!is_iv) {
       X <-
         model.matrix(object,
-          type = "rhs",
-          na.rm = TRUE,
-          collin.rm = TRUE
+                     type = "rhs",
+                     na.rm = TRUE,
+                     collin.rm = TRUE
         )
     } else {
       X_endog <-
         model.matrix(object,
-          type = "iv.endo",
-          na.rm = TRUE,
-          collin.rm = TRUE
+                     type = "iv.endo",
+                     na.rm = TRUE,
+                     collin.rm = TRUE
         )
       X_exog <-
         model.matrix(object,
-          type = "iv.exo",
-          na.rm = TRUE,
-          collin.rm = TRUE
+                     type = "iv.exo",
+                     na.rm = TRUE,
+                     collin.rm = TRUE
         )
       instruments <-
         model.matrix(object,
-          type = "iv.inst",
-          na.rm = TRUE,
-          collin.rm = TRUE
+                     type = "iv.inst",
+                     na.rm = TRUE,
+                     collin.rm = TRUE
         )
     }
-
     Y <- model.matrix(object, type = "lhs")
-
     weights <- weights(object)
     if (is.null(weights)) {
       has_weights <- FALSE
@@ -120,7 +106,6 @@ preprocess2.fixest <-
     } else {
       has_weights <- TRUE
     }
-
     if (has_weights) {
       if (!is.null(fe)) {
         rlang::abort(
@@ -130,12 +115,9 @@ preprocess2.fixest <-
         )
       }
     }
-
     fixed_effect <- NULL
     k2 <- 0
     W <- n_fe <- NULL
-
-
     if (has_fe) {
       # if(!is.null(fe)){
       get_fe <- transform_fe(
@@ -149,8 +131,6 @@ preprocess2.fixest <-
         bootstrap_type = bootstrap_type,
         clustid_char = clustid
       )
-
-
       X <- get_fe$X
       Y <- get_fe$Y
       fixed_effect <- get_fe$fixed_effect
@@ -159,8 +139,6 @@ preprocess2.fixest <-
[…]
