### 2023-07-02 7e70aa3
slightly clean up docs
