### 2023-07-04 a827a79
fix cmd check
