+++ R/boot_aggregate.R
@@ -398,7 +398,7 @@ boot_aggregate <- function(
     pvalues[i] <- pval(boot_fit)
     teststat[i] <- teststat(boot_fit)
     if(!is.null(clustid)){
-        conf_int[i,] <- fwildclusterboot:::confint.boottest(boot_fit)
+        conf_int[i,] <- confint.boottest(boot_fit)
     } else {
       conf_int[i,] <- rep(NA, 2)
     }
