+++ man/boottest.Rd
@@ -112,8 +112,8 @@ Webb, Matthew D.
 . No. 1315. Queen's Economics Department Working Paper, 2013.
 }
 \seealso{
-\link{boottest.lm},
-\link{boottest.fixest},
-\link{boottest.felm},
-\link{boottest.ivreg}
+\link[=boottest.lm]{boottest.lm},
+\link[=boottest.fixest]{boottest.fixest},
+\link[=boottest.felm]{boottest.felm},
+\link[=boottest.ivreg]{boottest.ivreg}
 }
