### 2023-07-05 00aa28e
allow to project out cluster fixed effects
