+++ tests/testthat/test-new-bootstrap-variants.R
@@ -506,16 +506,13 @@ test_that("new variants and fixed effects", {
 test_that("test cluster fixed effects", {
   
 
-  library(fixest)
-  library(fwildclusterboot)
-  
   B <- 9999
   
   data1 <<- fwildclusterboot:::create_data(
     N = 1000,
-    N_G1 = 20,
+    N_G1 = 30,
     icc1 = 0.5,
-    N_G2 = 20,
+    N_G2 = 10,
     icc2 = 0.2,
     numb_fe1 = 10,
     numb_fe2 = 10,
@@ -523,59 +520,52 @@ test_that("test cluster fixed effects", {
     weights = 1:N / N
   )
   
-  feols_fit <- feols(proposition_vote ~ treatment + log_income | group_id1 + group_id2,
+  feols_fit <- fixest::feols(proposition_vote ~ treatment + log_income | group_id1  + Q2_defense,
                      data = data1
   )
   
   
   for(bootstrap_type in c("11", "31")){
     
-    # set.seed(123); dqrng::dqset.seed(123)
-    # boot <- boottest(feols_fit,
-    #                    B = 9999,
-    #                    param = "treatment",
-    #                    clustid = "group_id1",
-    #                    bootstrap_type = bootstrap_type,
-    #                    ssc = boot_ssc(adj = FALSE, cluster.adj = FALSE)
-    # )
-    # 
-    # set.seed(123); dqrng::dqset.seed(123)
-    # boot_fe <- boottest(feols_fit,
-    #                  B = 9999,
-    #                  param = "treatment",
-    #                  clustid = "group_id1",
-    #                  bootstrap_type = bootstrap_type,
-    #                  ssc = boot_ssc(adj = FALSE, cluster.adj = FALSE), 
-    #                  fe = "group_id1"
-    # )
-    # 
-    # expect_equal(pval(boot), pval(boot_fe))
-    # expect_equal(teststat(boot), teststat(boot_fe))
-    # expect_equal(confint(boot), confint(boot_fe))
-    
-    
-    expect_error(
-      boot_fe <- boottest(feols_fit,
-                          B = 9999,
-                          param = "treatment",
-                          clustid = "group_id1",
-                          bootstrap_type = bootstrap_type,
-                          ssc = boot_ssc(adj = FALSE, cluster.adj = FALSE), 
-                          fe = "group_id1"
-      )
+    set.seed(123); dqrng::dqset.seed(123)
+    suppressWarnings(
+        boot <- boottest(feols_fit,
+                           B = 9999,
+                           param = "treatment",
+                           clustid = "group_id1",
+                           bootstrap_type = bootstrap_type,
+                           ssc = boot_ssc(adj = FALSE, cluster.adj = FALSE)
+        )
     )
-    
-    expect_error(
-      boot_fe <- boottest(feols_fit,
-                          B = 9999,
-                          param = "treatment",
-                          clustid = "group_id1",
-                          bootstrap_type = bootstrap_type,
-                          ssc = boot_ssc(adj = FALSE, cluster.adj = FALSE), 
-                          fe = "group_id1"
-      )
+    set.seed(123); dqrng::dqset.seed(123)
+    boot_fe <- boottest(feols_fit,
+                     B = 9999,
+                     param = "treatment",
+                     clustid = "group_id1",
+                     bootstrap_type = bootstrap_type,
+                     ssc = boot_ssc(adj = FALSE, cluster.adj = FALSE),
+                     fe = "group_id1"
     )
+    # set.seed(123); dqrng::dqset.seed(123)
+    # boot_jl <- boottest(feols_fit,
+    #                     B = 9999,
[…]
