+++ NAMESPACE
@@ -32,7 +32,6 @@ export(teststat)
 export(tidy)
 importFrom(JuliaConnectoR,juliaEval)
 importFrom(JuliaConnectoR,juliaImport)
-importFrom(MASS,ginv)
 importFrom(Matrix,Diagonal)
 importFrom(Matrix,Matrix)
 importFrom(Matrix,crossprod)
