+++ vignettes/articles/bibliography.bib
@@ -538,3 +538,13 @@ with Diagnostics},
     url = {https://github.com/s3alfisc/wildwyoung},
   }
 
+@article{djogbenou2019asymptotic,
+  title={Asymptotic theory and wild bootstrap inference with clustered errors},
+  author={Djogbenou, Antoine A and MacKinnon, James G and Nielsen, Morten {\O}rregaard},
+  journal={Journal of Econometrics},
+  volume={212},
+  number={2},
+  pages={393--412},
+  year={2019},
+  publisher={Elsevier}
+}
