+++ R/onLoad.R
@@ -19,5 +19,5 @@ function(libname, pkgname) {
   packageStartupMessage("\nPlease cite as: \n")
   packageStartupMessage(" Fischer & Roodman. (2021). fwildclusterboot: Fast Wild Cluster.")
   packageStartupMessage(" Bootstrap Inference for Linear Regression Models.")
-  packageStartupMessage(" Available from https://cran.r-project.org/package=fwildclusterboot\.")
+  packageStartupMessage(" Available from https://cran.r-project.org/package=fwildclusterboot/.")
 }
