+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   WildBootTests: WildBootTests.html
   fwildclusterboot: fwildclusterboot.html
   Literature: Literature.html
-last_built: 2023-04-18T08:30Z
+last_built: 2023-07-07T21:19Z
 urls:
   reference: https://s3alfisc.github.io/fwildclusterboot/reference
   article: https://s3alfisc.github.io/fwildclusterboot/articles
