+++ pkgdown.yml
@@ -6,9 +6,10 @@ articles:
   Multiple-Estimations-with-fixest: Multiple-Estimations-with-fixest.html
   Wild-Cluster-Bootstrap-Inference-in-Stata-Julia-and-Python: Wild-Cluster-Bootstrap-Inference-in-Stata-Julia-and-Python.html
   WildBootTests: WildBootTests.html
+  wild_bootstrap: wild_bootstrap.html
   fwildclusterboot: fwildclusterboot.html
   Literature: Literature.html
-last_built: 2023-07-07T21:19Z
+last_built: 2023-07-08T08:23Z
 urls:
   reference: https://s3alfisc.github.io/fwildclusterboot/reference
   article: https://s3alfisc.github.io/fwildclusterboot/articles
