+++ NEWS.md
@@ -1,3 +1,10 @@
+# fwildclusterboot 0.14.1
+
+- brings back the `print()` method, it had a use case after all
+- some fixes to the documentation
+
+
+
 # fwildclusterboot 0.14
 
 ## Breaking Changes
