+++ pkgdown.yml
@@ -9,7 +9,7 @@ articles:
   wild_bootstrap: wild_bootstrap.html
   fwildclusterboot: fwildclusterboot.html
   Literature: Literature.html
-last_built: 2023-07-08T08:40Z
+last_built: 2023-07-08T10:32Z
 urls:
   reference: https://s3alfisc.github.io/fwildclusterboot/reference
   article: https://s3alfisc.github.io/fwildclusterboot/articles
