+++ R/preprocess2.R
@@ -14,6 +14,7 @@
 preprocess2 <- function(object, ...) {
   UseMethod("preprocess2")
 }
+
 preprocess2.fixest <-
   function(object,
            clustid,
@@ -43,6 +44,7 @@ preprocess2.fixest <-
     #' @noRd
     #'
     #' @method preprocess2 fixest
+    
     call <- object$call
     call_env <- object$call_env
     fml <- formula(object)
@@ -242,6 +244,8 @@ preprocess2.felm <-
     #' @noRd
     #'
     #' @method preprocess2 felm
+    #' 
+  
     call <- object$call
     call_env <- environment(formula(object))
     fml <- formula(object)
@@ -412,6 +416,7 @@ preprocess2.lm <-
     #' @noRd
     #'
     #' @method preprocess2 lm
+    
     call <- object$call
     call_env <- environment(formula(object))
     fml <- formula(object)
@@ -819,54 +824,12 @@ get_cluster <-
     # doi: 10.18637/jss.v095.i01 (URL: https://doi.org/10.18637/jss.v095.i01).
     # changes by Alexander Fischer:
     # no essential changes, but slight reorganization of pieces of code
+    
     dreamerr::check_arg(clustid_char, "character scalar|charakter vector")
     dreamerr::check_arg(bootcluster, "character scalar | character vector")
     
     clustid_fml <- reformulate(clustid_char)
-    # Step 1: create cluster df
     
-    
-    manipulate_object <- function(object){
-      if(inherits(object, "fixest")){
-        if(!is.null(object$fixef_vars)){
-          update(object, . ~ + 1 | . + 1)
-        } else {
-          update(object, . ~ + 1 )
-        }
-      } else {
-        object
-      }
-    }
-    
-    cluster_tmp <-
-      if ("Formula" %in% loadedNamespaces()) {
-        ## FIXME to suppress potential warnings due to | in Formula
-        suppressWarnings(
-          expand.model.frame(
-            model =
-              manipulate_object(object),
-            extras = clustid_fml,
-            na.expand = FALSE,
-            envir = call_env
-          )
-        )
-      } else {
-        expand.model.frame(
-          model =
-            manipulate_object(object),
-          extras = clustid_fml,
-          na.expand = FALSE,
-          envir = call_env
-        )
-      }
-    
-    cluster_df <-
-      model.frame(clustid_fml, cluster_tmp, na.action = na.pass)
-    # without cluster intersection
-    N_G <-
-      vapply(cluster_df, function(x) {
-        length(unique(x))
-      }, numeric(1))
     # Step 1: decode bootcluster variable
     # create a bootcluster vector
     if (length(bootcluster) == 1) {
@@ -883,22 +846,103 @@ get_cluster <-
       bootcluster_char <- bootcluster
     }
     
-    # add bootcluster variable to formula of clusters
-    cluster_bootcluster_fml <-
-      update(
-        clustid_fml, paste(
-          "~ . +", paste(
-            bootcluster_char,
-            collapse = " + "
+    
+
+    if(inherits(object, "fixest")){
+      
+      data <- fetch_data(object)
+      # Check that cluster vars are in the original estimation dataset
+      cluster_vars_in_data <- clustid_char %in% colnames(data)
+      if (any(!(cluster_vars_in_data))) {
+        stop(paste0(
+          "The following variables are not found in the dataset used in your `feols` call: ",
+          paste(clustid_char[!cluster_vars_in_data], collapse = ", ")
+        ))
+      }
+      
+      clustid <- data[,clustid_char, drop = FALSE]
+      if(N != nrow(clustid)){
+        clustid <- clustid[unlist(object$obs_selection), drop = FALSE]
+      }
+      
+      bootcluster <- data[, bootcluster_char, drop = FALSE]
[…]
