### 2023-07-08 a6553e1
comment out varying slopes check
