+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: fwildclusterboot
 Title: Fast Wild Cluster Bootstrap Inference for Linear Models
-Version: 0.14.1
+Version: 0.15
 Authors@R: c(
     person("Alexander", "Fischer", , "alexander-fischer1801@t-online.de", role = c("aut", "cre")),
     person("David", "Roodman", role = "aut"),
@@ -75,6 +75,8 @@ Suggests:
     testthat (>= 3.0.0),
     tibble, 
     MASS
+Remotes: 
+    kylebutts/fixest/tree/sparse-matrix
 LinkingTo:
     Rcpp,
     RcppArmadillo,
