+++ NAMESPACE
@@ -54,6 +54,7 @@ importFrom(dreamerr,check_value_plus)
 importFrom(dreamerr,set_up)
 importFrom(dreamerr,validate_dots)
 importFrom(dreamerr,warn_up)
+importFrom(fixest,sparse_model_matrix)
 importFrom(generics,glance)
 importFrom(generics,tidy)
 importFrom(gtools,permutations)
