### 2023-07-08 313211a
add varying slopes check
