### 2023-07-09 ebbebf5
preprocess2 method to functions
