+++ R/boottest_ivreg.R
@@ -288,7 +288,7 @@ boottest.ivreg <- function(object,
 
 
   # preprocess data: X, Y, weights, fixed effects
-  preprocess <- preprocess2.ivreg(
+  preprocess <- preprocess_ivreg(
     object = object,
     clustid = clustid,
     R = R,
