+++ R/mboottest_lm.R
@@ -226,7 +226,7 @@ mboottest.lm <- function(object,
 
 
   # preprocess data: X, Y, weights, fixed effects
-  preprocess <- preprocess2.lm(
+  preprocess <- preprocess_lm(
     object = object,
     clustid = clustid,
     R = R,
