### 2023-07-09 5ced198
add cran notes
