+++ R/boottest_felm.R
@@ -200,14 +200,6 @@
 #' @section Multiple Fixed Effects:
 #' If your felm() model contains fixed effects, boottest() will internally convert all fixed
 #' effects but the one specified via the `fe` argument to dummy variables.
-#' @section Run `boottest` quietly:
-#' You can suppress all warning and error messages by setting the following global
-#' options:
-#' `options(rlib_warning_verbosity = "quiet")`
-#' `options(rlib_message_verbosity = "quiet")`
-#' Not that this will turn off all warnings (messages) produced via `rlang::warn()` and
-#' `rlang::inform()`, which might not be desirable if you use other software build on
-#' `rlang`, as e.g. the `tidyverse`.
 #' @section Stata, Julia and Python Implementations:
 #' The fast wild cluster bootstrap algorithms are further implemented in the
 #' following software packages:
@@ -390,47 +382,21 @@ boottest.felm <- function(object,
 
   check_arg(sampling, "charin(dqrng, standard)")
 
-  # remind packages users to set a global seed
-  inform_seed(
-    frequency_id = "seed-reminder-boot-felm",
-    engine = engine
-  )
 
   if(bootstrap_type %in% c("33", "13")){
-    rlang::abort(
-      c(
-        "The bootstrap types '13' and '33' are not yet supported for objects
-        of type felm. Support will be added in the future. Until then, you
-        can run the '13' and '33' bootstrap types with objects of type 'lm'
-        and 'fixest'."
-      ),
-      use_cli_format = TRUE
-    )
+    no_13_33_for_felm_error()
   }
 
   if(bootstrap_type != "fnw11"){
     if(engine == "R"){
       if(conf_int){
-        rlang::inform(
-        "Confidence Intervals are currently only supported for
-                the R engine with 'bootstrap_type = 'fnw11' '.",
-        use_cli_format = TRUE,
-        .frequency = "regularly",
-        .frequency_id = "CI only for fnw algo."
-        )
+        cis_only_for_fnw11()
       }
     }
   }
 
   if (!is.null(beta0)) {
-    rlang::abort(
-      c(
-        "The function argument 'beta0' is deprecated.
-         Please use the function argument 'r' instead,
-         by which it is replaced."
-      ),
-      use_cli_format = TRUE
-    )
+    arg_beta0_is_deprecated_error()
   }
 
   if (inherits(clustid, "formula")) {
