+++ R/invert_p_val.R
@@ -190,16 +190,7 @@ invert_p_val <- function(ABCD,
   #'   I have plans to revisit this going forward.
 
   if (sum(p_start < sign_level) < 2) {
-    rlang::abort(c(
-      "The inflation factor for initial guesses for standard errors was not
-      large enough. In consequence, the root-finding procedure to compute
-      confidence intervals via p-value inversion could not be initiated.
-      In a future release, it will be possible to specify a costum inflation
-      factor as a function argument to boottest(). Until then, you can still
-      use boottest() to calculate p-values by setting the boottest() function
-      argument conf_int to FALSE."),
-      use_cli_format = TRUE
-    )
+    ci_root_finding_initiation_failed()
   }
 
   # ----------------------------------------------------------------------

[…]
