### 2023-07-11 0834f04
fix test bugs; allow tests on CRAN as Julia never called
