### 2023-07-11 375a542
preprocess2 method to functions
