+++ R/mboottest_fixest.R
@@ -228,7 +228,7 @@ mboottest.fixest <- function(object,
     fe = fe
   )
 
-  preprocess <- preprocess2_fixest(
+  preprocess <- preprocess_fixest(
     object = object,
     clustid = clustid,
     R = R,
