+++ R/boottest_felm.R
@@ -458,7 +458,7 @@ boottest.felm <- function(object,
     fe = fe
   )
 
-  preprocess <- preprocess2_felm(
+  preprocess <- preprocess_felm(
     object = object,
     clustid = clustid,
     R = R_long,
