+++ R/boottest_fixest.R
@@ -484,7 +484,7 @@ boottest.fixest <- function(object,
   )
 
   # preprocess the data: Y, X, weights, fixed_effect
-  preprocess <- preprocess2_fixest(
+  preprocess <- preprocess_fixest(
     object = object,
     clustid = clustid,
     R = R_long,
