### 2023-09-10 c6d9d94
add se_guess function argument to boottest
