+++ R/boottest_felm.R
@@ -136,6 +136,8 @@
 #' This argument is mostly a convenience to control random number generation in
 #' a wrapper package around `fwildclusterboot`, `wildrwolf`.
 #' I recommend to use the fast' option.
+#' @param se_guess NULL by default. Please use it to update the initial 
+#'        guess for the standard error if the CI inversion throws an error.
 #' @param ... Further arguments passed to or from other methods.
 #'
 #' @importFrom dreamerr check_arg validate_dots
@@ -347,6 +349,7 @@ boottest.felm <- function(object,
                           maxmatsize = FALSE,
                           bootstrapc = FALSE,
                           getauxweights = FALSE,
+                          se_guess = NULL,
                           ...) {
   call <- match.call()
 
@@ -530,7 +533,9 @@ boottest.felm <- function(object,
       getauxweights = getauxweights ,
       maxmatsize = maxmatsize,
       sampling = sampling,
-      bootcluster = bootcluster
+      bootcluster = bootcluster,
+      se_guess = se_guess
+      
     )
 
 
