### 2023-09-13 a75865d
add engine_btype check; switch to R-lean whenever engine = "WildBootTests.jl" but the hetero bootstrap is run (prior, no clear error message) #135
