+++ R/boottest_fixest.R
@@ -431,7 +431,12 @@ boottest.fixest <- function(object,
       }
     }
   }
-
+  
+  check_engine_btype(
+    engine = engine,
+    bootstrap_type = bootstrap_type
+  )
+  
   if (!is.null(beta0)) {
     rlang::abort(
       c(
@@ -481,14 +486,18 @@ boottest.fixest <- function(object,
 
   if (is.null(clustid)) {
     heteroskedastic <- TRUE
-    if (engine == "R") {
+    if (engine %in% c("R", "WildBootTests.jl")) {
       # heteroskedastic models should always be run through R-lean
       engine <- "R-lean"
+      rlang::inform(
+        "The heteroskedastic bootstrap is only supported for `engine = 'R-lean'`, 
+        Resetting the engine argument to 'R-lean'."
+      )
+    } else {
+      heteroskedastic <- FALSE
     }
-  } else {
-    heteroskedastic <- FALSE
   }
-
+  
   check_bootstrap_types(
     param = param,
     bootstrap_type = bootstrap_type
