### 2023-09-13 181bf71
fix bug in print method
