+++ R/boottest_lm.R
@@ -420,9 +420,9 @@ boottest.lm <- function(object,
         "The heteroskedastic bootstrap is only supported for `engine = 'R-lean'`, 
         Resetting the engine argument to 'R-lean'."
       )
-    } else {
-      heteroskedastic <- FALSE
     }
+  } else {
+    heteroskedastic <- FALSE
   }
     
   check_bootstrap_types(
