+++ R/boottest_fixest.R
@@ -493,9 +493,9 @@ boottest.fixest <- function(object,
         "The heteroskedastic bootstrap is only supported for `engine = 'R-lean'`, 
         Resetting the engine argument to 'R-lean'."
       )
-    } else {
-      heteroskedastic <- FALSE
     }
+  } else {
+    heteroskedastic <- FALSE
   }
   
   check_bootstrap_types(
