+++ cran-comments.md
@@ -1,3 +1,39 @@
+## Submission of 0.15
+
+I have tested the package on: 
+- rhub
+- win devel
+- github actions
+
+and received the following notes: 
+
+* checking for non-standard things in the check directory ... NOTE
+Found the following files/directories:
+  ''NULL''
+
+* checking for detritus in the temp directory ... NOTE
+  'lastMiKTeXException'
+Found the following files/directories:
+
+* Found the following (possibly) invalid URLs:
+  URL: https://journals.sagepub.com/doi/pdf/10.1177/1536867X19830877
+    From: inst/doc/Literature.html
+    Status: 403
+    Message: Forbidden
+  URL: https://onlinelibrary.wiley.com/doi/abs/10.1002/jrsm.1554
+    From: inst/doc/Literature.html
+    Status: 403
+    Message: Forbidden
+  URL: https://www.tandfonline.com/doi/abs/10.1198/jbes.2009.07221
+    From: inst/doc/Literature.html
+          inst/doc/fwildclusterboot.html
+          README.md
+    Status: 403
+    Message: Forbidden
+    
+I have checked all links - all are valid.
+
+
 ## Submission of 0.13
 
 I have tested the package on: 
