+++ NAMESPACE
@@ -64,6 +64,7 @@ importFrom(rlang,warn)
 importFrom(stats,aggregate)
 importFrom(stats,as.formula)
 importFrom(stats,coef)
+importFrom(stats,confint)
 importFrom(stats,expand.model.frame)
 importFrom(stats,formula)
 importFrom(stats,model.frame)
