+++ R/boottest_felm.R
@@ -502,7 +502,7 @@ boottest.felm <- function(object,
     fe = fe
   )
 
-  preprocess <- preprocess2.felm(
+  preprocess <- preprocess_felm(
     object = object,
     clustid = clustid,
     R = R_long,
