+++ tests/testthat/test-method_equivalence.R
@@ -179,6 +179,7 @@ test_that("Do different, but equivalent ways to specify
   # one-way clustering
 
   create_models <- function(clustid) {
+    
     R1 <- clubSandwich::constrain_zero(2:3, coefs = coef(lm_fit))
     R2 <- matrix(c(1, 0, 0, 1, 0, 0), 2, 3)
     R3 <- clubSandwich::constrain_zero(1:2, coefs = coef(felm_fit8))
@@ -187,8 +188,8 @@ test_that("Do different, but equivalent ways to specify
     for (engine in c("R", "WildBootTests.jl")) {
       # boottest()
       cat("boottest()", "\n")
-      set.seed(8965)
-      dqrng::dqset.seed(765)
+      set.seed(89651)
+      dqrng::dqset.seed(74465)
 
       assign(paste0("boot_lm_", engine),
         suppressWarnings(
@@ -206,8 +207,8 @@ test_that("Do different, but equivalent ways to specify
         envir = .GlobalEnv
       )
 
-      set.seed(8965)
-      dqrng::dqset.seed(765)
+      set.seed(89651)
+      dqrng::dqset.seed(74465)
 
       assign(paste0("boot_fixest1_", engine),
         suppressWarnings(
@@ -225,8 +226,8 @@ test_that("Do different, but equivalent ways to specify
         envir = .GlobalEnv
       )
 
-      set.seed(8965)
-      dqrng::dqset.seed(765)
+      set.seed(89651)
+      dqrng::dqset.seed(74465)
 
       assign(paste0("boot_fixest2_", engine),
         suppressWarnings(
@@ -244,8 +245,8 @@ test_that("Do different, but equivalent ways to specify
         envir = .GlobalEnv
       )
 
-      set.seed(8965)
-      dqrng::dqset.seed(765)
+      set.seed(89651)
+      dqrng::dqset.seed(74465)
 
       assign(paste0("boot_fixest3_", engine),
         suppressWarnings(
@@ -263,8 +264,8 @@ test_that("Do different, but equivalent ways to specify
         envir = .GlobalEnv
       )
 
-      set.seed(8965)
-      dqrng::dqset.seed(765)
+      set.seed(89651)
+      dqrng::dqset.seed(74465)
 
       assign(paste0("boot_fixest4_", engine),
         suppressWarnings(
@@ -282,8 +283,8 @@ test_that("Do different, but equivalent ways to specify
         envir = .GlobalEnv
       )
 
-      set.seed(8965)
-      dqrng::dqset.seed(765)
+      set.seed(89651)
+      dqrng::dqset.seed(74465)
 
       assign(paste0("boot_fixest5_", engine),
         suppressWarnings(
@@ -301,8 +302,8 @@ test_that("Do different, but equivalent ways to specify
         envir = .GlobalEnv
       )
 
-      set.seed(8965)
-      dqrng::dqset.seed(765)
+      set.seed(89651)
+      dqrng::dqset.seed(74465)
 
       assign(paste0("boot_fixest6_", engine),
         suppressWarnings(
@@ -320,8 +321,8 @@ test_that("Do different, but equivalent ways to specify
         envir = .GlobalEnv
       )
 
-      set.seed(8965)
-      dqrng::dqset.seed(765)
+      set.seed(89651)
+      dqrng::dqset.seed(74465)
 
       assign(paste0("boot_fixest7_", engine),
         suppressWarnings(
@@ -339,8 +340,8 @@ test_that("Do different, but equivalent ways to specify
         envir = .GlobalEnv
       )
 
-      set.seed(8965)
-      dqrng::dqset.seed(765)
+      set.seed(89651)
+      dqrng::dqset.seed(74465)
 
[…]
