+++ R/preprocess_models.R
@@ -1,20 +1,4 @@
-#' `preprocess2` is a S3 method that fetches data from several model
-#' objectects for use with `boottest()`.
-#'
-#' @param object An objectect of type lm, fixest, felm or ivreg
-#' @param ... other arguments
-#' @noRd
-#'
-#' @importFrom Matrix Diagonal
-#' @return An object of class `preprocess2`.
-#' @srrstats {G2.8} *Software should provide appropriate conversion or dispatch
-#'  routines as part of initial pre-processing to ensure that all other
-#'  sub-functions of a package receive inputs of a single defined class or
-#'  type.* Similar preprocess method + run_bootstrap function.
-preprocess2 <- function(object, ...) {
-  UseMethod("preprocess2")
-}
-preprocess2.fixest <-
+preprocess_fixest <-
   function(object,
            clustid,
            R,
@@ -213,7 +197,7 @@ preprocess2.fixest <-
     }
     res
   }
-preprocess2.felm <-
+preprocess_felm <-
   function(object,
            clustid,
            R,
@@ -388,7 +372,7 @@ preprocess2.felm <-
     }
     res
   }
-preprocess2.lm <-
+preprocess_lm <-
   function(object,
            clustid,
            R,
@@ -492,7 +476,7 @@ preprocess2.lm <-
     class(res) <- c("preprocess", "ols")
     res
   }
-preprocess2.ivreg <-
+preprocess_ivreg <-
   function(object,
            clustid,
            R,
