+++ R/mboottest_felm.R
@@ -229,7 +229,7 @@ mboottest.felm <- function(object,
 
   fedfadj <- 0L
 
-  preprocess <- preprocess2.felm(
+  preprocess <- preprocess_felm(
     object = object,
     clustid = clustid,
     R = R,
