+++ man/boottest.Rd
@@ -22,7 +22,7 @@ the fast wild bootstrap algorithm developed in Roodman et al., 2019.
 \details{
 Technical Details
 For technical details, either take a look at the references below, or check
-out the \href{wild_bootstrap.html}{wild (cluster) bootstrap vignette}.
+out the \href{https://s3alfisc.github.io/fwildclusterboot/articles/wild_bootstrap.html}{wild (cluster) bootstrap vignette}.
 }
 \section{Setting Seeds}{
 
