+++ vignettes/Literature.Rmd
@@ -33,7 +33,7 @@ Methods for Cluster-Robust Inference (Journal of Applied Econometrics), 2023](ht
 
 + [Davidson & MacKinnon - Wild bootstrap tests for IV regression (Journal of Economic and Business Statistics), 2010](https://www.tandfonline.com/doi/abs/10.1198/jbes.2009.07221) Introduces the WRE bootstrap / wild cluster bootstrap for instrumental variables regression.
 
-+ [Cameron, Gelbach & Miller - Bootstrap-based improvements for inference with clustered errors (Review of Economics and Statistics)](https://www.nber.org/system/files/working_papers/t0344/t0344.pdf) The paper that started the literature on wild cluster bootstrap inference. Simulation evidence that the wild cluster bootstrap works remarkably well when there are only few clusters. 
++ [Cameron, Gelbach & Miller - Bootstrap-based improvements for inference with clustered errors (Review of Economics and Statistics)](https://www.econstor.eu/bitstream/10419/31404/1/514886757.pdf) The paper that started the literature on wild cluster bootstrap inference. Simulation evidence that the wild cluster bootstrap works remarkably well when there are only few clusters. 
 
 + [Flachaire - Bootstrapping heteroskedastic regression models: wild bootstrap vs. pairs bootstrap (Computational Statistics & Data Analysis), 2005](https://shs.hal.science/file/index/docid/175910/filename/Flachaire_03a.pdf) Provides simulation evidence of the performance of *heteroskedastic* wild bootstrap procedures. 
 
@@ -41,7 +41,7 @@ Methods for Cluster-Robust Inference (Journal of Applied Econometrics), 2023](ht
 
 + [Joshi, Megha, James E. Pustejovsky, and S. Natasha Beretvas - "Cluster wild bootstrapping to handle dependent effect sizes in meta‐analysis with a small number of studies." (Research Synthesis Methods), 2022](https://onlinelibrary.wiley.com/doi/abs/10.1002/jrsm.1554) Nice simulations on the empirical performance of the wild cluster bootstrap for tests of multiple joint hypotheses. 
  
-+ [Kline & Santos - A Score Based Approach to Wild Bootstrap Inference (Journal of Econometric Methods), 2012](https://www.nber.org/system/files/working_papers/w16127/w16127.pdf) Introduces a score based wild bootstrap for non-linear regression models. 
++ [Kline & Santos - A Score Based Approach to Wild Bootstrap Inference (Journal of Econometric Methods), 2012](https://eml.berkeley.edu/~pkline/papers/ScoreFinal_web.pdf) Introduces a score based wild bootstrap for non-linear regression models. 
 
 ## Links, blogposts, etc
 
