### 2023-09-16 47b7f6a
fixes for CRAN release
