+++ tests/testthat/test-error_warning.R
@@ -1,3 +1,6 @@
+fwildclusterboot::setBoottest_nthreads(1)
+
+
 test_that("errors and warnings q = 1", {
 
 
