+++ tests/testthat/test-method_equivalence.R
@@ -1,3 +1,6 @@
+fwildclusterboot::setBoottest_nthreads(1)
+
+
 test_that("Do different, but equivalent ways to specify
           linear models lead to equivalent results?", {
   skip_on_cran()
