+++ tests/testthat/test-convergence.R
@@ -1,3 +1,6 @@
+fwildclusterboot::setBoottest_nthreads(1)
+
+
 test_that("algorithm performance test", {
 
 
