### 2023-09-17 8203ef1
document setNthreads arg
