+++ R/utils.R
@@ -263,7 +263,6 @@ setBoottest_nthreads <- function(nthreads) {
   #' By default, only one thread is used
   #' @param nthreads Integer. Number of threads to be used
   #' @return No return value
-  #' @noRd
   #' @importFrom parallel detectCores
   #' @export
 
@@ -398,16 +397,9 @@ to_integer <- function(vec) {
 
 # functions taken from 'clubSandwich' package
 matrix_split <- function (x, fac, dim) {
-  # if (is.vector(x)) {
-  #   if (dim != "both")
-  #     stop(paste0("Object must be a matrix in order to subset by ",
-  #                 dim, "."))
-  #   x_list <- split(x, fac)
-  #   lapply(x_list, function(x) diag(x, nrow = length(x)))
-  # }
-  # else {
-    lapply(levels(fac), sub_f(x, fac, dim))
-  #}
+  
+  lapply(levels(fac), sub_f(x, fac, dim))
+  
 }
 
 sub_f <- function (x, fac, dim){
