### 2023-09-17 41a2032
no absolute value for inversion starting values
