### 2023-12-08 #546 @jhollist
@s3alfisc I am currently serving as the rOpenSci EIC and just checking in on things.

How's it coming on the updates to your package from the review process?
