### 2023-12-15 #546 @s3alfisc
Hi @jhollist , thanks for checking up on this.

The honest status update is that I have, to some degree, given up on adressing **all** of the comments I received out of the review process. The main hurdle are the inline standards documentation via roxygen tags. To some degree, I feel that many of them do not really apply to the package, as `fwildclusterboot` is not a package for regression fitting but for regression **inference**. And then with the tags that do apply, it seems like an incredible amount of work to put them in every single spot of the package where they apply. I've started this multiple times but have repeatedly given up.

On the other hand, I did incorporate many of the other points I got feedback on:

- I have significantly revamped the vignettes. Most importantly, I added a dedicated [vignette that explains all concepts around the wild cluster bootstrap](https://s3alfisc.github.io/fwildclusterboot/articles/wild_bootstrap.html). I have also added a reference section with short descriptions of key paper from the [literature on the wild cluster bootstrap](https://s3alfisc.github.io/fwildclusterboot/articles/Literature.html). This actually took a lot of time, but I think it was worth the effort.
- In the vignette, I also improved the explanation of what the plot method does, though I have noticed I still need to update the documentation of `plot.boottest()`.
- I have added a details section to [the documentation of `boottest()`](https://s3alfisc.github.io/fwildclusterboot/reference/boottest.html) that links to the detailed vignettes.
- Since the review, I have incorporated a set of new estimators developed in work by [MacKinnon, Nielsen & Webb (Journal of Applied Econometrics, 2023)](https://arxiv.org/pdf/2301.04527.pdf). This also took a good amount of time.

I'm not really sure what the best way forward will be. I have quite a bit of time off until the New Year, so chances are I pull myself together and make sure all the tags can be found in the codebase. To put some pressure on me - how about we keep this review open until Jan 2, and if I haven't made sufficient progress until then, we close the review?

Best, Alex
