+++ R/preprocess.R
@@ -1,3 +1,12 @@
+#' @srrstatsNA {RE1.3} *Regression Software which passes or otherwise
+#' transforms aspects of input data onto output structures should ensure
+#' that those output structures retain all relevant aspects of input data,
+#' notably including row and column names, and potentially information from
+#' other `attributes()`.*
+#' The `preprocess_` set of functions ensures that all relevant information
+#' from the supported model classes is retained and appropriately transformed.
+
+
 preprocess_fixest <-
   function(object,
            clustid,
@@ -25,7 +34,7 @@ preprocess_fixest <-
     #' by MacKinnon, Nielsen & Webb (2022)
     #'
     #' @noRd
-    
+
     call <- object$call
     call_env <- object$call_env
     fml <- formula(object)
@@ -39,7 +48,7 @@ preprocess_fixest <-
     if (object$method != "feols") {
       only_ols_for_fixest_error()
     }
-    
+
     is_iv <- ifelse(!is.null(object$fml_all$iv), TRUE, FALSE)
     has_fe <- ifelse(!is.null(object$fml_all$fixef), TRUE, FALSE)
     if (!is_iv) {
@@ -207,7 +216,7 @@ preprocess_felm <-
     #' see the paper by MacKinnon, Nielsen & Webb (2022)
     #'
     #' @noRd
-    
+
     call <- object$call
     call_env <- environment(formula(object))
     fml <- formula(object)
@@ -367,7 +376,7 @@ preprocess_lm <-
     #' @importFrom Matrix sparse.model.matrix
     #'
     #' @noRd
-    
+
     call <- object$call
     call_env <- environment(formula(object))
     fml <- formula(object)
@@ -469,7 +478,7 @@ preprocess_ivreg <-
     #' the bootcluster variables. Alternatively, "min" or "max"
     #'
     #' @noRd
-    
+
     call <- object$call
     call_env <- environment(formula(object))
     fml <- formula(object)
@@ -670,7 +679,7 @@ transform_fe <-
       }
       # project out fe
       if (engine == "R") {
-        
+
         if(bootstrap_type != "fnw11"){
           if(clustid_char != fe){
             no_fixef_for_fast_reliable_error()
@@ -687,7 +696,7 @@ transform_fe <-
         W <- prep_fe$W
         n_fe <- prep_fe$n_fe
       }
-      
+
     } else {
       add_fe <- all_fe
       add_fe_names <- names(add_fe)
@@ -695,17 +704,17 @@ transform_fe <-
       if(engine == "R" && bootstrap_type %in% c("11", "31", "13","33")){
         add_fe_dummies <-
           Matrix::sparse.model.matrix(fml_fe, model.frame(fml_fe, data = as.data.frame(add_fe)))
-        
+
       } else {
-        
+
         add_fe_dummies <-
           model.matrix(fml_fe, model.frame(fml_fe, data = as.data.frame(add_fe)))
       }
-      
+
       X <- cbind(X, add_fe_dummies)
-      
+
     }
-    
+
     res <- list(
       X = X,
       Y = Y,
@@ -768,11 +777,11 @@ get_cluster <-
     # no essential changes, but slight reorganization of pieces of code
     dreamerr::check_arg(clustid_char, "character scalar|charakter vector")
     dreamerr::check_arg(bootcluster, "character scalar | character vector")
-    
+
     clustid_fml <- reformulate(clustid_char)
     # Step 1: create cluster df
-    
-    
+
+
     manipulate_object <- function(object){
       if(inherits(object, "fixest")){
         if(!is.null(object$fixef_vars)){
@@ -784,7 +793,7 @@ get_cluster <-
         object
       }
     }
-    
+
[…]
