+++ tests/testthat/test-seed.R
@@ -13,6 +13,10 @@ test_that("seed works for OLS", {
   #' [NIST Standard Reference Datasets](https://www.itl.nist.gov/div898/strd/),
   #'  or data sets provided by other widely-used R packages).* All tests are based
   #'  on random data sets.
+  #'
+  #' @srrstatsTODO {G5.1} *Data sets created within, and used to test, a package should be exported
+  #'  (or otherwise made generally available) so that users can confirm tests and run examples.*
+
 
   data1 <<-
     fwildclusterboot:::create_data(
