+++ R/srr-stats-standards.R
@@ -14,31 +14,32 @@
 #' time.
 #'
 #' @srrstats {G1.3} *All statistical terminology should be clarified and
-#'  unambiguously defined.* I have prepared a small paper that attempts to
-#'  clarify all statistical terms, which I am happy to send to reviewers.
+#'  unambiguously defined.*  I have added extensive sets of vignettes,
+#' e.g. in this [vignette](https://s3alfisc.github.io/fwildclusterboot/articles/wild_bootstrap.html).
 #'
 #' @srrstats {G1.4} *`roxygen2` is used throughout the package to document
-#' all functions.*
+#' all functions.* You can open any package or check the docs to verify this.
 #'
 #' @srrstats {G1.4a} *All internal (non-exported) functions should also be
 #' documented in standard [`roxygen2`](https://roxygen2.r-lib.org/) format,
 #' along with a final `@noRd` tag to suppress automatic generation of `.Rd`
-#' files.* Done.
+#' files.* Everything is documented or has a noRd tag.
 #'
 #' @srrstats {G1.6} *Software should include code necessary to compare
 #' performance claims with alternative implementations in other R packages.*
-#' Is a link to this blog post sufficient?
+#' Here is a link to a blog post that demonstrates that the package is fast.
+#' But I don't think the package makes any performance claims against other
+#' packages, just mentions that it is fast =).
 #' ://s3alfisc.github.io/blog/post/1000x-faster-wild-cluster-bootstrap
 #' -inference-in-r-with-fwildclusterboot/
 #'
 #' @srrstats {G2.0a} Provide explicit secondary documentation of any
-#' expectations on lengths of inputs. I think this is sufficiently documented.
-#'
-
+#' expectations on lengths of inputs. - I am not sure where to best place
+#' this tag in my codebase.
 #'
 #' @srrstats {G2.6} *Software which accepts one-dimensional input should ensure
-#'  values are appropriately pre-processed regardless of class structures.* (?)
-#'   different methods are checked via 'check_methods_equivalence.R'
+#'  values are appropriately pre-processed regardless of class structures.* -
+#'  I don't really understand this, sorry =D
 #'
 #' @srrstats {G2.7} *Software should accept as input as many of the above
 #' standard tabular forms as possible, including extension to domain-specific
@@ -65,37 +66,44 @@
 #'   Does not happen. NA, as no tabular input data.
 #'
 #'
-#' @srrstats {G2.14} *see G2.13*
+#' @srrstats {G2.14} *Where possible, all functions should provide options for users to specify how to handle missing (`NA`) data, with options minimally including:*
+#' This is not applicable. The data is pre-processed in the regression class, there is no need to do so again for `boottest()`.
 #'
-#' @srrstats {G2.14a} *see G2.13*
+#'@srrstatsTODO {G2.14a} *error on missing data*. *see G2.13*
 #'
 #' @srrstats {G2.14b} *see G2.13*
 #'
 #' @srrstats {G2.14c} *see G2.13*
 #'
 #' @srrstats {G3.0} *No floating point numbers (rational numbers) are
-#'  compared with equality.* Yes, checked.
+#'  compared with equality.* Yes, checked. Though I don't know where to
+#' document this in the codebase.
 #'
 #' @srrstats {G3.1} *Statistical software which relies on covariance
 #' calculations should enable users to choose between different algorithms
 #' for calculating covariances, and should not rely solely on covariances
 #' from the `stats::cov` function.* The package deals with "clustered"
-#' standard errors, but does not produce covariance matrices.
-#'  Non-clustered tests based on the 'regular' wild bootstrap
-#'  are also supported.
+#' standard errors, but does not produce covariance matrices. Nevertheless,
[…]

[…]
