### 2024-01-13 b152e68
some updates on ssr standards and reformat via styler
