### 2024-01-13 6961e31
prepare 0.15.0, ropensci srr's, move to drqng::dqrrademacher
