+++ R/boot_algo_fastnwild.R
@@ -65,9 +65,9 @@ boot_algo_fastnwild <-
     #' @param sampling 'dqrng' or 'standard'. If 'dqrng', the 'dqrng' package is
     #' used for random number generation (when available). If 'standard', 
     #' functions from the 'stats' package are used when available. 
-    #' This argument is mostly a convenience to control random number generation in 
-    #' a wrapper package around `fwildclusterboot`, `wildrwolf`. 
-    #' I recommend to use the fast' option. 
+    #' This argument is mostly a convenience to control random 
+    #' number generation in a wrapper package around `fwildclusterboot`, 
+    #' `wildrwolf`. I recommend to use the fast' option. 
     #' @return A list of ...
     #' @importFrom collapse fsum GRP
     #' @importFrom stats as.formula coef model.matrix model.response
