### 2024-03-02 #546 @s3alfisc
Hi, reposting for visibility a message I sent to @mpadge on the topic:

> In my memory, the main thing that I still needed to do for the `fwildclusterboot` review was to include the srr tags into the fwildclusterboot codebase. For every tag that was relevant, I placed it in at least one point in the codebase - I decided not to place the same tags in each boottest() method, e.g. only `boottest_fixest` is properly tagged whereas `boottest_lm` and `boottest_lfe` are not.
> For quite a few tags, I could not find the "one" spot in the codebase, so I left them in the srr file with a comment. For example for a tag as "never compare ints with bools", I could not find a proper spot. I just don't think I do so! Otherwise, many tags are just not relevant - as fwildclusterboot is a post regression method, NA values need to be handled by the regression package which fwildclusterboot supports; or all tags relating predictions.
> Beyond my main takeaway from the second review was that I needed to fix the standards and improve the documentation of statistical terms. https://github.com/ropensci/software-review/issues/546#issuecomment-1566285672 For this, I have significantly revised the vignettes at some point, added details sections to the docs, etc. There are other things that I am not perfectionist enough to fix, i.e. a consistent citation format for all references, and then things that I simply cannot fix.
