### 2024-03-04 #546 @helske
@mpadge please correct me if I'm wrong, but my impression on the standard tags is that it is fine to have some general tags relevant for the whole package for example in the main package documentation, e.g., in `fwildclusterboot-package.R` which could contain some overall info on the package accessible by `?fwildclusterboot` (similar text as in the `DESCRIPTION`). For non-applicable standards, you could also include these in the above file as well, with a short description on why they do not apply, which I see you have already done in the `srr-stats-standards.R`.

Other than than the placement of the standards, if you feel that you have adequately reacted to all points by the reviewers, then we can ask them to take a look and see if they feel ready to approve the package.
