### 2024-03-04 #546 @mpadge
@ldecicco-USGS and @helske I've had a look through the code, and from my side at least would now be happy with the way that the standards have been documented in-line within the code. I've opened a short TODO list for @s3alfisc at https://github.com/s3alfisc/fwildclusterboot/issues/149. Alex, please report back once you've addressed those, and all should be good to proceed. Thanks :+1:
