### 2024-03-04 s3alfisc/fwildclusterboot#149 @s3alfisc
Awesome, thank you @mpadge! I'll start working on this tomorrow and am very positive that I'll be able to address all your points by the end of the weekend!
