### 2024-03-06 #546 @s3alfisc
I'll tackle all of your suggestions this weekend. Thanks for the feedback @mpadge! I'll report here once I think I am done =)
