### 2024-03-10 b80f9c7
move all non-applicable standards to NA section
