+++ R/srr-stats-standards.R
@@ -39,40 +39,6 @@
 #' expectations on lengths of inputs. - I am not sure where to best place
 #' this tag in my codebase.
 #'
-#' @srrstats {G2.6} *Software which accepts one-dimensional input should ensure
-#'  values are appropriately pre-processed regardless of class structures.* -
-#'  I don't really understand this, sorry =D
-#'
-#' @srrstats {G2.7} *Software should accept as input as many of the above
-#' standard tabular forms as possible, including extension to domain-specific
-#' forms.* No tabular data is passed to boottest().
-#'
-#' @srrstats {G2.10} *Software should ensure that extraction or filtering of
-#'  single columns from tabular inputs should not presume any particular
-#'  default behaviour, and should ensure all column-extraction operations
-#'  behave consistently regardless of the class of tabular data used as input.*
-#'  NA, as no tabular data inputs.
-#'
-#' @srrstats {G2.11} *Software should ensure that `data.frame`-like tabular
-#'  objects which have columns which do not themselves have standard class
-#'   attributes (typically, `vector`) are appropriately processed, and do
-#'   not error without reason. This behaviour should be tested. Again, columns
-#'    created by the [`units` package](https://github.com/r-quantities/units/)
-#'    provide a good test case.* NA, as no tabular input data.
-#'
-#' @srrstats {G2.12} *Software should ensure that `data.frame`-like tabular
-#'  objects which have list columns should ensure that those columns are
-#'  appropriately pre-processed either through being removed, converted to
-#'  equivalent vector columns where appropriate, or some other appropriate
-#'  treatment such as an informative error. This behaviour should be tested.*
-#'   Does not happen. NA, as no tabular input data.
-#'
-#'
-#' @srrstats {G2.14} *Where possible, all functions should provide options 
-#' for users to specify how to handle missing (`NA`) data, with options 
-#' minimally including:* This is not applicable. The data is pre-processed 
-#' in the regression class, there is no need to do so again for `boottest()`.
-#'
 #' @srrstatsTODO {G2.14a} *error on missing data*. *see G2.13*
 #'
 #' @srrstats {G2.14b} *see G2.13*
@@ -113,19 +79,6 @@
 #' with data with extreme properties including but not limited to:* Tests for
 #'  non-randomness under full enumeration:
 #'
-#' @srrstats {RE1.1} *Regression Software should document how formula interfaces
-#'  are converted to matrix representations of input data.* Not applicable. 
-#'  Formulas only create one dimensional columns (for cluster variables or
-#'  fixed effect).
-#'
-#' @srrstats {RE1.4} *Regression Software should document any assumptions made
-#' with regard to input data; for example distributional assumptions, or
-#' assumptions that predictor data have mean values of zero. Implications of
-#' violations of these assumptions should be both documented and tested.*
-#' The bootstrap weight options are described in a separate vignette article.
-#' In general, the wild bootstrap does not make any distributional assumptions
-#' for estimation beyond the assumption of a linear regression model.
-#'
 #' @srrstats {RE4.4} *The specification of the model, generally as a
 #' formula (via `formula()`)* Not applicable.
 #'
@@ -141,14 +94,6 @@
 #'  variables/columns) and speed of algorithm.* I don't really understand
 #' this requirement.
 #'
-#'
-#' @srrstats {RE6.1} *Where the default `plot` method is **NOT** a generic
-#'  `plot` method dispatched on the class of return objects (that is,
-#'  through an S3-type `plot.<myclass>` function or equivalent), that
-#'  method dispatch (or equivalent) should nevertheless exist in order
[…]
