### 2024-03-10 40a6c4f
update wording of compliance in srr-stats-standards.R
