+++ vignettes/articles/wild_bootstrap.Rmd
@@ -13,6 +13,13 @@ knitr::opts_chunk$set(
 
 ### Wild Cluster Bootstrap 101
 
+#' @srrstatsNA {RE1.4} *Regression Software should document any assumptions made
+#' with regard to input data; for example distributional assumptions, or
+#' assumptions that predictor data have mean values of zero. Implications of
+#' violations of these assumptions should be both documented and tested.*
+#' The wild bootstrap does not make any distributional assumptions
+#' for estimation beyond the assumption of a linear regression model.
+
 We are interested in testing a linear hypothesis $H_0: \beta_{j} = 0$ against $H_1: \beta_{j} \neq 0$
 
 for a linear model of interest
@@ -43,7 +50,7 @@ u_{G}
 
 with $E(u|X) = 0$, where group $g$ contains $N_{g}$ observations so that $N = \sum_{g = 1}^{G} N_{g}$. The regression residuals $u$ are allowed to be correlated within clusters, but are assumed to be uncorrelated across clusters. In consequence, the models' covariance matrix is block diagonal. For each cluster, we denote $E(u_{g} u_{g}') =\Sigma_{g}$ and $E(u u') =\Sigma$.
 
-A generic wild bootstrap test then proceeds in the following steps: 
+A generic wild bootstrap test then proceeds in the following steps:
 
 ::: callout
 * Step 1: Either
@@ -92,9 +99,9 @@ The subscript $x$ for the variance-covariance estimate $\hat{\Sigma}_{jj,x}$ den
 
 ### Multiple Variants of the Wild Cluster Bootstrap
 
-The above algorithm leads to several variants of the wild cluster bootstrap: first, the bootstrap variants may differ in the choice of the variance-covariance estimator $\Sigma_{x}$. Second, they may differ in the functional transformation of the estimated residuals $f(\hat{u})$. Finally, the choice of imposing or not imposing the null on the bootstrap data generating process, as well as the choice of bootstrap weights lead to different bootstrap variants. 
+The above algorithm leads to several variants of the wild cluster bootstrap: first, the bootstrap variants may differ in the choice of the variance-covariance estimator $\Sigma_{x}$. Second, they may differ in the functional transformation of the estimated residuals $f(\hat{u})$. Finally, the choice of imposing or not imposing the null on the bootstrap data generating process, as well as the choice of bootstrap weights lead to different bootstrap variants.
 
-Based on recent work by @mackinnon2022fast, we will put emphasis on the choice of the variance-covariance estimator $\Sigma_{x}$ and how the residuals $f(\hat{u})$ are transformed. 
+Based on recent work by @mackinnon2022fast, we will put emphasis on the choice of the variance-covariance estimator $\Sigma_{x}$ and how the residuals $f(\hat{u})$ are transformed.
 
 For reasons of computational feasibility, @mackinnon2022fast focus on CRV1 and CRV3 covariance matrix estimators.
 
@@ -182,7 +189,7 @@ For Rademacher and Mammen weights, only $2^G$ unique combinations of draws exist
 
 ### Wild Bootstrap Confidence Intervals
 
-In theory, multiple ways to calculate wild (cluster) bootstrapped confidence intervals exists [@mackinnon2022cluster]. 
+In theory, multiple ways to calculate wild (cluster) bootstrapped confidence intervals exists [@mackinnon2022cluster].
 <!-- If the null hypothesis is not imposed on the bootstrap dgp, -->
 <!-- The most straightforward approach is to calculate a standard error $\hat{s}^{*}(\hat{\beta})$ based on the bootstrapped regression coefficients $\hat{\beta}_{b}^{*}$ and a plug-in estimator: -->
 
@@ -202,7 +209,7 @@ In theory, multiple ways to calculate wild (cluster) bootstrapped confidence int
 
[…]
