### 2024-04-16 6a12790
update github action version for julia installation
