### 2024-04-16 s3alfisc/fwildclusterboot#149 @s3alfisc
Hi @mpadge ,

- I have moved all non applicable standards to NA section: https://github.com/s3alfisc/fwildclusterboot/pull/148/commits/b80f9c767601bc6140b6a721d92d306b8348c89d
- I moved RE1.4 and RE4.17 into the suggested spots https://github.com/s3alfisc/fwildclusterboot/pull/148/commits/517b46efe98f1432806c16947d3f0d969feb1495
- I have added details of computational complexity of the most costly operation of the algorithm https://github.com/s3alfisc/fwildclusterboot/pull/148/commits/523bae4eaa742c275aaa03e0530c29060004b584

Once all unit tests ([via GHA](https://github.com/s3alfisc/fwildclusterboot/actions/runs/8711654572/job/23896240704?pr=148)) pass, I will merge version 0.15.0 in to main =)
