### 2024-04-16 0b87b07
delete is_congruent function
