+++ NAMESPACE
@@ -56,7 +56,6 @@ importFrom(dreamerr,validate_dots)
 importFrom(dreamerr,warn_up)
 importFrom(generics,glance)
 importFrom(generics,tidy)
-importFrom(gtools,permutations)
 importFrom(parallel,detectCores)
 importFrom(stats,aggregate)
 importFrom(stats,as.formula)
