+++ tests/testthat/test-method_equivalence.R
@@ -6,22 +6,22 @@ test_that("Do different, but equivalent ways to specify
     message = "skip test as julia installation not found."
   )
 
-  set.seed(2351)
-  dqrng::dqset.seed(2351)
+  set.seed(23351)
+  dqrng::dqset.seed(23513)
 
 
   print_results <- FALSE
 
   data1 <<-
     fwildclusterboot:::create_data(
-      N = 10000,
+      N = 15000,
       N_G1 = 20,
       icc1 = 0.01,
       N_G2 = 10,
       icc2 = 0.01,
       numb_fe1 = 10,
       numb_fe2 = 10,
-      seed = 71986045
+      seed = 7198045
     )
   # sapplydata1, class)
 
@@ -188,7 +188,7 @@ test_that("Do different, but equivalent ways to specify
       # boottest()
       cat("boottest()", "\n")
       set.seed(8965)
-      dqrng::dqset.seed(765)
+      dqrng::dqset.seed(12111221)
 
       assign(paste0("boot_lm_", engine),
         suppressWarnings(
@@ -207,7 +207,7 @@ test_that("Do different, but equivalent ways to specify
       )
 
       set.seed(8965)
-      dqrng::dqset.seed(765)
+      dqrng::dqset.seed(12111221)
 
       assign(paste0("boot_fixest1_", engine),
         suppressWarnings(
@@ -226,7 +226,7 @@ test_that("Do different, but equivalent ways to specify
       )
 
       set.seed(8965)
-      dqrng::dqset.seed(765)
+      dqrng::dqset.seed(12111221)
 
       assign(paste0("boot_fixest2_", engine),
         suppressWarnings(
@@ -245,7 +245,7 @@ test_that("Do different, but equivalent ways to specify
       )
 
       set.seed(8965)
-      dqrng::dqset.seed(765)
+      dqrng::dqset.seed(12111221)
 
       assign(paste0("boot_fixest3_", engine),
         suppressWarnings(
@@ -264,7 +264,7 @@ test_that("Do different, but equivalent ways to specify
       )
 
       set.seed(8965)
-      dqrng::dqset.seed(765)
+      dqrng::dqset.seed(12111221)
 
       assign(paste0("boot_fixest4_", engine),
         suppressWarnings(
@@ -283,7 +283,7 @@ test_that("Do different, but equivalent ways to specify
       )
 
       set.seed(8965)
-      dqrng::dqset.seed(765)
+      dqrng::dqset.seed(12111221)
 
       assign(paste0("boot_fixest5_", engine),
         suppressWarnings(
@@ -302,7 +302,7 @@ test_that("Do different, but equivalent ways to specify
       )
 
       set.seed(8965)
-      dqrng::dqset.seed(765)
+      dqrng::dqset.seed(12111221)
 
       assign(paste0("boot_fixest6_", engine),
         suppressWarnings(
@@ -321,7 +321,7 @@ test_that("Do different, but equivalent ways to specify
       )
 
       set.seed(8965)
-      dqrng::dqset.seed(765)
+      dqrng::dqset.seed(12111221)
 
       assign(paste0("boot_fixest7_", engine),
         suppressWarnings(
@@ -340,7 +340,7 @@ test_that("Do different, but equivalent ways to specify
       )
 
       set.seed(8965)
-      dqrng::dqset.seed(765)
+      dqrng::dqset.seed(12111221)
 
       assign(paste0("boot_fixest8_", engine),
         suppressWarnings(
@@ -359,7 +359,7 @@ test_that("Do different, but equivalent ways to specify
       )
 
       set.seed(8965)
-      dqrng::dqset.seed(765)
+      dqrng::dqset.seed(12111221)
 
       assign(paste0("boot_fixest9_", engine),
         suppressWarnings(
[…]
