+++ DESCRIPTION
@@ -1,72 +1,72 @@
 Package: fwildclusterboot
 Title: Fast Wild Cluster Bootstrap Inference for Linear Models
-Version: 0.13.0
-Authors@R: 
+Version: 0.13.1
+Authors@R:
     c(
     person(given = "Alexander",
            family = "Fischer",
            role = c("aut", "cre"),
            email = "alexander-fischer1801@t-online.de"),
-    person(given = "David", 
-           family = "Roodman", 
-           role = "aut"), 
-    person(given = "Achim", 
+    person(given = "David",
+           family = "Roodman",
+           role = "aut"),
+    person(given = "Achim",
            family = "Zeileis",
            role = c("ctb"),
            comment = "Author of included sandwich fragments"),
 	person(given = "Nathaniel",
 	       family = "Graham",
-	       role = c("ctb"), 
+	       role = c("ctb"),
            comment = "Contributor to included sandwich fragments"),
 	person(given = "Susanne",
-	       family = "Koell", 
-	       role = c("ctb"), 
+	       family = "Koell",
+	       role = c("ctb"),
            comment = "Contributor to included sandwich fragments"),
-	person("Laurent", 
+	person("Laurent",
 	        "Berge",
-	        role = c("ctb"), 
+	        role = c("ctb"),
             comment = "Author of included fixest fragments"),
-	 person("Sebastian", 
-	        "Krantz", 
+	 person("Sebastian",
+	        "Krantz",
 	         role = c("ctb"))
 	)
-Description: Implementation of fast algorithms for wild cluster bootstrap 
+Description: Implementation of fast algorithms for wild cluster bootstrap
              inference developed in 'Roodman et al' (2019, 'STATA' Journal,
-             <doi:10.1177/1536867X19830877>) and 'MacKinnon et al' (2022), 
-             which makes it feasible to quickly calculate bootstrap test 
-             statistics based on a large number of bootstrap draws even for 
-             large samples. Multiple bootstrap types as described in 'MacKinnon, 
-             Nielsen & Webb' (2022) are supported. 
-             Further, 'multiway' clustering, regression weights, 
+             <doi:10.1177/1536867X19830877>) and 'MacKinnon et al' (2022),
+             which makes it feasible to quickly calculate bootstrap test
+             statistics based on a large number of bootstrap draws even for
+             large samples. Multiple bootstrap types as described in 'MacKinnon,
+             Nielsen & Webb' (2022) are supported.
+             Further, 'multiway' clustering, regression weights,
              bootstrap weights, fixed effects and 'subcluster' bootstrapping
              are supported. Further, both restricted ('WCR') and unrestricted
-             ('WCU') bootstrap are supported. Methods are provided for a variety 
-             of fitted models, including 'lm()', 'feols()' 
-             (from package 'fixest') and 'felm()' (from package 'lfe'). 
-             Additionally implements a 'heteroskedasticity-robust' ('HC1') wild 
+             ('WCU') bootstrap are supported. Methods are provided for a variety
+             of fitted models, including 'lm()', 'feols()'
+             (from package 'fixest') and 'felm()' (from package 'lfe').
+             Additionally implements a 'heteroskedasticity-robust' ('HC1') wild
              bootstrap.
              Last, the package provides an R binding to 'WildBootTests.jl',
-             which provides additional speed gains and functionality, 
-             including the 'WRE' bootstrap for instrumental variable models 
+             which provides additional speed gains and functionality,
[…]
