### 2024-08-05 e096b66
fix failing tests, update version
