### 2024-08-05 0f06778
adress more cran comments, fix typo
