+++ R/utils.R
@@ -101,17 +101,17 @@ create_data <-
         state = rep(1:(N / 20), 20),
         year = sort(rep(1:20, N / 20))
       )
-    
+
     voters$Q1_immigration <- as.factor(voters$Q1_immigration)
     voters$Q2_defense <- as.factor(voters$Q2_defense)
-    
-    
+
+
     voters$log_income <- log(voters$income)
     voters$Q1_immigration <- as.factor(voters$Q1_immigration)
-    
+
     # add weights
     voters$weights <- sample(1:10, N, replace = TRUE) / 10
-    
+
     voters
   }
 
@@ -129,7 +129,7 @@ create_data <-
 #     #' @param set logical flag for duplicates
 #     #' @param repeats.allowed logical flag
 #     #' @noRd
-#     
+#
 #     if (mode(n) != "numeric" || length(n) != 1 || n < 1 ||
 #         (n %% 1) != 0) {
 #       stop("bad value of n")
@@ -190,31 +190,31 @@ create_data <-
 #   }
 
 setBoottest_engine <- function(engine) {
-  
-  #' Sets the default bootstrap algo for the current R session 
+
+  #' Sets the default bootstrap algo for the current R session
   #' to be run via `boottest()` and `mboottest()`
-  #' 
+  #'
   #' @param engine Character scalar. Either 'R' or 'WildBootTests.jl'.
   #'  Default is 'R'
   #' @return No return value
-  #' 
+  #'
   #' @export
-  #' 
+  #'
   #' @examples
   #' \dontrun{
   #' setBoottest_engine(engine = "R")
   #' setBoottest_engine(engine = "WildBootTests.jl")
   #' }
-  
+
   if (missing(engine) || is.null(engine)) {
     # New default: one cores used
     engine <- "R"
   }
-  
+
   engine <- set_engine(engine)
-  
+
   options("boottest_engine" = engine)
-  
+
   invisible()
 }
 
@@ -222,7 +222,7 @@ getBoottest_engine <- function() {
   #' get the bootstrap algorithm to be run via `boottest()` and `waldboottest()`
   #' @return The number of threads currently used by boottest as set in options
   #' @noRd
-  
+
   x <- getOption("boottest_engine")
   if (!(x %in% c("R", "WildBootTests.jl"))) {
     rlang::abort(
@@ -236,13 +236,13 @@ getBoottest_engine <- function() {
 }
 
 set_engine <- function(engine) {
-  
+
   #' check the bootstrap algo
   #' @param engine character scalar
   #' @noRd
-  
+
   dreamerr::check_value(engine, "charin(R, WildBootTests.jl)")
-  
+
   engine
 }
 
@@ -257,29 +257,29 @@ set_engine <- function(engine) {
 # 2. the default number of threads in check_set_nthreads is set to 1
 
 setBoottest_nthreads <- function(nthreads) {
-  
+
   #' Set the number of threads for use with open mp via options
   #' By default, only one thread is used
   #' @param nthreads Integer. Number of threads to be used
   #' @return No return value
   #' @noRd
   #' @importFrom parallel detectCores
-  
+
   max_CRAN <- as.numeric(Sys.getenv("OMP_THREAD_LIMIT"))
   max_CRAN[is.na(max_CRAN)] <- 1000
-  
+
   max_threads <- min(cpp_get_nb_threads(), 1000, max_CRAN)
   # we cap at 1k nthreads
-  
+
   if (missing(nthreads) || is.null(nthreads)) {
     # New default: one cores used
     nthreads <- 1
   }
-  
+
   nthreads <- check_set_nthreads(nthreads)
-  
+
   options("boottest_nthreads" = nthreads)
-  
+
   invisible()
 }
 
@@ -288,12 +288,12 @@ getBoottest_nthreads <- function() {
[…]
