### 2024-08-24 #546 @adamhsparks
Hi @s3alfisc, I'm going through open issues as the current EIC and wanted to check on this one to see if you had any updates on your progress addressing @mpadge's TODO list?
