### 2024-11-10 #546 @emilyriederer
Hey @s3alfisc ! I'm taking over as EoC this quarter and doing a review of currently open issues. Are you still actively working on these changes or should we consider putting this review on hold?
