### 2025-02-02 #546 @maurolepore
@s3alfisc I just started my EiC rotation and see this issue flagged as "most urgent". Are you able to continue or should we label as holding?

I see the last contact intent was more than a month ago -- by [emilyriederer](https://github.com/emilyriederer)
[on Nov 10, 2024](https://github.com/ropensci/software-review/issues/546#issuecomment-2466989297).

Our policy (pasted below) states we should close this issue and let you know also by email. I see @mpadge things this is very close to the finish line so I'll postpone the closing a few days hoping you'll read this message during the upcoming work week.

In any case, I totally understand things can change and you may not be able to continue.

---

https://devguide.ropensci.org/softwarereview_policies.html#policiesreviewprocess

> If the author hasn’t requested a holding label, but is simply not responding, we should close the issue within one month after the last contact intent. This intent will include a comment tagging the author, but also an email using the email address listed in the DESCRIPTION of the package which is one of the rare cases where the editor will try to contact the author by email.
