### 2025-05-04 #546 @maurolepore
@s3alfisc I'm wrapping up my EiC rotation. I'm closing this submission for lack of response. As mentioned above our
policy states we should close this issue and let you. I just sent you an email letting you know.

I hope you're well, and I totally understand things can change and you may not be able to continue.

Thanks so much for sharing your work with rOpenSci and I look forward to your next submission whenever it's a good time for it.
